### 2025-08-04 #8749 @editorialbot
[PRE REVIEW]: scarlet2: Astronomical scene modeling in jax
**Submitting author:** @pmelchior (<a href="http://orcid.org/0000-0002-8873-5065">Peter Melchior</a>)
**Repository:** https://github.com/pmelchior/scarlet2/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.0
