+++ src/scarlet2/lsst_utils.py
@@ -1,0 +1,268 @@
+import time
+import numpy as np
+import uuid
+import os
+import glob
+import math
+import pandas
+import matplotlib.pyplot as plt
+import tempfile
+import getpass
+from PIL import Image
+from IPython.display import Image as dimg
+
+import lsst.geom as geom
+import lsst.resources
+import lsst.afw.display as afwDisplay
+from lsst.afw.image import Exposure, ExposureF
+from lsst.pipe.tasks.registerImage import RegisterConfig, RegisterTask
+from lsst.rsp import get_tap_service
+from lsst.rsp.utils import get_access_token
+from lsst.afw.fits import MemFileManager
+
+import pyvo
+from pyvo.dal.adhoc import DatalinkResults, SodaQuery
+
+import astropy
+from astropy import units as u
+from astropy.coordinates import SkyCoord, Angle
+from astropy.io import fits
+from astropy.wcs import WCS
+
+import scarlet2
+import matplotlib.pyplot as plt
+import numpy as np
+from scipy.optimize import curve_fit
+from scipy.stats import skew
+import getpass
+
+import lsst.daf.butler as dafButler
+import lsst.afw.display as afwDisplay
+from lsst.geom import Point2D, radToDeg, SpherePoint, degrees
+
+from astropy.wcs import WCS
+from astropy.io import fits
+import jax.numpy as jnp
+import lsst.afw.geom as afwGeom
+import lsst.afw.image as afwImage
+import lsst.afw.display as afwDisplay
+import lsst.afw.table as afwTable
+import lsst.geom as geom
+from lsst.afw.detection import Psf
+from lsst.meas.algorithms import WarpedPsf
+
+def warp_img(ref_img, img_to_warp, ref_wcs, wcs_to_warp):
+    """Warp and rotate an image onto the coordinate system of another image
+
+   Parameters
+    ----------
+    ref_img: 'ExposureF'
+        is the reference image for the re-projection
+    img_to_warp: 'ExposureF'
+        the image to rotate and warp onto the reference image's wcs
+    ref_wcs: 'wcs object'
+        the wcs of the reference image (i.e. ref_img.getWcs() )
+    wcs_to_warp: 'wcs object'
+        the wcs of the image to warp (i.e. img_to_warp.getWcs() )
+   Returns
+    -------
+    warpedExp: 'ExposureF'
+        a reprojected, rotated image that is aligned and matched to ref_image
+     """
+    config = RegisterConfig()
+    task = RegisterTask(name="register", config=config)
+    warpedExp = task.warpExposure(img_to_warp, wcs_to_warp, ref_wcs,
+                                  ref_img.getBBox())
+
+    return warpedExp
+
+def read_cutout_mem(sq):
+    """Read the cutout into memory
+
+    Parameters
+    ----------
+    sq : 'dict'
+        returned from SodaQuery.from_resource()
+
+    Returns
+    -------
+    exposure : 'ExposureF'
+        the cutout in exposureF format
+    """
+
+    cutout_bytes = sq.execute_stream().read()
+    sq.raise_if_error()
+    mem = MemFileManager(len(cutout_bytes))
+    mem.setData(cutout_bytes, len(cutout_bytes))
+    exposure = ExposureF(mem)
+
+    return exposure
+
+def make_image_cutout(tap_service, ra, dec, dataId, cutout_size=0.01,
+                      imtype=None, filename=None):
+    """Wrapper function to generate a cutout using the cutout tool
+
+   Parameters
+    ----------
+    tap_service : an instance of the TAP service
+    ra, dec : 'float'
+        the ra and dec of the cutout center
+    dataId : 'dict'
+        the dataId of the image to make a cutout from. The format
+        must correspond to that provided for parameter 'imtype'
[…]
