+++ src/scarlet2/lsst_utils.py
@@ -1,60 +1,25 @@
-import time
-import numpy as np
-import uuid
-import os
-import glob
-import math
-import pandas
-import matplotlib.pyplot as plt
-import tempfile
-import getpass
-from PIL import Image
-from IPython.display import Image as dimg
-
-import lsst.geom as geom
-import lsst.resources
-import lsst.afw.display as afwDisplay
-from lsst.afw.image import Exposure, ExposureF
-from lsst.pipe.tasks.registerImage import RegisterConfig, RegisterTask
-from lsst.rsp import get_tap_service
-from lsst.rsp.utils import get_access_token
-from lsst.afw.fits import MemFileManager
-
-import pyvo
-from pyvo.dal.adhoc import DatalinkResults, SodaQuery
-
-import astropy
-from astropy import units as u
-from astropy.coordinates import SkyCoord, Angle
-from astropy.io import fits
-from astropy.wcs import WCS
-
-import scarlet2
-import matplotlib.pyplot as plt
-import numpy as np
-from scipy.optimize import curve_fit
-from scipy.stats import skew
-import getpass
-
-import lsst.daf.butler as dafButler
-import lsst.afw.display as afwDisplay
-from lsst.geom import Point2D, radToDeg, SpherePoint, degrees
-
-from astropy.wcs import WCS
-from astropy.io import fits
 import jax.numpy as jnp
 import lsst.afw.geom as afwGeom
-import lsst.afw.image as afwImage
-import lsst.afw.display as afwDisplay
-import lsst.afw.table as afwTable
 import lsst.geom as geom
-from lsst.afw.detection import Psf
+import matplotlib.pyplot as plt
+import numpy as np
+from astropy import units as u
+from astropy.io import fits
+from astropy.wcs import WCS
+from lsst.afw.fits import MemFileManager
+from lsst.afw.image import ExposureF
+from lsst.geom import Point2D
 from lsst.meas.algorithms import WarpedPsf
+from lsst.pipe.tasks.registerImage import RegisterConfig, RegisterTask
+from pyvo.dal.adhoc import DatalinkResults, SodaQuery
+
+import scarlet2
+
 
 def warp_img(ref_img, img_to_warp, ref_wcs, wcs_to_warp):
     """Warp and rotate an image onto the coordinate system of another image
 
-   Parameters
+    Parameters
     ----------
     ref_img: 'ExposureF'
         is the reference image for the re-projection
@@ -64,17 +29,17 @@ def warp_img(ref_img, img_to_warp, ref_wcs, wcs_to_warp):
         the wcs of the reference image (i.e. ref_img.getWcs() )
     wcs_to_warp: 'wcs object'
         the wcs of the image to warp (i.e. img_to_warp.getWcs() )
-   Returns
+    Returns
     -------
     warpedExp: 'ExposureF'
         a reprojected, rotated image that is aligned and matched to ref_image
-     """
+    """
     config = RegisterConfig()
     task = RegisterTask(name="register", config=config)
-    warpedExp = task.warpExposure(img_to_warp, wcs_to_warp, ref_wcs,
-                                  ref_img.getBBox())
+    warpedExp = task.warpExposure(img_to_warp, wcs_to_warp, ref_wcs, ref_img.getBBox())
 
     return warpedExp
+
 
 def read_cutout_mem(sq):
     """Read the cutout into memory
@@ -98,11 +63,11 @@ def read_cutout_mem(sq):
 
     return exposure
 
-def make_image_cutout(tap_service, ra, dec, dataId, cutout_size=0.01,
-                      imtype=None, filename=None):
+
+def make_image_cutout(tap_service, ra, dec, dataId, cutout_size=0.01, imtype=None, filename=None):
[…]
