### 2025-08-08 4e32e4d
add back in file writing for wcs reasons
