### 2025-08-14 a87228d
add back in plotting
