### 2025-08-14 427ec3a
Adding simple tests for SourceValidation.
