### 2025-08-14 128f1a3
more plotting tweaks
