### 2025-08-14 3da68ff
Renamed a test file for consistency. Added unit tests for `check_source_boxes_comparable_to_observation`.
