+++ tests/scarlet2/test_scene_checks.py
@@ -1,0 +1,173 @@
+from functools import partial
+
+import numpy as np
+import pytest
+from huggingface_hub import hf_hub_download
+from numpyro.distributions import constraints
+from scarlet2 import init
+from scarlet2.frame import Frame
+from scarlet2.module import Parameter, relative_step
+from scarlet2.observation import Observation
+from scarlet2.psf import ArrayPSF
+from scarlet2.scene import Scene, SceneValidator
+from scarlet2.source import PointSource, Source
+from scarlet2.validation_utils import (
+    ValidationInfo,
+    ValidationWarning,
+    set_validation,
+)
+
+
+@pytest.fixture(autouse=True)
+def setup_validation():
+    """Automatically disable validation for all tests. This permits the creation
+    of intentionally invalid Observation objects."""
+    set_validation(False)
+
+
+@pytest.fixture()
+def data_file():
+    """Download and load a realistic test file. This is the same data used in the
+    quickstart notebook. The data will be manipulated to create invalid inputs for
+    the `bad_obs` fixture."""
+    filename = hf_hub_download(
+        repo_id="astro-data-lab/scarlet-test-data", filename="hsc_cosmos_35.npz", repo_type="dataset"
+    )
+    return np.load(filename)
+
+
+@pytest.fixture()
+def bad_obs(data_file):
+    """Create an observation that should fail multiple validation checks."""
+
+    data = np.asarray(data_file["images"])
+    channels = [str(f) for f in data_file["filters"]]
+    weights = np.asarray(1 / data_file["variance"])
+    psf = np.asarray(data_file["psfs"])
+
+    weights = weights[:-1]  # Remove the last weight to create a mismatch in dimensions
+    weights[0][0] = np.inf  # Set one weight to infinity
+    weights[1][0] = -1.0  # Set one weight to a negative value
+    psf = psf[:-1]  # Remove the last PSF to create a mismatch in dimensions
+    psf = psf[0] + 0.001
+
+    return Observation(
+        data=data,
+        weights=weights,
+        channels=channels,
+        psf=ArrayPSF(psf),
+    )
+
+
+@pytest.fixture()
+def good_obs(data_file):
+    """Create an observation that should pass all validation checks."""
+    data = np.asarray(data_file["images"])
+    channels = [str(f) for f in data_file["filters"]]
+    weights = np.asarray(1 / data_file["variance"])
+    psf = np.asarray(data_file["psfs"])
+
+    return Observation(
+        data=data,
+        weights=weights,
+        channels=channels,
+        psf=ArrayPSF(psf),
+    )
+
+
+@pytest.fixture()
+def full_overlap_scene(good_obs, data_file):
+    """Assemble a scene from the good observation and the data file."""
+    model_frame = Frame.from_observations(good_obs)
+    centers = np.array([(src["y"], src["x"]) for src in data_file["catalog"]])  # Note: y/x convention!
+
+    with Scene(model_frame) as scene:
+        spectrum = init.pixel_spectrum(good_obs, centers[0], correct_psf=True)
+        PointSource(centers[0], spectrum)
+
+    return scene
+
+
+@pytest.fixture()
+def partial_overlap_scene(good_obs):
+    """Assemble a scene from the good observation and the data file."""
+    model_frame = Frame.from_observations(good_obs)
+
+    with Scene(model_frame) as scene:
+        spectrum = init.pixel_spectrum(good_obs, np.array([1, 1]), correct_psf=True)
+        PointSource(np.array([1, 1]), spectrum)
+
+    return scene
+
+
+@pytest.fixture()
+def no_overlap_scene(good_obs):
+    """Assemble a scene from the good observation and the data file."""
+    model_frame = Frame.from_observations(good_obs)
+
+    with Scene(model_frame) as scene:
[…]
