### 2025-08-14 6eb151f
Added unit tests for `check_parameters_have_necessary_fields`.
