### 2025-08-15 ae86742
Adding unit tests that mock chi-squared values to confirm the logic in the FitValidation checks.
