### 2025-08-15 a043aa3
Cleaning up validation check tests. Moved most pytest.fixtures to conftest.py to reduce duplicated code and shrink the number of lines in the test modules.
