+++ src/scarlet2/lsst_utils.py
@@ -1,5 +1,5 @@
 import jax.numpy as jnp
-import lsst.afw.geom as afwGeom
+import lsst.afw.geom as afw_geom
 import lsst.geom as geom
 import matplotlib.pyplot as plt
 import numpy as np
@@ -30,14 +30,14 @@ def warp_img(ref_img, img_to_warp, ref_wcs, wcs_to_warp):
         the wcs of the image to warp (i.e. img_to_warp.getWcs() )
     Returns
     -------
-    warpedExp: 'ExposureF'
+    warped_exp: 'ExposureF'
         a reprojected, rotated image that is aligned and matched to ref_image
     """
     config = RegisterConfig()
     task = RegisterTask(name="register", config=config)
-    warpedExp = task.warpExposure(img_to_warp, wcs_to_warp, ref_wcs, ref_img.getBBox())
+    warped_exp = task.warpExposure(img_to_warp, wcs_to_warp, ref_wcs, ref_img.getBBox())
 
-    return warpedExp
+    return warped_exp
 
 
 def read_cutout_mem(sq):
@@ -63,15 +63,16 @@ def read_cutout_mem(sq):
     return exposure
 
 
-def make_image_cutout(tap_service, ra, dec, dataId, cutout_size=0.01, imtype=None, filename=None):
+def make_image_cutout(tap_service, ra, dec, data_id, cutout_size=0.01, imtype=None):
     """Wrapper function to generate a cutout using the cutout tool
 
     Parameters
     ----------
-    tap_service : an instance of the TAP service
+    tap_service : `pyvo.dal.tap.TAPService`
+        the TAP service to use for querying the cutouts
     ra, dec : 'float'
         the ra and dec of the cutout center
-    dataId : 'dict'
+    data_id : 'dict'
         the dataId of the image to make a cutout from. The format
         must correspond to that provided for parameter 'imtype'
     cutout_size : 'float', optional
@@ -80,8 +81,6 @@ def make_image_cutout(tap_service, ra, dec, dataId, cutout_size=0.01, imtype=Non
         string containing the type of LSST image to generate
         a cutout of (e.g. deepCoadd, calexp). If imtype=None,
         the function will assume a deepCoadd.
-    filename : 'string', optional
-        filename of the resulting cutout (which has fits format)
 
     Returns
     -------
@@ -89,7 +88,7 @@ def make_image_cutout(tap_service, ra, dec, dataId, cutout_size=0.01, imtype=Non
         the cutout in exposureF format
     """
 
-    spherePoint = geom.SpherePoint(ra * geom.degrees, dec * geom.degrees)
+    sphere_point = geom.SpherePoint(ra * geom.degrees, dec * geom.degrees)
 
     if imtype == "calexp":
         query = (
@@ -99,20 +98,20 @@ def make_image_cutout(tap_service, ra, dec, dataId, cutout_size=0.01, imtype=Non
             + "AND obs_collection = 'LSST.DP02' "
             + "AND dataproduct_subtype = 'lsst.calexp' "
             + "AND lsst_visit = "
-            + str(dataId["visit"])
+            + str(data_id["visit"])
             + " "
             + "AND lsst_detector = "
-            + str(dataId["detector"])
+            + str(data_id["detector"])
         )
         results = tap_service.search(query)
 
     else:
         # Find the tract and patch that contain this point
-        tract = dataId["tract"]
-        patch = dataId["patch"]
+        tract = data_id["tract"]
+        patch = data_id["patch"]
 
-        # add optional default band if it is not contained in the dataId
-        band = dataId["band"]
+        # add optional default band if it is not contained in the data_id
+        band = data_id["band"]
 
[…]
