### 2025-08-19 5b83646
Update validation UX, include final set of requested validation checks (#191)
* WIP - Adding initial Scene.fit validation checks.

* Starting to add 'after initialization' validation checks.

* Introducing new ValidationResults levels, print all validation check results, update corresponding unit tests.

* Implementing more "after initialization" validation checks.

* Adding simple tests for SourceValidation.

* Renamed a test file for consistency. Added unit tests for `check_source_boxes_comparable_to_observation`.

* Added unit tests for `check_parameters_have_necessary_fields`.

* Adding un
[…]
