+++ src/scarlet2/scene.py
@@ -1,5 +1,4 @@
 from collections import defaultdict
-from typing import Optional
 
 import equinox as eqx
 import jax
@@ -10,7 +9,14 @@ from .bbox import overlap_slices
 from .frame import Frame
 from .module import Module, Parameters
 from .renderer import ChannelRenderer
-from .validation_utils import ValidationError, ValidationMethodCollector
+from .validation_utils import (
+    ValidationError,
+    ValidationInfo,
+    ValidationMethodCollector,
+    ValidationResult,
+    ValidationWarning,
+    print_validation_results,
+)
 
 
 class Scene(Module):
@@ -340,12 +346,10 @@ class Scene(Module):
         if VALIDATION_SWITCH:
             from .validation import check_fit
 
-            validation_errors = check_fit(returned_scene)
-            if validation_errors:
-                raise ValueError(
-                    "Fit validation failed. The following errors were found:\n"
-                    + "\n".join(str(error) for error in validation_errors)
-                )
+            for i, obs in enumerate(observations):
+                print(f"Running validation checks on the fit of the scene for observation {i}.")
+                validation_results = check_fit(returned_scene, obs)
+                print_validation_results("Fit validation results", validation_results)
 
         return returned_scene
 
@@ -526,15 +530,241 @@ class FitValidator(metaclass=ValidationMethodCollector):
     validation methods in this class into a single class attribute list called
     `validation_checks`. This allows for easy iteration over all checks."""
 
-    def __init__(self, scene: Scene):
-        self.scene = scene
+    def __init__(self, scene: Scene, observation):
+        """Initialize the FitValidator.
 
-    def check_fit_example(self) -> Optional[ValidationError]:
-        """Check that the fit was successful.
+        Parameters
+        ----------
+        scene : Scene
+            The scene object to validate.
+        observation : Observation
+            The observation object containing the data to validate against.
+        """
+        self.scene = scene
+        self.observation = observation
+
+        # These are placeholders, waiting for the actual width of distribution to be
+        # implemented - see issue https://github.com/pmelchior/scarlet2/issues/192
+        self.chi2_tolerable_threshold = 1.5
+        self.chi2_critical_threshold = 5.0
+
+    def check_goodness_of_fit(self) -> ValidationResult:
+        """Evaluate the goodness of the model fit to the data by calling the Observation
+        class's `goodness_of_fit` method.
+
+        For a Gaussian noise model, the gof is defined as the averaged squared deviation of the model from the
+        data, scaled by the variance of the data, aka mean chi squared
+        :math:`\frac{1}{N}\\sum_i=1^N w_i (m_i - d_i)^2` with inverse variance weights :math:`w_i`.
+
+        Up to a normalization, the gof is identical to :py:class:`~scarlet2.Observation.log-likelihood`.
+
+        Parameters
+        ----------
+        model: array
+            The (pre-rendered) predicted data cube, typically from evaluating :py:class:`~scarlet2.Scene`
 
         Returns
         -------
-        ValidationError or None
-            Returns a ValidationError if the check fails, otherwise None.
+        float
         """
-        return None
+        obs = self.observation
+
+        chi2 = obs.goodness_of_fit(self.scene())
+
+        ret_val: ValidationResult = ValidationInfo("The model fit is good.", check=self.__class__.__name__)
+        if self.chi2_tolerable_threshold <= chi2 < self.chi2_critical_threshold:
+            ret_val = ValidationWarning(
+                "The model fit is acceptable, but the goodness of fit is not optimal.",
+                check=self.__class__.__name__,
+                context={"chi2": chi2},
+            )
+        elif chi2 >= self.chi2_critical_threshold:
+            ret_val = ValidationError(
[…]
