+++ docs/2-validation.md
@@ -1,0 +1,277 @@
+# Validation
+
+## Automatic validation
+There are validation checks that run automatically at various points in a ``scarlet2``
+workflow.
+* Observation class checks
+* Scene class checks
+* Scene.fit output checks
+* Source class checks
+
+All the validation checks are enabled by default, they run quickly, so there is
+minimal time penalty for leaving them on by default.
+
+Three tiers Info, Warn, Error. When validation runs, every check will return some result.
+The results are written out as log lines.
+None of the results of validation checks will cause the program to halt.
+i.e. even a ValidationError will not cause the execution of ``scarlet2`` to stop.
+
+If you want to turn off automatic validation, run: 
+```
+set_validation=False
+```
+
+If you would like to run validation checks manually, use one of the following:
+* ``check_fit(scene, observation)``
+* ``check_observation(observation)``
+* ``check_scene(scene, observation, parameters)``
+* ``check_source(source)``
+
+## Dev guide - Implementing validation checks
+The goal of the validation checks is to provide guidance for users as they work with ``scarlet2``.
+Since guidance changes and the code base evolves, we want to encourage developing
+and maintaining validation checks.
+
+Here we include two examples to walk a developer through implementing new validation
+checks within the framework that has been established.
+
+
+### Example - Adding a new validation check
+Ideally when modifying existing code or extending API functionality, the developer
+will add corresponding validation checks.
+
+Here we'll imagine that a hypothetical method has been added to the ``Observation``
+class that will take the square root of a new input parameter ``obs_sqrt``.
+The user is free to provide any value for the new input parameter.
+We'll write a validation check to provide some guardrails for the user.
+
+#### Write the basic validation check
+Validation checks are implemented as methods in a ``*Validation`` class that
+is co-located with the class it is validating.
+For this example all the validation checks for the ``Observation`` class are
+contained in the ``observation.py`` module in a class named
+``ObservationValidator``.
+
+Well add the following method to the ``ObservationValidator`` class:
+```
+import numbers
+
+def check_sqrt_parameter(self) -> ValidationResult:
+    """Check that the parameter to be passed to the ``obs_sqrt`` function is
+    reasonable.
+
+    Returns
+    -------
+    ValidationResult
+        An appropriate ValidationResult subclass based on the value of ``obs_sqrt``.
+
+    obs_sqrt = self.observation.obs_sqrt
+
+    if not isinstance(obs_sqrt, numbers.Number):
+        return ValidationError(
+            message="The value of `obs_sqrt` is not numeric",
+            check=self.__class__.__name__,
+            context={"obs_sqrt": obs_sqrt}
+        )
+
+    elif(obs_sqrt < 0.):
+        return ValidationWarning(
+            message="The value of `obs_sqrt` is < 0. This might be a mistake.",
+            check=self.__class__.__name__,
+            context={"obs_sqrt": obs_sqrt}
+        )
+    else:
+        return ValidationInfo(
+            message="The value of `obs_sqrt` looks good.",
+            check=self.__class__.__name__,
+        )
+```
+
+When adding a new validation check, it is required to use the following naming scheme:
+``check_<something>``.
+Failure to do so will mean the check won't be run automatically with the rest of the checks.
+
+A validation check should always return at least one instance of a subclass of ``ValidationResult``.
+Returning a list of ``ValidationResults`` subclasses is also ok.
+You should not return an instance of the ``ValidationResults`` directly.
+The available subclasses are:
+* ValidationInfo - for checks passed without any concerns.
+* ValidationWarning - for something concerning, but probably won't prevent downstream tasks from completing.
[…]
