### 2025-08-22 e904260
Updating docstring and return type hints. Updated a couple variable names in validation.py.
