### 2025-08-22 fd6f019
Accept copilot recommendation for src/scarlet2/scene.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
