### 2025-08-22 fb6cadf
Fixing up the botched Github copilot commit.
