### 2025-08-22 55b8b20
Clean up docstrings and type hints in validation files (#208)
* Updating docstring and return type hints. Updated a couple variable names in validation.py.

* Accept copilot recommendation for src/scarlet2/scene.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Fixing up the botched Github copilot commit.

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
