### 2025-08-22 c09d6cf
Filled out the user guide section for automatic validation.
