### 2025-08-22 5e3bd0e
use jinja for output box
