+++ pyproject.toml
@@ -31,6 +31,7 @@ dependencies = [
     "pyyaml",
     "markdown",
     "pygments",
+    "jinja2"
 ]
 
 [project.urls]
@@ -57,9 +58,6 @@ build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
 write_to = "src/scarlet2/_version.py"
-
-[tool.setuptools.package-data]
-"scarlet2.questionnaire" = ["questions.yaml"]
 
 [tool.pytest.ini_options]
 testpaths = [
