### 2025-08-22 9ec1a66
use jinja for output box
