+++ tests/scarlet2/questionnaire/conftest.py
@@ -1,4 +1,17 @@
+import json
+import re
+from importlib.resources import files
+
+from pygments import highlight
+from pygments.formatters.html import HtmlFormatter
+from pygments.lexers.python import PythonLexer
 from pytest import fixture
+
+from ipywidgets import HTML, HBox, VBox, Button, Label
+
+from scarlet2.questionnaire.models import Questionnaire
+from scarlet2.questionnaire.questionnaire import OUTPUT_BOX_LAYOUT, QUESTION_BOX_LAYOUT, VIEWS_PACKAGE_PATH, \
+    OUTPUT_BOX_STYLE_FILE
 
 
 @fixture
@@ -14,7 +27,7 @@ def example_questionnaire_dict():
                     {
                         "answer": "Example answer",
                         "tooltip": "This is an example tooltip.",
-                        "templates": [{"replacement": "{{code}}", "code": "example_code {{follow}}"}],
+                        "templates": [{"replacement": "code", "code": "example_code {{follow}}"}],
                         "followups": [
                             {
                                 "question": "Follow-up question?",
@@ -22,23 +35,125 @@ def example_questionnaire_dict():
                                     {
                                         "answer": "Follow-up answer",
                                         "tooltip": "This is a follow-up tooltip.",
-                                        "templates": [{"replacement": "{{follow}}", "code": "followup_code"}],
+                                        "templates": [{"replacement": "follow", "code": "followup_code\n{{code}}"}],
                                         "followups": [],
                                         "commentary": "",
+                                    },
+                                    {
+                                        "answer": "Second follow-up answer",
+                                        "tooltip": "This is a second follow-up tooltip.",
+                                        "templates": [{"replacement": "follow", "code": "second_followup_code\n{{code}}"}],
+                                    }
+                                ],
+                            },
+                            {
+                                "question": "Another follow-up question?",
+                                "answers": [
+                                    {
+                                        "answer": "Another follow-up answer",
+                                        "templates": [],
                                     }
                                 ],
                             }
                         ],
-                        "commentary": "",
+                        "commentary": "This is some commentary.",
                     },
                     {
                         "answer": "Another answer",
                         "tooltip": "This is another tooltip.",
-                        "templates": [{"replacement": "{{code}}", "code": "another_code"}],
+                        "templates": [{"replacement": "code", "code": "another_code\n{{code}}"}],
                         "followups": [],
-                        "commentary": "",
+                        "commentary": "Some other commentary.",
                     },
+                ],
+            },
+            {
+                "question": "Second question?",
+                "answers": [
+                    {
+                        "answer": "Second answer",
+                        "tooltip": "This is a second tooltip.",
+                        "templates": [{"replacement": "code", "code": "second_code"}],
+                        "followups": [],
+                        "commentary": "Next commentary.",
+                    }
                 ],
             }
         ],
     }
+
+@fixture
+def example_questionnaire(example_questionnaire_dict):
+    return Questionnaire.model_validate(example_questionnaire_dict)
+
+class Helpers:
+
+    @staticmethod
+    def assert_widget_ui_matches_state(widget):
+        assert isinstance(widget.ui, HBox)
+        assert widget.ui.children == (widget.question_box, widget.output_box)
+
+        assert isinstance(widget.output_box, VBox)
+        assert widget.output_box.children == (widget.output_container,)
+        assert widget.output_box.layout == OUTPUT_BOX_LAYOUT
+
[…]
