+++ pyproject.toml
@@ -47,6 +47,7 @@ dev = [
     "pre-commit", # Used to run checks before finalizing a git commit
     "pytest",
     "pytest-cov", # Used to report total code coverage
+    "pytest-mock", # Used to mock objects in tests
     "ruff", # Used for static linting of files
 ]
 
