+++ src/scarlet2/questionnaire/__init__.py
@@ -1,1 +1,3 @@
 from .questionnaire import QuestionnaireWidget, run_questionnaire
+
+__all__ = ["QuestionnaireWidget", "run_questionnaire"]
