### 2025-08-28 58c5d7b
change min to max fft shape resampling renderer
