+++ tests/scarlet2/test_renderer.py
@@ -8,9 +8,10 @@ import warnings
 import astropy.io.fits as fits
 import jax
 import jax.numpy as jnp
-import scarlet2
 from astropy.wcs import WCS
 from huggingface_hub import hf_hub_download
+
+import scarlet2
 
 warnings.filterwarnings("ignore")
 
@@ -105,7 +106,7 @@ def test_hst_to_hsc_against_galsim():
     # Deconvolution, Resampling and Reconvolution
     hst_resampled = obs_hsc.render(data_hst)
 
-    assert jnp.allclose(out_galsim, hst_resampled[0], atol=1.3e-4)
+    assert jnp.allclose(out_galsim, hst_resampled[0], atol=2.3e-4)
 
 
 # Rotate WCS
@@ -145,7 +146,7 @@ def test_hst_to_hsc_against_galsim_rotated_wcs():
     # Deconvolution, Resampling and Reconvolution
     hst_resampled = obs_hsc.render(data_hst_rot)
 
-    assert jnp.allclose(out_galsim, hst_resampled[0], atol=1.3e-4)
+    assert jnp.allclose(out_galsim, hst_resampled[0], atol=2.3e-4)
 
 
 def test_no_channel_axis_in_obs_psf():
@@ -164,7 +165,7 @@ def test_no_channel_axis_in_obs_psf():
     # Deconvolution, Resampling and Reconvolution
     hst_resampled = obs_hsc.render(data_hst)
 
-    assert jnp.allclose(out_galsim, hst_resampled[0], atol=1.3e-4)
+    assert jnp.allclose(out_galsim, hst_resampled[0], atol=2.3e-4)
 
 
 if __name__ == "__main__":
