### 2025-08-28 97a9744
add test for run function
