### 2025-08-28 1e8cc04
Change `min` to `max`  FFT shape in resampling renderer (#215)
