### 2025-08-28 d0e656c
adding CorrelatedNoiseObservation class
