+++ src/scarlet2/__init__.py
@@ -25,7 +25,7 @@ from .morphology import (
     SersicMorphology,
     StarletMorphology,
 )
-from .observation import Observation
+from .observation import CorrelatedNoiseObservation, Observation
 from .psf import PSF, ArrayPSF, GaussianPSF
 from .scene import Scene
 from .source import Component, DustComponent, PointSource, Source
@@ -52,6 +52,7 @@ __all__ = [
     "SersicMorphology",
     "StarletMorphology",
     "Observation",
+    "CorrelatedNoiseObservation",
     "PSF",
     "ArrayPSF",
     "GaussianPSF",
