+++ .github/workflows/pre-commit-ci.yml
@@ -14,7 +14,7 @@ jobs:
   pre-commit-ci:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
+    - uses: actions/checkout@v5
       with:
         fetch-depth: 0 
     - name: Set up Python
