### 2025-09-03 c291c39
Update src/scarlet2/questionnaire/questions.yaml
Co-authored-by: Drew Oldag <47493171+drewoldag@users.noreply.github.com>
