+++ src/scarlet2/questionnaire/questions.yaml
@@ -20,7 +20,7 @@ questions:
                     code: "channels=[(\"g\",0)]
                            \nchannels=[(\"g\",1)]
                            \n\n{{obs}}"
-                commentary: "For each image, the channels should be a list of tuples where each tuple provides the filter and an epoch identifier which is unique for each image of a particular filter. For example, you might have two g-band images from different times, which would be initialized separately with `channels=[(‘g’,0)]` and `channels=[(‘g’,1’)]`. You might have two other r-band images taken at different times again. These would be initialized with `channels=[(‘r’,0)]` and `channels=[(‘r’,1’)]`. You may wish to take note of any epochs where the transient is known to be ‘off’ - you can add this constraint later if desired. Each image can be initialized"
+                commentary: "For each image, the channels should be a list of tuples where each tuple provides the filter and an epoch identifier which is unique for each image of a particular filter. For example, you might have two g-band images from different times, which would be initialized separately with `channels=[(‘g’,0)]` and `channels=[(‘g’,1’)]`. You might have two other r-band images taken at different times again. These would be initialized with `channels=[(‘r’,0)]` and `channels=[(‘r’,1’)]`. You may wish to take note of any epochs where the transient is known to be ‘off’ - you can add this constraint later if desired."
               - answer: "No"
                 templates:
                   - replacement: code
@@ -165,14 +165,14 @@ questions:
                       \n)
                       \n
                       \n# Select the transient-free epochs to initialize amplitudes for the static source. 
-                      \n#Make sure this matches the number of filters included in the modeling!
+                      \n# Make sure this matches the number of filters included in the modeling!
                       \n# These will be shared across all epochs
                       \nspectrum = spectrum[0:2]
                       \n
                       \nscarlet2.Source(
                       \n    center, scarlet2.StaticArraySpectrum(spectrum, bands=bands, band_selector=band_selector), sersic
                       \n)
-                      \n)
+                      \n
                       \n\n{{source_init}}"
         - value: null
           questions:
@@ -224,7 +224,7 @@ questions:
                       \nscarlet2.Source(
                       \n    center, spectrum, sersic
                       \n)
-                      \n)
+                      \n
                       \n\n{{source_init}}"
 
     - answer: "No"
