### 2025-09-03 8e4d7b0
Add Questionnaire Variable Feature (#216)
* add questionnaire variables feature

* lint

* Update src/scarlet2/questionnaire/questions.yaml typo

Co-authored-by: Drew Oldag <47493171+drewoldag@users.noreply.github.com>

* Update src/scarlet2/questionnaire/questions.yaml

Co-authored-by: Drew Oldag <47493171+drewoldag@users.noreply.github.com>

* Update src/scarlet2/questionnaire/questions.yaml add backticks

Co-authored-by: Drew Oldag <47493171+drewoldag@users.noreply.github.com>

* Update src/scarlet2/questionnaire/questions.yaml

Co-authored-by: Drew Oldag <47493171+drewoldag@users.noreply.g
[…]
