+++ tests/scarlet2/questionnaire/test_questionnaire.py
@@ -77,6 +77,201 @@ def test_questionnaire_complete_all_questions(example_questionnaire, helpers):
     helpers.assert_widget_ui_matches_state(widget)
 
 
+def test_questionnaire_switch_variable(example_questionnaire_with_switch, helpers):
+    """Test that switch-case logic in the questionnaire works correctly."""
+
+    # Test that the switch-case logic works correctly for the first case.
+    widget = QuestionnaireWidget(example_questionnaire_with_switch)
+
+    first_question = example_questionnaire_with_switch.questions[0]
+    first_answer = first_question.answers[0]
+    first_button = widget.question_box.children[1]  # First button after question label
+
+    first_button.click()
+
+    assert widget.code_output == first_answer.templates[0].code
+    assert widget.commentary == first_answer.commentary
+
+    assert widget.question_answers == [(first_question, 0)]
+    assert widget.variables == {first_question.variable: 0}
+
+    switch = example_questionnaire_with_switch.questions[1]
+    case = switch.cases[0]
+
+    assert widget.current_question == case.questions[0]
+    assert widget.questions_stack == case.questions[1:] + example_questionnaire_with_switch.questions[2:]
+
+    helpers.assert_widget_ui_matches_state(widget)
+
+    # Now test that the switch-case logic works correctly for the second case.
+    widget = QuestionnaireWidget(example_questionnaire_with_switch)
+    first_question = example_questionnaire_with_switch.questions[0]
+    second_answer = first_question.answers[1]
+    second_button = widget.question_box.children[2]  # Second button after question label
+
+    second_button.click()
+
+    assert widget.code_output == second_answer.templates[0].code
+    assert widget.commentary == second_answer.commentary
+    assert widget.question_answers == [(first_question, 1)]
+    assert widget.variables == {first_question.variable: 1}
+
+    switch = example_questionnaire_with_switch.questions[1]
+    case = switch.cases[1]
+
+    assert widget.current_question == case.questions[0]
+    assert widget.questions_stack == case.questions[1:] + example_questionnaire_with_switch.questions[2:]
+
+    helpers.assert_widget_ui_matches_state(widget)
+
+    # Test that the default case is used if no case matches.
+    widget = QuestionnaireWidget(example_questionnaire_with_switch)
+
+    first_question = example_questionnaire_with_switch.questions[0]
+    third_answer = first_question.answers[2]
+    third_button = widget.question_box.children[3]  # Third button after question label
+
+    third_button.click()
+
+    assert widget.code_output == third_answer.templates[0].code
+    assert widget.commentary == third_answer.commentary
+    assert widget.question_answers == [(first_question, 2)]
+    assert widget.variables == {first_question.variable: 2}
+
+    switch = example_questionnaire_with_switch.questions[1]
+    case = switch.cases[2]  # Default case where value is None
+
+    assert widget.current_question == case.questions[0]
+    assert widget.questions_stack == case.questions[1:] + example_questionnaire_with_switch.questions[2:]
+
+    helpers.assert_widget_ui_matches_state(widget)
+
+
+def test_questionnaire_followup_switch(example_questionnaire_with_followup_switch, helpers):
+    """Test that a switch based on a follow-up question works correctly."""
+
+    widget = QuestionnaireWidget(example_questionnaire_with_followup_switch)
+
+    # Select the first answer to the first question (which has a follow-up question)
+    first_question = example_questionnaire_with_followup_switch.questions[0]
+    first_answer = first_question.answers[0]
+    first_button = widget.question_box.children[1]  # First button after question label
+
+    first_button.click()
+
+    assert widget.code_output == first_answer.templates[0].code
+    assert widget.commentary == first_answer.commentary
+
+    assert widget.question_answers == [(first_question, 0)]
+
[…]

[…]
