### 2025-09-04 0bcd296
style prev questions + tooltip
