### 2025-09-04 929e547
style prev questions + tooltip
