### 2025-09-04 c2fa9cd
style prev questions + tooltip
