### 2025-09-05 ec158fc
refactor resource loading
