### 2025-09-05 bcb7319
refactor resource loading
