+++ src/scarlet2/questionnaire/questionnaire.py
@@ -18,77 +18,7 @@ QUESTIONS_FILE_NAME = "questions.yaml"
 VIEWS_PACKAGE_PATH = "scarlet2.questionnaire.views"
 OUTPUT_BOX_TEMPLATE_FILE = "output_box.html.jinja"
 OUTPUT_BOX_STYLE_FILE = "output_box.css"
-
-_PREV_TOOLTIP_CSS = HTML("""
-    <style>
-    /* Add overflow-x: hidden to the question box to prevent horizontal scrolling */
-    .question-box-container {
-        overflow-x: hidden;
-    }
-    .prev-item { 
-        position: relative; 
-        overflow: visible;
-    }
-    .prev-btn { 
-        width: auto; 
-        text-align: left; 
-        border: none;
-        background: none;
-        box-shadow: none;
-        color: #555;
-        text-decoration: none;
-        cursor: pointer;
-        font-size: 0.9em;
-        margin: 2px;
-        padding: 2px 4px;
-        overflow: hidden;
-        max-width: 100%; /* Ensure button doesn't exceed container width */
-        text-overflow: ellipsis; /* Add ellipsis for text that overflows */
-    }
-
-    /* Container for the tooltip */
-    .prev-item .tooltip {
-      position: absolute;
-      background: #333;
-      color: #fff;
-      padding: 6px 8px;
-      border-radius: 8px;
-      font-size: 12px;
-      line-height: 1.3;
-      box-shadow: 0 4px 12px rgba(0,0,0,.2);
-      white-space: normal;
-      min-width: 160px;
-      max-width: 320px;
-      opacity: 0;
-      visibility: hidden;
-      transition: opacity .12s ease;
-      pointer-events: none;
-      z-index: 9999;
-
-      /* Default position to the right of the button */
-      left: 100%;
-      top: 50%;
-      transform: translateY(-50%) translateX(8px);
-    }
-
-    /* When hovering over the container, show the tooltip */
-    .prev-item:hover .tooltip {
-      opacity: 1;
-      visibility: visible;
-    }
-
-    /* Arrow for the tooltip */
-    .prev-item .tooltip::before {
-      content: "";
-      position: absolute;
-      border-width: 6px;
-      border-style: solid;
-      left: -6px;
-      top: 50%;
-      transform: translateY(-50%);
-      border-color: transparent #333 transparent transparent;
-    }
-    """)
+QUESTION_BOX_STYLE_FILE = "question_box.css"
 
 
 QUESTION_BOX_LAYOUT = Layout(
@@ -113,11 +43,34 @@ class QuestionnaireWidget:
         self.initial_template = questionnaire.initial_template
         self.initial_commentary = questionnaire.initial_commentary
 
+        # Load resources once at initialization
+        self._load_resources()
+
         self._init_state()
         self._init_ui()
 
         self._render_output_box()
         self._render_next_question()
+
+    def _load_resources(self):
+        # Load question box CSS
+        question_box_css_file = files(VIEWS_PACKAGE_PATH).joinpath(QUESTION_BOX_STYLE_FILE)
+        with question_box_css_file.open("r") as f:
+            self.question_box_css = f.read()
+
+        # Load output box CSS
+        output_box_css_file = files(VIEWS_PACKAGE_PATH).joinpath(OUTPUT_BOX_STYLE_FILE)
+        with output_box_css_file.open("r") as f:
+            self.output_box_css = f.read()
+
+        # Load Jinja template
+        template_file = files(VIEWS_PACKAGE_PATH).joinpath(OUTPUT_BOX_TEMPLATE_FILE)
+        with template_file.open("r") as f:
+            template_source = f.read()
+        self.output_box_template = jinja2.Template(template_source)
+
+        # Create HTML widget with question box CSS
[…]
