+++ docs/3-questionnaire.md
@@ -1,0 +1,140 @@
+# Questionnaire
+
+## User Guide - Interactive Project Setup
+
+The scarlet2 questionnaire is an interactive tool designed to help you quickly set up a new scarlet2 project 
+by generating a customized code template that matches your use case.
+
+### Running the Questionnaire
+
+To run the questionnaire, you need to import and call the `run_questionnaire` function from the `scarlet2.questionnaire` module:
+
+```python
+from scarlet2.questionnaire import run_questionnaire
+run_questionnaire()
+```
+
+![Questionnaire Screenshot](_static/questionnaire_screenshot.png)
+
+This will launch an interactive widget in your Jupyter notebook that guides you through a series of questions about your project and your data.
+
+The questionnaire presents one question at a time with multiple-choice answers. For each answer you select, 
+the questionnaire will update the code template to match your choices and display some explanatory text to 
+help you understand the code being generated.
+
+### Using the Generated Template
+
+As you progress through the questionnaire, a code template is dynamically generated based on your answers.
+The code will outline the steps needed to set up your scarlet2 project, but it will require some manual editing
+to fill in specific details about your data. For example, the code will include definitions of example values 
+for variables like `channels` that you will need to replace with values that match your data, and will 
+reference variables like `data` and `psf` that you will need to change to use your actual data.
+
+In the bottom right of the code output, you can click the "📋 Copy" button at any time to copy the generated 
+code to your clipboard.
+
+> Note: The template is a starting point that you can modify to fit your specific data. It provides the structure for your project based on your use case.
+
+## Developer Guide - Questionnaire Architecture
+
+The questionnaire module is designed to be extensible and maintainable. This section explains how the questionnaire works internally and how to modify or extend it.
+
+### Module Structure
+
+The questionnaire module consists of several key components:
+
+1. **models.py**: Defines the data structures used by the questionnaire that map to the YAML file
+2. **questionnaire.py**: Contains the main `QuestionnaireWidget` class that uses ipywidgets to render the UI and handle user interactions
+3. **questions.yaml**: Stores the actual questions, answers, and templates
+
+### YAML Structure
+
+Setting up the questions in the questionnaire is done by modifying the `questions.yaml` file. 
+The questions are defined in this YAML file with the following structure:
+
+```yaml
+initial_template: '{{code}}'  # The starting template with placeholders
+initial_commentary: 'Welcome message'  # (Optional) The initial commentary text before any questions are answered
+questions:  # List of top-level questions
+  - question: "Question text" # Each question object has a question text, answers, and optionally a variable
+    variable: "variable_name"  # Optional variable to store the answer to be referenced later
+    answers:
+      - answer: "Answer option 1" # Each answer has an answer text
+        tooltip: "Helpful tooltip" # Optional tooltip for the answer that appears on hover of the button
+        templates:  # A list of code snippets to apply if selected
+          - replacement: code  # The placeholder to replace
+            code: "# Your code here\n{{next_placeholder}}"  # The replacement code
+        commentary: "Explanation of this choice" # The commentary text to display when this answer is selected, can include markdown formatting
+        followups:  # Additional questions to ask immediately if this answer is selected. This list of followups matches the structure of top-level questions, and can include question objects or switch/case objects
+          - question: "Follow-up question"
+            answers: [...]
+  - question: "..." # More questions
+    answers: [...]
+  # ...
+  - switch: "variable_name"  # Conditional branching based on a previous answer
+    cases:
+      - value: 0  # If the question with variable "variable_name" was answered with the first answer (index 0)
+        questions: [...] # The questions to ask in this case. This list matches the structure of top-level questions, and can include question objects or switch/case objects
[…]
