+++ src/scarlet2/questionnaire/questions.yaml
@@ -1,4 +1,5 @@
 initial_template: '{{code}}'
+feedback_url: "https://docs.google.com/forms/d/e/1FAIpQLScKHbiqxhizacgzRx3xHEdqqjgtZBsxjtQZFJlYBdLcbOnfBg/viewform"
 questions:
   - question: Do you have a single image, or multiple images?
     answers:
