+++ src/scarlet2/questionnaire/questions.yaml
@@ -114,6 +114,7 @@ questions:
         - value: 0
           questions:
           - question: Would you like to model galaxies with?
+            variable: "galaxy_model"
             answers:
             - answer: flexible non-parametric morphology models (starlets)
               templates:
@@ -180,6 +181,7 @@ questions:
         - value: null
           questions:
             - question: Would you like to model galaxies with?
+              variable: "galaxy_model"
               answers:
                 - answer: flexible non-parametric morphology models (starlets)
                   templates:
@@ -233,3 +235,178 @@ questions:
                       code: "import scarlet2.init as init\n{{imports}}"
 
     - answer: "No"
+  - question: Do you have any point sources to model?
+    variable: "point_sources"
+    answers:
+    - answer: "Yes"
+      followups:
+      - question: Are any of the point sources transient or variable?
+        answers:
+        - answer: "No"
+          commentary: "Point sources can be initialized as follows"
+          templates:
+          - replacement: source_init
+            code: "# 2) Point source for the transient
+                  \n# Define the epochs where the transient is allowed to have a non-zero amplitude (this could be all epochs)
+                  \nepochs = [2, 3]
+                  \n# As we already know that the transient is present, we can measure the flux at the center location
+                  \n# Initializing as zero also works
+                  \nspectrum = scarlet2.init.pixel_spectrum(obs, center)
+                  \n 
+                  \nscarlet2.PointSource(
+                  \n    center, scarlet2.TransientArraySpectrum(spectrum, epochs=epochs, epoch_selector=epoch_selector)
+                  \n)
+                  \n\n{{source_init}}"
+        - answer: "Yes"
+          commentary: "Variable point sources can be initialized as follows"
+          templates:
+          - replacement: source_init
+            code: "# 2) Point source for the transient
+                  \n# Define the epochs where the transient is allowed to have a non-zero amplitude (this could be all epochs)
+                  \nepochs = [2, 3]
+                  \n# As we already know that the transient is present, we can measure the flux at the center location
+                  \n# Initializing as zero also works
+                  \nspectrum = scarlet2.init.pixel_spectrum(obs, center)
+                  \n 
+                  \nscarlet2.PointSource(
+                  \n    center, scarlet2.TransientArraySpectrum(spectrum, epochs=epochs, epoch_selector=epoch_selector)
+                  \n)
+                  \n\n{{source_init}}"
+    - answer: "No"
+  - question: Click to continue
+    answers:
+    - answer: Continue
+      templates:
+        - replacement: source_init
+          code: "print(scene.sources)
+                \n\n{{parameters}}"
+      commentary: "After initializing all sources, You can check the final list of sources as follows"
+  - question: Click to continue
+    answers:
+    - answer: Continue
+      templates:
+        - replacement: parameters
+          code: "pos_step = 1e-2
+                \nmorph_step = lambda p: scarlet2.relative_step(p, factor=1e-3)
+                \nSED_step = lambda p: scarlet2.relative_step(p, factor=5e-2)
+
+                \nparameters = scene.make_parameters()
+                \n\n{{parameters}}"
+        - replacement: imports
+          code: "from numpyro.distributions import constraints\n{{imports}}"
+      commentary: "You can initialize parameters as follows"
+  - switch: galaxy_model
+    cases:
+    - value: 0
+      questions:
+      - question: Click to continue
+        answers:
+        - answer: Continue
+          commentary: "We set the parameters for the galaxy model."
+          templates:
+          - replacement: parameters
+            code: "parameters += scarlet2.Parameter(
[…]
