### 2025-09-08 8e4c934
add loading unit tests
