### 2025-09-08 dae828a
add loading unit tests
