+++ src/scarlet2/questionnaire/questionnaire.py
@@ -12,7 +12,12 @@ from pygments.formatters import HtmlFormatter
 from pygments.lexers import PythonLexer
 
 from scarlet2.questionnaire.models import (
-    Question, Questionnaire, QuestionAnswer, QuestionAnswers, Switch, Template
+    Question,
+    QuestionAnswer,
+    QuestionAnswers,
+    Questionnaire,
+    Switch,
+    Template,
 )
 
 PACKAGE_PATH = "scarlet2.questionnaire"
@@ -41,7 +46,12 @@ OUTPUT_BOX_LAYOUT = Layout(
 class QuestionnaireWidget:
     """A widget to run an interactive questionnaire in a Jupyter notebook."""
 
-    def __init__(self, questionnaire: Questionnaire, save_path: str | None = None, initial_answers: QuestionAnswers | None = None):
+    def __init__(
+        self,
+        questionnaire: Questionnaire,
+        save_path: str | None = None,
+        initial_answers: QuestionAnswers | None = None,
+    ):
         self.questions = questionnaire.questions
         self.initial_template = questionnaire.initial_template
         self.initial_commentary = questionnaire.initial_commentary
@@ -171,17 +181,18 @@ class QuestionnaireWidget:
             color = "green" if message_type == "success" else "orange"
 
             # Create message HTML
-            message_html = HTML(
-                f'<div class="save-message" style="color: {color}">{message_text}</div>'
-            )
+            message_html = HTML(f'<div class="save-message" style="color: {color}">{message_text}</div>')
             save_components = [message_html, save_button]
             self.save_message = None
 
         # Create a container for the save button
-        save_button_container = VBox(save_components, layout=Layout(
-            width="100%",
-            padding="0",
-        ))
+        save_button_container = VBox(
+            save_components,
+            layout=Layout(
+                width="100%",
+                padding="0",
+            ),
+        )
         save_button_container.add_class("save-button-container")
 
         if self.current_question is None:
@@ -202,7 +213,9 @@ class QuestionnaireWidget:
             # Wrap the final message in a container
             final_message_container = VBox([HTML(final_message)], layout=Layout(margin="0 0 0 0"))
 
-            self.question_box.children = [self.question_box_css_html] + previous_qs + [final_message_container, save_button_container]
+            self.question_box.children = (
+                [self.question_box_css_html] + previous_qs + [final_message_container, save_button_container]
+            )
             return
 
         q_label = HTML(f"<b>{self.current_question.question}</b>")
@@ -225,7 +238,9 @@ class QuestionnaireWidget:
         # Create a container for the buttons
         buttons_container = VBox(buttons, layout=Layout(margin="0"))
 
-        self.question_box.children = [self.question_box_css_html] + previous_qs + [q_label, buttons_container, save_button_container]
+        self.question_box.children = (
+            [self.question_box_css_html] + previous_qs + [q_label, buttons_container, save_button_container]
+        )
 
     def _generate_previous_questions(self):
         items = []
@@ -237,7 +252,7 @@ class QuestionnaireWidget:
                 description=f"{q.question} — {ans.answer}",
                 tooltip="Click to go back",  # native title tooltip as a fallback
                 layout=Layout(width="auto", margin="2px 0"),
-                style={'button_color': 'transparent', 'font_weight': 'normal'},
+                style={"button_color": "transparent", "font_weight": "normal"},
             )
             btn.add_class("prev-btn")
 
@@ -249,7 +264,7 @@ class QuestionnaireWidget:
             btn.on_click(on_click_handler)
 
             # The custom tooltip node (shown on hover via CSS)
-            tip_html = HTML(f"<div class='tooltip'>Click to go back</div>")
+            tip_html = HTML("<div class='tooltip'>Click to go back</div>")
 
             # Wrap button + tooltip in a positioned container
[…]
