+++ src/scarlet2/questionnaire/questionnaire.py
@@ -1,3 +1,4 @@
+import os
 import re
 from importlib.resources import files
 
@@ -10,7 +11,14 @@ from pygments import highlight
 from pygments.formatters import HtmlFormatter
 from pygments.lexers import PythonLexer
 
-from scarlet2.questionnaire.models import Question, Questionnaire, Switch, Template
+from scarlet2.questionnaire.models import (
+    Question,
+    QuestionAnswer,
+    QuestionAnswers,
+    Questionnaire,
+    Switch,
+    Template,
+)
 
 PACKAGE_PATH = "scarlet2.questionnaire"
 QUESTIONS_FILE_NAME = "questions.yaml"
@@ -18,6 +26,7 @@ QUESTIONS_FILE_NAME = "questions.yaml"
 VIEWS_PACKAGE_PATH = "scarlet2.questionnaire.views"
 OUTPUT_BOX_TEMPLATE_FILE = "output_box.html.jinja"
 OUTPUT_BOX_STYLE_FILE = "output_box.css"
+QUESTION_BOX_STYLE_FILE = "question_box.css"
 
 
 QUESTION_BOX_LAYOUT = Layout(
@@ -37,36 +46,88 @@ OUTPUT_BOX_LAYOUT = Layout(
 class QuestionnaireWidget:
     """A widget to run an interactive questionnaire in a Jupyter notebook."""
 
-    def __init__(self, questionnaire: Questionnaire):
+    def __init__(
+        self,
+        questionnaire: Questionnaire,
+        save_directory: str | None = None,
+        initial_answers: QuestionAnswers | None = None,
+    ):
         self.questions = questionnaire.questions
-        self.code_output = questionnaire.initial_template
-        self.commentary = questionnaire.initial_commentary
+        self.initial_template = questionnaire.initial_template
+        self.initial_commentary = questionnaire.initial_commentary
         self.feedback_url = questionnaire.feedback_url
+        self.save_directory = save_directory
 
-        self._init_questions()
+        self._load_resources()
+
+        self._init_state()
         self._init_ui()
 
-        self._render_output_box()
-        self._render_next_question()
+        if initial_answers is not None:
+            self._start_with_answers(initial_answers)
+        else:
+            self._render_output_box()
+            self._render_next_question()
 
-    def _init_questions(self):
+    def _load_resources(self):
+        question_box_css_file = files(VIEWS_PACKAGE_PATH).joinpath(QUESTION_BOX_STYLE_FILE)
+        with question_box_css_file.open("r") as f:
+            self.question_box_css = f.read()
+
+        output_box_css_file = files(VIEWS_PACKAGE_PATH).joinpath(OUTPUT_BOX_STYLE_FILE)
+        with output_box_css_file.open("r") as f:
+            self.output_box_css = f.read()
+
+        template_file = files(VIEWS_PACKAGE_PATH).joinpath(OUTPUT_BOX_TEMPLATE_FILE)
+        with template_file.open("r") as f:
+            template_source = f.read()
+        self.output_box_template = jinja2.Template(template_source)
+
+        self.question_box_css_html = HTML(f"<style>{self.question_box_css}</style>")
+
+    def _init_state(self):
         self.current_question = None
         self.questions_stack = []
         self.question_answers = []
         self.variables = {}
+        self.code_output = self.initial_template
+        self.commentary = self.initial_commentary
+        self.save_message = None
         self._add_questions_to_stack(self.questions)
 
     def _init_ui(self):
         self.output_container = HTML()
         self.question_box = VBox([], layout=QUESTION_BOX_LAYOUT)
+        # Add a class to the question box for CSS targeting
+        self.question_box.add_class("question-box-container")
         self.output_box = VBox([self.output_container], layout=OUTPUT_BOX_LAYOUT)
 
[…]
