+++ tests/scarlet2/questionnaire/test_questionnaire.py
@@ -1,3 +1,8 @@
+import os
+from pathlib import Path
+
+import yaml
+from ipywidgets import HTML, Button
 from scarlet2.questionnaire import QuestionnaireWidget, run_questionnaire
 from scarlet2.questionnaire.models import Questionnaire
 from scarlet2.questionnaire.questionnaire import load_questions
@@ -35,7 +40,7 @@ def test_questionnaire_handle_answer_selection(example_questionnaire, helpers):
 
     first_question = example_questionnaire.questions[0]
     first_answer = first_question.answers[0]
-    first_button = widget.question_box.children[1]  # First button after question label
+    first_button = helpers.get_answer_button(widget, 0)  # First answer button
 
     first_button.click()
 
@@ -69,7 +74,7 @@ def test_questionnaire_complete_all_questions(example_questionnaire, helpers):
         assert widget.question_answers == list(zip(expected_questions[:i], answer_inds[:i], strict=False))
         helpers.assert_widget_ui_matches_state(widget)
 
-        button = widget.question_box.children[i + 1 + ans_ind]
+        button = helpers.get_answer_button(widget, ans_ind)
         button.click()
 
     assert widget.current_question is None
@@ -85,7 +90,7 @@ def test_questionnaire_switch_variable(example_questionnaire_with_switch, helper
 
     first_question = example_questionnaire_with_switch.questions[0]
     first_answer = first_question.answers[0]
-    first_button = widget.question_box.children[1]  # First button after question label
+    first_button = helpers.get_answer_button(widget, 0)  # First answer button
 
     first_button.click()
 
@@ -107,7 +112,7 @@ def test_questionnaire_switch_variable(example_questionnaire_with_switch, helper
     widget = QuestionnaireWidget(example_questionnaire_with_switch)
     first_question = example_questionnaire_with_switch.questions[0]
     second_answer = first_question.answers[1]
-    second_button = widget.question_box.children[2]  # Second button after question label
+    second_button = helpers.get_answer_button(widget, 1)  # Second answer button
 
     second_button.click()
 
@@ -129,7 +134,7 @@ def test_questionnaire_switch_variable(example_questionnaire_with_switch, helper
 
     first_question = example_questionnaire_with_switch.questions[0]
     third_answer = first_question.answers[2]
-    third_button = widget.question_box.children[3]  # Third button after question label
+    third_button = helpers.get_answer_button(widget, 2)  # Third answer button
 
     third_button.click()
 
@@ -155,7 +160,7 @@ def test_questionnaire_followup_switch(example_questionnaire_with_followup_switc
     # Select the first answer to the first question (which has a follow-up question)
     first_question = example_questionnaire_with_followup_switch.questions[0]
     first_answer = first_question.answers[0]
-    first_button = widget.question_box.children[1]  # First button after question label
+    first_button = helpers.get_answer_button(widget, 0)  # First answer button
 
     first_button.click()
 
@@ -173,7 +178,7 @@ def test_questionnaire_followup_switch(example_questionnaire_with_followup_switc
     # Select the first answer to the follow-up question (which sets the switch variable)
 
     followup_answer = followup_question.answers[0]
-    followup_button = widget.question_box.children[2]  # first button after question label and previous qs
+    followup_button = helpers.get_answer_button(widget, 0)  # First answer button for the followup question
 
     followup_button.click()
 
@@ -203,7 +208,7 @@ def test_questionnaire_followup_switch(example_questionnaire_with_followup_switc
     # Select the first answer to the first question (which has a follow-up question)
     first_question = example_questionnaire_with_followup_switch.questions[0]
     first_answer = first_question.answers[0]
[…]

[…]
