+++ .github/workflows/pre-commit-ci.yml
@@ -18,7 +18,7 @@ jobs:
       with:
         fetch-depth: 0 
     - name: Set up Python
-      uses: actions/setup-python@v5
+      uses: actions/setup-python@v6
       with:
         python-version: '3.11'
     - name: Install dependencies
