### 2025-10-04 681ba34
correct treatment of WCS angle, flip, scale (fixes #218); fixed tests of renderer and moments
