+++ tests/scarlet2/test_renderer.py
@@ -6,6 +6,7 @@
 import warnings
 
 import astropy.io.fits as fits
+import astropy.units as u
 import jax
 import jax.numpy as jnp
 import scarlet2
@@ -123,7 +124,7 @@ wcs_hst_rot.wcs.crpix = [crpix1_new, crpix2_new]
 
 # # Mock a rotation of 90 deg counter-clockwise of the HST WCS
 M = get_affine(wcs_hst)
-phi = 90 / 180 * jnp.pi  # in rad
+phi = 90 * u.deg
 R = _rot_matrix(phi)
 if jnp.linalg.det(M) < 0:  # improper rotation: flip y, rotate, flip back
     F = jnp.diag(jnp.array((1, -1)))
