### 2025-10-04 48896cc
fixed pixels size comparison
