### 2025-10-05 69c7d11
removed units from angle and scale because not supported by jax;
one could use unxt, but with several new dependencies
