### 2025-10-05 f73188b
moved set_spectra_to_match to measure.forced_photometry
