+++ src/scarlet2/scene.py
@@ -1,3 +1,4 @@
+import logging
 from collections import defaultdict
 
 import equinox as eqx
@@ -8,7 +9,6 @@ from . import Scenery
 from .bbox import overlap_slices
 from .frame import Frame
 from .module import Module, Parameters
-from .renderer import ChannelRenderer
 from .validation_utils import (
     ValidationError,
     ValidationInfo,
@@ -354,94 +354,10 @@ class Scene(Module):
         return returned_scene
 
     def set_spectra_to_match(self, observations, parameters):
-        """Sets the spectra of every source in the scene to match the observations
-
-        Computes the best-fit amplitude of the rendered model of all components in every
-        channel of every observation as a linear inverse problem.
-
-        Parameters
-        ----------
-        observations: :py:class:`~scarlet2.Observation` or list
-            The observations used to set the spectra.
-        parameters: :py:class:`~scarlet2.Parameters`
-            Parameters to adjust. This method will ignore all spectrum parameters that are not arrays
-            or not included in this list.
-
-        Returns
-        -------
-        None
-        """
-
-        if not hasattr(observations, "__iter__"):
-            observations = (observations,)
-
-        # extract multi-channel model for every source
-        spectrum_parameters = []
-        models = []
-        for i, src in enumerate(self.sources):
-            # search for spectrum in parameters: only works for standard arrays
-            if isinstance(src.spectrum, jnp.ndarray):
-                for p in parameters:
-                    if p.node is src.spectrum:
-                        spectrum_parameters.append(i)
-                        # update source to have flat spectrum
-                        src = eqx.tree_at(lambda src: src.spectrum, src, jnp.ones_like(p.node))
-                        break
-
-            # evaluate the model for any source so that fit includes it even if its spectrum is not updated
-            model = self.evaluate_source(src)  # assumes all sources are single components
-
-            # check for models with identical initializations, see scarlet repo issue #282
-            # if duplicate: raise ValueError
-            for model_indx in range(len(models)):
-                if jnp.allclose(model, models[model_indx]):
-                    message = f"Source {i} has a model identical to source {model_indx}.\n"
-                    message += "This is likely not intended, and the second source should be deleted."
-                    raise ValueError(message)
-            models.append(model)
-
-        models = jnp.array(models)
-        num_models = len(models)
-
-        for obs in observations:
-            # independent channels, no mixing
-            # solve the linear inverse problem of the amplitudes in every channel
-            # given all the rendered morphologies
-            # spectrum = (M^T Sigma^-1 M)^-1 M^T Sigma^-1 * im
-            num_channels = obs.frame.C
-            images = obs.data
-            weights = obs.weights
-            morphs = jnp.stack([obs.render(model) for model in models], axis=0)
-            spectra = jnp.zeros((num_models, num_channels))
-            for c in range(num_channels):
-                im = images[c].reshape(-1)
-                w = weights[c].reshape(-1)
-                m = morphs[:, c, :, :].reshape(num_models, -1)
-                mw = m * w[None, :]
-                # check if all components have nonzero flux in c.
-                # because of convolutions, flux can be outside the box,
-                # so we need to compare weighted flux with unweighted flux,
-                # which is the same (up to a constant) for constant weights.
-                # so we check if *most* of the flux is from pixels with non-zero weight
-                nonzero = jnp.sum(mw, axis=1) / jnp.sum(m, axis=1) / jnp.mean(w) > 0.1
-                nonzero = jnp.flatnonzero(nonzero)
-                if len(nonzero) == num_models:
[…]
