### 2025-10-05 ce65c15
removed check string from validation output
