### 2025-10-05 487083d
moved source checks to Scene.__exit__
