+++ src/scarlet2/module.py
@@ -6,6 +6,15 @@ import jax.tree_util as jtu
 import varname
 from astropy.coordinates import SkyCoord
 
+from . import Parameterization
+from .validation_utils import (
+    ValidationError,
+    ValidationInfo,
+    ValidationMethodCollector,
+    ValidationResult,
+    print_validation_results,
+)
+
 
 class Module(eqx.Module):
     """Scarlet2 base module
@@ -123,6 +132,14 @@ class Parameter:
         self.prior = prior
         self.stepsize = stepsize
 
+        # add this source to the active scene
+        try:
+            Parameterization.parameters += self
+        except AttributeError:
+            print("A Parameter instance can only be created within the context of Parameters")
+            print("Use 'with Parameters(scene) as p: Parameter(...)'")
+            raise
+
     def apply_constraint(self):
         """Transform the value of the parameter to the unconstrained region"""
 
@@ -195,6 +212,25 @@ class Parameters:
         self._base_leaves = jtu.tree_leaves(base)
         self._params = list()
         self._leave_idx = list()
+
+    def __enter__(self):
+        # context manager to register sources
+        # purpose is to provide scene.frame to source inits that will need some of its information
+        # also allows us to append the sources automatically to the scene
+        Parameterization.parameters = self
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        Parameterization.parameters = None
+
+        # (re)-import `VALIDATION_SWITCH` at runtime to avoid using a static/old value
+        from .validation_utils import VALIDATION_SWITCH
+
+        if VALIDATION_SWITCH:
+            from .validation import check_parameters
+
+            validation_results = check_parameters(self)
+            print_validation_results("Parameters validation results", validation_results)
 
     def __repr__(self):
         # equinox-like formatting
@@ -368,3 +404,95 @@ def relative_step(x, *args, factor=0.01, minimum=1e-6):
         factor*norm(x), or `minimum`, whichever is larger.
     """
     return jnp.maximum(minimum, factor * jnp.linalg.norm(x))
+
+
+class ParameterValidator(metaclass=ValidationMethodCollector):
+    """A class containing all of the validation checks for a Scene object.
+    A convenience function exists that will run all the checks in this class and
+    returns a list of validation results:
+
+    :py:func:`~scarlet2.validation.check_scene`.
+    """
+
+    def __init__(self, parameter: Parameter):
+        """Initialize the SceneValidator.
+
+        Parameters
+        ----------
+        parameter : Parameter
+            The parameter to validate.
+        """
+        self.parameter = parameter
+
+    def check_constrained_parameter_is_feasible(self) -> list[ValidationResult]:
+        """Check that a constrained parameter has a feasible value.
+
+        Returns
+        -------
+        list[ValidationResult]
+            A list of validation results, which can be either `ValidationInfo`
+            or `ValidationError`.
+        """
+        validation_results: list[ValidationResult] = []
+        param = self.parameter
+        constraint_is_none = param.constraint is None
+        if not constraint_is_none:
+            is_feasible = param.constraint.check(param.node)
+        if constraint_is_none or (not constraint_is_none and is_feasible.all()):
+            validation_results.append(
+                ValidationInfo(
+                    f"Parameter {param.name} value is feasible.",
+                    check=self.__class__.__name__,
+                )
+            )
+        else:
+            validation_results.append(
+                ValidationError(
+                    f"Parameter {param.name} value is infeasible.",
+                    check=self.__class__.__name__,
+                    context={
+                        "name": param.name,
+                        "constraint": param.constraint,
+                        "feasible": is_feasible,
+                    },
+                )
+            )
+        return validation_results
+
+    def check_parameter_has_necessary_fields(self) -> list[ValidationResult]:
[…]
