### 2025-10-06 965a17a
separated fitting and sampling test
