### 2025-10-06 a7646a8
Validation (#226)
* removed check string from validation output

* moved source checks to Scene.__exit__

* Parameters work as context; parameter test activated

* use Parameters context

* allowed backward compatible declaration of Parameters

* separated fitting and sampling test
