+++ src/scarlet2/scene.py
@@ -105,6 +105,15 @@ class Scene(Module):
     def __exit__(self, exc_type, exc_value, traceback):
         Scenery.scene = None
 
+        # (re)-import `VALIDATION_SWITCH` at runtime to avoid using a static/old value
+        from .validation_utils import VALIDATION_SWITCH
+
+        if VALIDATION_SWITCH:
+            from .validation import check_scene
+
+            validation_results = check_scene(self)
+            print_validation_results("Source validation results", validation_results)
+
     def sample(
         self, observations, parameters, seed=0, num_warmup=100, num_samples=200, progress_bar=True, **kwargs
     ):
@@ -647,118 +656,6 @@ class FitValidator(metaclass=ValidationMethodCollector):
                         f"The chi-square in the border for source {i} is poor.",
                         check=self.__class__.__name__,
                         context={"chi2_border": chi2_outside, "source": i},
-                    )
-                )
-
-        return validation_results
-
-
-class SceneValidator(metaclass=ValidationMethodCollector):
-    """A class containing all of the validation checks for a Scene object.
-    A convenience function exists that will run all the checks in this class and
-    returns a list of validation results:
-
-    :py:func:`~scarlet2.validation.check_scene`.
-    """
-
-    def __init__(self, scene: Scene, parameters: Parameters, observation):
-        """Initialize the SceneValidator.
-
-        Parameters
-        ----------
-        scene : Scene
-            The scene object to validate.
-        parameters : Parameters
-            The parameters of the scene to validate.
-        observation : Observation
-            The observation object containing the data to validate against.
-        """
-        self.scene = scene
-        self.parameters = parameters
-        self.observation = observation
-
-    def check_source_boxes_comparable_to_observation(self) -> list[ValidationResult]:
-        """Check that the bounding boxes of all sources are comparable to the observation.
-
-        This checks that the bounding boxes of all sources are contained within the
-        observation's bounding box.
-
-        Returns
-        -------
-        list[ValidationResult]
-            A list of validation results, which can be either `ValidationInfo` or
-            `ValidationWarning`.
-        """
-        validation_results: list[ValidationResult] = []
-        for i, source in enumerate(self.scene.sources):
-            source_and_observation_box = source.bbox & self.observation.frame.bbox
-            if source_and_observation_box == source.bbox:
-                validation_results.append(
-                    ValidationInfo(
-                        f"Source {i} has a bounding box inside the observation.",
-                        check=self.__class__.__name__,
-                        context={"bbox": source.bbox, "observation_bbox": self.observation.frame.bbox},
-                    )
-                )
-            elif source_and_observation_box.spatial.get_extent() == [0, 0, 0, 0]:
-                validation_results.append(
-                    ValidationWarning(
-                        f"Source {i} has a bounding box outside of the observation.",
-                        check=self.__class__.__name__,
-                        context={"bbox": source.bbox, "observation_bbox": self.observation.frame.bbox},
-                    )
-                )
-
-            else:
-                validation_results.append(
-                    ValidationWarning(
-                        f"Source {i} has a bounding box that is partially inside the observation.",
-                        check=self.__class__.__name__,
-                        context={"bbox": source.bbox, "observation_bbox": self.observation.frame.bbox},
-                    )
-                )
-
-        return validation_results
-
-    def check_parameters_have_necessary_fields(self) -> list[ValidationResult]:
-        """Check that all parameters in the scene have the necessary fields set.
-
-        This checks that all parameters in the scene have the `prior` or `stepsize`
-        attributes set, but not both.
-
-        Returns
-        -------
-        list[ValidationResult]
-            A list of validation results, which can be either `ValidationInfo`
-            or `ValidationError`.
-        """
-        validation_results: list[ValidationResult] = []
-        for i in range(len(self.parameters)):
[…]
