### 2025-10-08 73bc7ab
corrected plotting centers and boxes for WCS
