### 2025-10-09 0e8f3ae
create dummy WCS if not set
