+++ src/scarlet2/plot.py
@@ -703,11 +703,11 @@ def sources(
     show_rendered=False,
     show_spectrum=True,
     model_mask=None,
-    add_markers=True,
+    add_labels=False,
     add_boxes=False,
     fig_kwargs=None,
     title_kwargs=None,
-    marker_kwargs=None,
+    label_kwargs=None,
     box_kwargs=None,
 ):
     """Plot all sources in `scene`
@@ -736,16 +736,15 @@ def sources(
         Whether to show the spectrum of each source
     model_mask: array, optional
         A mask to apply to the model. If not given, no mask is applied
-    add_markers: bool, optional
-        Whether to plot a marker at the center for each source
-        Requires the source to have a `center` attribute.
+    add_labels: bool, optional
+        Whether each source is labeled with its numerical index in the source list
     add_boxes: bool, optional
         Whether to plot the bounding box of each source
     fig_kwargs: dict, optional
         Additional arguments for `mpl.subplots`
     title_kwargs: dict, optional
         Additional arguments for `mpl.set_title`
-    marker_kwargs: dict, optional
+    label_kwargs: dict, optional
         Additional arguments for `mpl.plot` of the source centers. Defaults to
         {"color": "w", "marker": "x", "mew": 1, "ms": 10}
     box_kwargs: dict, optional
@@ -760,8 +759,8 @@ def sources(
         fig_kwargs = {}
     if title_kwargs is None:
         title_kwargs = {}
-    if marker_kwargs is None:
-        marker_kwargs = {"color": "w", "marker": "x", "mew": 1, "ms": 10}
+    if label_kwargs is None:
+        label_kwargs = {"color": "w", "ha": "center", "va": "center"}
     if box_kwargs is None:
         box_kwargs = {"facecolor": "none", "edgecolor": "w", "lw": 0.5}
 
@@ -791,13 +790,13 @@ def sources(
                 origin="lower",
             )
             ax[k][panel].set_title(f"Model Source {k}", **title_kwargs)
-            if add_markers:
+            if add_labels:
                 center = src.center
-                ax[k][panel].plot(*(center[::-1]), **marker_kwargs)  # needs x,y
+                ax[k][panel].text(*(center[::-1]), k, **label_kwargs)  # x,y
             panel += 1
 
         if show_rendered or show_observed:
-            if add_markers:
+            if add_labels:
                 center_obs = observation.frame.get_pixel(scene.frame.get_sky_coord(center)).flatten()
             if add_boxes:
                 start, stop = src.bbox.spatial.start, src.bbox.spatial.stop
@@ -816,8 +815,8 @@ def sources(
                 origin="lower",
             )
             ax[k][panel].set_title(f"Model Source {k} Rendered", **title_kwargs)
-            if add_markers:
-                ax[k][panel].plot(*(center_obs[::-1]), **marker_kwargs)
+            if add_labels:
+                ax[k][panel].text(*(center_obs[::-1]), k, **label_kwargs)  # x,y
             if add_boxes:
                 poly = Polygon(corners_obs[:, ::-1], closed=True, **box_kwargs)
                 ax[k][panel].add_artist(poly)
@@ -830,8 +829,8 @@ def sources(
                 origin="lower",
             )
             ax[k][panel].set_title("Observation".format(), **title_kwargs)
-            if add_markers:
-                ax[k][panel].plot(*(center_obs[::-1]), **marker_kwargs)
+            if add_labels:
+                ax[k][panel].text(*(center_obs[::-1]), k, **label_kwargs)  # x,y
             if add_boxes:
                 poly = Polygon(corners_obs[:, ::-1], closed=True, **box_kwargs)
                 ax[k][panel].add_artist(poly)
@@ -961,11 +960,6 @@ def scene(
         panel = 0
[…]
