### 2025-10-09 12d9489
added check for sub-pixel shift
