### 2025-10-09 8963cef
added moment normalization
