### 2025-10-10 6891e7e
Correct rotations even when improper (#225)
* correct treatment of WCS angle, flip, scale (fixes #218); fixed tests of renderer and moments

* make Moments fluent

* angle and scale get astropy units

* fixed pixels size comparison

* removed units from angle and scale because not supported by jax;

one could use unxt, but with several new dependencies

* moved set_spectra_to_match to measure.forced_photometry

* removed set_spectra_to_match

* corrected plotting centers and boxes for WCS

* create dummy WCS if not set

* minor fixes

* added check for sub-pixel shift

* added moment normaliza
[…]
