### 2025-10-10 b4e8408
spectrum normalization switched back on; docs updates (#227)
