+++ src/scarlet2/renderer.py
@@ -187,7 +187,7 @@ class TrimSpatialBox(Renderer):
         )
 
     def __call__(self, model, key=None):
-        """What to run when AdjustToFrame is called"""
+        """What to run when TrimSpatialBox is called"""
         sub = model[:, self.slices[-2].get_slice(), self.slices[-1].get_slice()]
         return sub
 
@@ -196,14 +196,15 @@ class ResamplingRenderer(Renderer):
     """Renderer to resample image to different pixel grid (subpixel position, resolution, orientation)"""
 
     padding: int
-    fft_shape_target: int = eqx.field(repr=False)
-    fft_shape_model_im: int = eqx.field(repr=False)
+    shape_obs: tuple = eqx.field(repr=False)
+    shape_obs_k: int = eqx.field(repr=False)
+    shape_model_k: int = eqx.field(repr=False)
     scale: float
     angle: float
     flip: int
-    model_kpsf_interp: jnp.array = eqx.field(repr=False)
-    obs_kpsf_interp: jnp.array = eqx.field(repr=False)
-    real_shape_target: tuple = eqx.field(repr=False)
+    shift: tuple
+    psf_model_k: jnp.array = eqx.field(repr=False)
+    psf_obs_k: jnp.array = eqx.field(repr=False)
 
     def __init__(self, model_frame, obs_frame, padding=4):
         """Initialize preprocess renderer in multi-resolution mapping
@@ -230,9 +231,6 @@ class ResamplingRenderer(Renderer):
 
         # create PSF model
         psf_model = model_frame.psf()
-        if len(psf_model.shape) == 2:
-            psf_model = psf_model[None, ...]
-
         if len(psf_model.shape) == 2:  # only one image for all bands
             psf_model = jnp.tile(psf_model, (obs_frame.bbox.shape[0], 1, 1))
 
@@ -240,90 +238,92 @@ class ResamplingRenderer(Renderer):
         if len(psf_obs.shape) == 2:
             psf_obs = psf_obs[None, ...]
 
-        self.fft_shape_model_im = good_fft_size(padding * max(model_frame.bbox.shape))
-        fft_shape_model_psf = good_fft_size(padding * max(psf_model.shape))
-        fft_shape_obs_psf = good_fft_size(padding * max(psf_obs.shape))
+        self.shape_obs = obs_frame.bbox.shape
+        self.shape_obs_k = good_fft_size(padding * max(obs_frame.bbox.spatial.shape))
+        # if self.shape_obs_k % 2 == 0:
+        #     self.shape_obs_k += 1
+        self.shape_model_k = good_fft_size(padding * max(model_frame.bbox.spatial.shape))
+        # if self.shape_model_k % 2 == 0:
+        #     self.shape_model_k += 1
 
-        # Fourier transform model and observation PSFs
-        model_kpsf = jnp.fft.fftshift(
-            transform(psf_model, (fft_shape_model_psf, fft_shape_model_psf), (-2, -1)), (-2)
+        # Fourier transform of model and observation PSFs
+        self.psf_model_k = jnp.fft.fftshift(
+            transform(psf_model, (self.shape_model_k, self.shape_model_k), (-2, -1)), (-2)
         )
-        obs_kpsf = jnp.fft.fftshift(
-            transform(psf_obs, (fft_shape_obs_psf, fft_shape_obs_psf), (-2, -1)), (-2)
+        self.psf_obs_k = jnp.fft.fftshift(
+            transform(psf_obs, (self.shape_obs_k, self.shape_obs_k), (-2, -1)), (-2)
         )
 
-        # Get maximum of the fft shapes to interpolate on the highest resolved FFT image
-        # odd shape is required for k-wrapping later
-        self.fft_shape_target = max(self.fft_shape_model_im, fft_shape_obs_psf) + 1
[…]
