### 2025-10-21 b7edfe0
insert_into as standalone function
