### 2025-10-21 b477d7e
added Moments.convolve (+test); increased ProfileMorphology shape to 15*size
