### 2025-10-21 bc10c88
undid my changes to resampling renderer
