+++ src/scarlet2/renderer.py
@@ -198,15 +198,15 @@ class ResamplingRenderer(Renderer):
     """Renderer to resample image to different pixel grid (subpixel position, resolution, orientation)"""
 
     padding: int
-    shape_obs: tuple = eqx.field(repr=False)
-    shape_obs_k: int = eqx.field(repr=False)
-    shape_model_k: int = eqx.field(repr=False)
+    fft_shape_target: int = eqx.field(repr=False)
+    fft_shape_model_im: int = eqx.field(repr=False)
     scale: float
     angle: float
     flip: int
     shift: tuple
-    psf_model_k: jnp.array = eqx.field(repr=False)
-    psf_obs_k: jnp.array = eqx.field(repr=False)
+    model_kpsf_interp: jnp.array = eqx.field(repr=False)
+    obs_kpsf_interp: jnp.array = eqx.field(repr=False)
+    real_shape_target: tuple = eqx.field(repr=False)
 
     def __init__(self, model_frame, obs_frame, padding=4):
         """Initialize preprocess renderer in multi-resolution mapping
@@ -240,25 +240,21 @@ class ResamplingRenderer(Renderer):
         if len(psf_obs.shape) == 2:
             psf_obs = psf_obs[None, ...]
 
-        self.shape_obs = obs_frame.bbox.shape
-        self.shape_obs_k = good_fft_size(padding * max(obs_frame.bbox.spatial.shape))
-        # if self.shape_obs_k % 2 == 0:
-        #     self.shape_obs_k += 1
-        self.shape_model_k = good_fft_size(padding * max(model_frame.bbox.spatial.shape))
-        # if self.shape_model_k % 2 == 0:
-        #     self.shape_model_k += 1
+        self.fft_shape_model_im = good_fft_size(padding * max(model_frame.bbox.shape))
+        fft_shape_model_psf = good_fft_size(padding * max(psf_model.shape))
+        fft_shape_obs_psf = good_fft_size(padding * max(psf_obs.shape))
 
-        # Fourier transform of model and observation PSFs
-        self.psf_model_k = jnp.fft.fftshift(
-            transform(psf_model, (self.shape_model_k, self.shape_model_k), (-2, -1)), (-2)
+        # Fourier transform model and observation PSFs
+        model_kpsf = jnp.fft.fftshift(
+            transform(psf_model, (fft_shape_model_psf, fft_shape_model_psf), (-2, -1)), (-2)
         )
-        self.psf_obs_k = jnp.fft.fftshift(
-            transform(psf_obs, (self.shape_obs_k, self.shape_obs_k), (-2, -1)), (-2)
+        obs_kpsf = jnp.fft.fftshift(
+            transform(psf_obs, (fft_shape_obs_psf, fft_shape_obs_psf), (-2, -1)), (-2)
         )
 
-        # # Get maximum of the fft shapes to interpolate on the highest resolved FFT image
-        # # odd shape is required for k-wrapping later
-        # self.shape_obs_k = max(self.shape_model_k, fft_shape_obs_psf) + 1
+        # Get maximum of the fft shapes to interpolate on the highest resolved FFT image
+        # odd shape is required for k-wrapping later
+        self.fft_shape_target = max(self.fft_shape_model_im, fft_shape_obs_psf) + 1
 
         # Extract rotation angle, flip, scale between WCSs
         M_in = get_affine(model_frame.wcs)  # noqa: N806
@@ -270,63 +266,70 @@ class ResamplingRenderer(Renderer):
         obs_center = obs_frame.bbox.spatial.center
         self.shift = tuple(model_center_in_obs[c].item() - obs_center[c] for c in range(2))
 
-        # self.model_kpsf_interp = resample_ops(
-        #     model_kpsf,
-        #     model_kpsf.shape[-2],
[…]
