+++ tests/scarlet2/test_renderer.py
@@ -14,12 +14,11 @@ from huggingface_hub import hf_hub_download
 
 import scarlet2
 from scarlet2.frame import _rot_matrix, get_affine
-from scarlet2.validation_utils import set_validation
 
 warnings.filterwarnings("ignore")
 
 # turn off automatic validation checks
-set_validation(False)
+scarlet2.set_validation(False)
 
 # Load the HSC image data
 # load test data from HSC and HST
@@ -112,7 +111,7 @@ def test_hst_to_hsc_against_galsim():
     # Deconvolution, Resampling and Reconvolution
     hst_resampled = obs_hsc.render(data_hst)
 
-    assert jnp.allclose(out_galsim, hst_resampled[0], atol=3)  # 2.3e-4)
+    assert jnp.allclose(out_galsim, hst_resampled[0], atol=2.3e-4)
 
 
 # Rotate WCS
@@ -156,7 +155,7 @@ def test_hst_to_hsc_against_galsim_rotated_wcs():
     # Deconvolution, Resampling and Reconvolution
     hst_resampled = obs_hsc.render(data_hst_rot)
 
-    assert jnp.allclose(out_galsim, hst_resampled[0], atol=3)  # 2.3e-4)
+    assert jnp.allclose(out_galsim, hst_resampled[0], atol=2.3e-4)
 
 
 if __name__ == "__main__":
