### 2025-10-30 6cd230f
added Moment methods; redefined center shift for ProfileMorphology accordingly
