+++ src/scarlet2/measure.py
@@ -5,7 +5,7 @@ import copy
 import numpy as jnp
 import numpy.ma as ma
 
-from .frame import get_affine, get_scale_angle_flip
+from .frame import get_affine, get_scale_angle_flip_shift
 from .source import Component
 
 
@@ -196,14 +196,16 @@ class Moments(dict):
                 self[m, n - m] = (grid_y**m * grid_x ** (n - m) * model * weight).sum(axis=(-2, -1))
 
             if n == 1:
+                self._centroid = jnp.array((self[1, 0] / self[0, 0], self[0, 1] / self[0, 0]))
                 # shift grid to produce centered moments
                 if center is None:
-                    center = self[1, 0] / self[0, 0], self[0, 1] / self[0, 0]  # centroid
+                    center = self._centroid
                     self[1, 0] = jnp.zeros_like(self[1, 0])
                     self[0, 1] = jnp.zeros_like(self[0, 1])
                 else:
                     center = jnp.asarray(center)
                     # centroid wrt given center
+                    self._centroid -= center * self[0, 0]
                     self[1, 0] -= center[0] * self[0, 0]
                     self[0, 1] -= center[1] * self[0, 0]
                 if model.ndim == 3 and center[0].ndim == 1:
@@ -223,6 +225,16 @@ class Moments(dict):
         return max(key[0] for key in self.keys())
 
     @property
+    def flux(self):
+        """Determine flux from 0th moment
+
+        Returns
+        -------
+        float
+        """
+        return self[0, 0]
+
+    @property
     def centroid(self):
         """Determine centroid from moments
 
@@ -231,7 +243,7 @@ class Moments(dict):
         array
             Coordinates of the centroid in the pixel frame of the data that defines these moments
         """
-        return jnp.array((self[1, 0] / self[0, 0], self[0, 1] / self[0, 0]))
+        return self._centroid
 
     @property
     def size(self):
@@ -353,6 +365,8 @@ class Moments(dict):
         This operation arises when one adjust the moments for a change in the size
         of pixels of the defining frame, e.g. when asking "what would the moments
         be if the pixels were factor c smaller (or the source c times larger)"?
+        The fluxes are unchanged, which corresponds to preservation of photons under resizing.
+
         The moments are changed in place.
 
         See Teague (1980), "Image  analysis via the general  theory of moments", eq. 34 for details.
@@ -366,15 +380,47 @@ class Moments(dict):
         -------
         self
         """
-        # Teague (1980), eq. 34
         if jnp.isscalar(c):
+            assert c > 0
+            flux_change = c**2
             for e in self:
-                self[e] = self[e] * c ** (2 + e[0] + e[1])
+                self[e] = self[e] * c ** (2 + e[0] + e[1]) / flux_change
         elif len(c) == 2:
+            assert c[0] > 0 and c[1] > 0
+            flux_change = c[0] * c[1]
             for e in self:
-                self[e] = self[e] * c[0] ** (e[0] + 1) * c[1] ** (e[1] + 1)
+                self[e] = self[e] * c[0] ** (e[0] + 1) * c[1] ** (e[1] + 1) / flux_change
         else:
             raise AttributeError("c must be a scalar of a list or array of two components")
+
+        return self
+
+    def fliplr(self):
+        """Flip moments along the x-axis
+
+        The moments are changed in place.
+
+        Returns
+        -------
+        self
+        """
+        for e in self:
+            if e[1] % 2 == 1:
+                self[e] *= -1
+        return self
+
+    def flipud(self):
+        """Flip moments along the y-axis
+
+        The moments are changed in place.
+
+        Returns
+        -------
+        self
+        """
+        for e in self:
+            if e[0] % 2 == 1:
+                self[e] *= -1
         return self
 
     def rotate(self, phi):
[…]
