### 2025-10-30 3a937ad
introduced translation to ResamplingRenderer; whole slew of minor corrections of Frame/WCS
