+++ src/scarlet2/frame.py
@@ -29,9 +29,7 @@ class Frame(eqx.Module):
         self.bbox = bbox
         self.psf = psf
         if wcs is None:
-            wcs = astropy.wcs.WCS(naxis=2)  # Dummy WCS
-            wcs._naxis = bbox.spatial.shape[::-1]
-            wcs.wcs.ctype = ["RA---TAN", "DEC--TAN"]
+            wcs = _wcs_default(bbox.spatial.shape)
         self.wcs = wcs
 
         if channels is None:
@@ -70,9 +68,10 @@ class Frame(eqx.Module):
         pixel coordinates in the model frame
         """
         if isinstance(pos, SkyCoord):
-            wcs_ = self.wcs.celestial  # only use celestial portion
-            pixel = jnp.asarray(pos.to_pixel(wcs_), dtype="float32").T
-            return pixel[..., ::-1]
+            # WCS uses x/y convention
+            wcs = self.wcs.celestial  # only use celestial portion
+            x, y = pos.to_pixel(wcs)
+            pos = jnp.stack([y, x], axis=-1)  # convert to y/x
         return pos
 
     def get_sky_coord(self, pos):
@@ -87,9 +86,10 @@ class Frame(eqx.Module):
         ----------
         astropy.coordinates.SkyCoord
         """
-        pixels = pos.reshape(-1, 2)
         wcs = self.wcs.celestial  # only use celestial portion
-        sky_coord = SkyCoord.from_pixel(pixels[:, 1], pixels[:, 0], wcs)
+        if jnp.ndim(pos) > 1:
+            pos = jnp.asarray(pos).reshape(-1, 2).T.reshape(2, *jnp.shape(pos)[:-1])
+        sky_coord = SkyCoord.from_pixel(pos[1], pos[0], wcs)
         return sky_coord
 
     def convert_pixel_to(self, target, pixel=None):
@@ -303,42 +303,87 @@ def get_psf_size(psf):
     return sigma3
 
 
-def get_affine(wcs):
-    """Return the WCS transformation matrix"""
+def get_affine(wcs=None, linear=True):
+    """Return the WCS transformation matrix
+
+    The transformation to intermediate world coordinates is given by the equation
+    $q = M\\cdot (p - r)$, where $p$ is the pixel coordinate, $r$ is `CRPIX`, and $M$ is the `CD` matrix.
+
+    This method provides the augmented matrix of the affine transformation:
+    $T = \begin{bmatrix} M & -M\\cdot r\\ 0 & 1\\end{bmatrix}$, for the extended vector $(p,1)$.
+
+    See Greisen & Calabretta (2002) for details.
+
+    Parameters
+    ----------
+    wcs: `astropy.wcs.WCS`
+        WCS structure
+    linear: `bool`
+        Return only linear 2x2 matrix
+
+    Returns
+    -------
+    array (3x3 or 2x2)
+    """
     if wcs is None:
-        return jnp.diag(jnp.ones(2))
+        return jnp.diag(jnp.ones(3))
     wcs_ = wcs.celestial
     try:
-        model_affine = wcs_.wcs.pc
+        m = wcs_.wcs.pc
     except AttributeError:
         try:
-            model_affine = wcs_.cd
+            m = wcs_.cd
         except AttributeError:
-            model_affine = wcs_.wcs.cd
-    return model_affine[:2, :2]
+            m = wcs_.wcs.cd
+    m = m[:2, :2]  # avoid using channel information that is not declared "spectral" in the WCS
+    if linear:
+        return m
[…]
