### 2025-10-30 b7baeb1
use jacobian only for resample_fourier
