### 2025-10-30 b1894b2
made PSF (de)convolution optional in ResamplingRenderer
