+++ tests/scarlet2/test_renderer.py
@@ -3,161 +3,190 @@
 # ruff: noqa: D103
 # ruff: noqa: D106
 
-import warnings
-
-import astropy.io.fits as fits
 import astropy.units as u
-import jax
 import jax.numpy as jnp
-from astropy.wcs import WCS
-from huggingface_hub import hf_hub_download
+from numpy.testing import assert_allclose
 
 import scarlet2
-from scarlet2.frame import _rot_matrix, get_affine
+from scarlet2 import ArrayPSF
+from scarlet2.frame import _flip_matrix, _rot_matrix, _wcs_default, get_affine
+from scarlet2.measure import Moments
+from scarlet2.morphology import GaussianMorphology
 
-warnings.filterwarnings("ignore")
+# resampling renderer to test
+cls = scarlet2.renderer.ResamplingRenderer
 
-# turn off automatic validation checks
-scarlet2.set_validation(False)
+# create a Gaussian as model
+T = 10
+eps = jnp.array((0.5, 0.3))
+model = GaussianMorphology(size=T, ellipticity=eps, shape=(150, 151))()[None, ...]  # test even & odd
+g = scarlet2.measure.Moments(component=model[0], N=2)
 
-# Load the HSC image data
-# load test data from HSC and HST
-filename = hf_hub_download(
-    repo_id="astro-data-lab/scarlet-test-data",
-    filename="test_resampling/Cut_HSC1.fits.gz",
-    repo_type="dataset",
-)
-obs_hdu = fits.open(filename)
-data_hsc = jnp.array(obs_hdu[0].data, jnp.float32)
-wcs_hsc = WCS(obs_hdu[0].header)
-channels_hsc = ["g", "r", "i", "z", "y"]
-obs_hdu.close()
-
-# Load the HSC PSF data
-filename = hf_hub_download(
-    repo_id="astro-data-lab/scarlet-test-data",
-    filename="test_resampling/PSF_HSC.fits.gz",
-    repo_type="dataset",
-)
-psf_hsc_data = jnp.array(fits.open(filename)[0].data, jnp.float32)
-Np1, Np2 = psf_hsc_data[0].shape
-psf_hsc_data = jnp.pad(psf_hsc_data, ((0, 0), (1, 0), (1, 0)))
-psf_hsc_single_band_ = psf_hsc_data[:1]
-psf_hsc = scarlet2.ArrayPSF(psf_hsc_data)
-psf_hsc_single_band = scarlet2.ArrayPSF(psf_hsc_single_band_)
-
-# Load the HST image data
-filename = hf_hub_download(
-    repo_id="astro-data-lab/scarlet-test-data",
-    filename="test_resampling/Cut_HST1.fits.gz",
-    repo_type="dataset",
-)
-obs_hdu = fits.open(filename)
-data_hst = jnp.array(obs_hdu[0].data, jnp.float32)
-wcs_hst = WCS(obs_hdu[0].header)
-channels_hst = ["F814W"]
-obs_hdu.close()
-
-# Load the HST PSF data
-filename = hf_hub_download(
-    repo_id="astro-data-lab/scarlet-test-data",
-    filename="test_resampling/PSF_HST.fits.gz",
-    repo_type="dataset",
-)
-psf_hst = jnp.array(fits.open(filename)[0].data, jnp.float32)
-psf_hst = psf_hst[None, :, :]
-psf_hst = jnp.pad(psf_hst, ((0, 0), (1, 0), (1, 0)))
-psf_hst_ = jnp.repeat(psf_hst, 5, 0)
-psf_hst_single_band_ = psf_hst_[:1]
-psf_hst = scarlet2.ArrayPSF(psf_hst_)
[…]
