### 2025-10-30 73d0009
minor fixes; removed source validation checks
