+++ src/scarlet2/source.py
@@ -10,10 +10,7 @@ from .module import Module
 from .morphology import Morphology
 from .spectrum import Spectrum
 from .validation_utils import (
-    ValidationError,
-    ValidationInfo,
     ValidationMethodCollector,
-    ValidationResult,
 )
 
 
@@ -254,24 +251,24 @@ class SourceValidator(metaclass=ValidationMethodCollector):
     def __init__(self, source: Source):
         self.source = source
 
-    def check_source_has_positive_contribution(self) -> ValidationResult:
-        """Check that the source has a positive contribution i.e. that the result
-        of evaluating self.source() does not contain negative values.
-
-        Returns
-        -------
-        ValidationResult
-            A subclass of ValidationResult indicating the result of the check.
-        """
-        model = self.source()
-        if jnp.any(model < 0):
-            return ValidationError(
-                "Source model has negative contributions.",
-                check=self.__class__.__name__,
-                context={"source": self.source},
-            )
-        else:
-            return ValidationInfo(
-                "Source model has positive contributions.",
-                check=self.__class__.__name__,
-            )
+    # def check_source_has_positive_contribution(self) -> ValidationResult:
+    #     """Check that the source has a positive contribution i.e. that the result
+    #     of evaluating self.source() does not contain negative values.
+    #
+    #     Returns
+    #     -------
+    #     ValidationResult
+    #         A subclass of ValidationResult indicating the result of the check.
+    #     """
+    #     model = self.source()
+    #     if jnp.any(model < 0):
+    #         return ValidationError(
+    #             "Source model has negative contributions.",
+    #             check=self.__class__.__name__,
+    #             context={"source": self.source},
+    #         )
+    #     else:
+    #         return ValidationInfo(
+    #             "Source model has positive contributions.",
+    #             check=self.__class__.__name__,
+    #         )
