+++ .github/workflows/asv-pr.yml
@@ -64,7 +64,7 @@ jobs:
       env:
         STEP_URL: ${{steps.jobs.outputs.html_url}}#step:10:1
     - name: Upload artifacts (PR number and benchmarks output)
-      uses: actions/upload-artifact@v4
+      uses: actions/upload-artifact@v5
       with:
         name: benchmark-artifacts
         path: ${{env.ARTIFACTS_DIR}}
