### 2025-11-04 cc96293
fixes #229
