### 2025-11-07 b6c2c73
ResamplingRenderer (#230)
* k-space resampling (as I understand it but not working that well)

* insert_into as standalone function

* added Lanczos renderer (test is not good yet); reversed coordinate convention for resample2d

* added Moments.convolve (+test); increased ProfileMorphology shape to 15*size

* fixed import for Lanczos resampler

* undid my changes to resampling renderer

* added Moment methods; redefined center shift for ProfileMorphology accordingly

* introduced translation to ResamplingRenderer; whole slew of minor corrections of Frame/WCS

* use jacobian only for resample_f
[…]
