+++ src/scarlet2/renderer.py
@@ -1,13 +1,15 @@
 """Renderer classes"""
 
+from functools import partial
+
 import equinox as eqx
 import jax
 import jax.numpy as jnp
 
 from .bbox import Box, overlap_slices
 from .fft import _get_fast_shape, _trim, _wrap_hermitian_x, convolve, deconvolve, good_fft_size, transform
-from .frame import get_affine, get_scale_angle_flip
-from .interpolation import resample_ops
+from .frame import _minmax_int, get_affine, get_scale_angle_flip_shift
+from .interpolation import Interpolant, Lanczos, resample3d, resample_fourier
 
 
 class Renderer(eqx.Module):
@@ -187,7 +189,7 @@ class TrimSpatialBox(Renderer):
         )
 
     def __call__(self, model, key=None):
-        """What to run when AdjustToFrame is called"""
+        """What to run when TrimSpatialBox is called"""
         sub = model[:, self.slices[-2].get_slice(), self.slices[-1].get_slice()]
         return sub
 
@@ -196,13 +198,17 @@ class ResamplingRenderer(Renderer):
     """Renderer to resample image to different pixel grid (subpixel position, resolution, orientation)"""
 
     padding: int
-    fft_shape_target: int = eqx.field(repr=False)
-    fft_shape_model_im: int = eqx.field(repr=False)
     scale: float
     angle: float
-    flip: int
-    model_kpsf_interp: jnp.array = eqx.field(repr=False)
-    obs_kpsf_interp: jnp.array = eqx.field(repr=False)
+    handedness: int
+    shift: tuple
+    has_psf_in: bool
+    has_psf_out: bool
+    fft_shape_target: int = eqx.field(repr=False)
+    fft_shape_model_im: int = eqx.field(repr=False)
+    jacobian: jnp.array = eqx.field(repr=False)
+    model_kpsf_interp: jnp.array = eqx.field(repr=False, default=None)
+    obs_kpsf_interp: jnp.array = eqx.field(repr=False, default=None)
     real_shape_target: tuple = eqx.field(repr=False)
 
     def __init__(self, model_frame, obs_frame, padding=4):
@@ -228,80 +234,91 @@ class ResamplingRenderer(Renderer):
         # resampling to obs frame, followed by a convolution in obs frame because the difference
         # kernel would be expressed in obs pixel and can thus easily undersample the model PSF.
 
-        # create PSF model
-        psf_model = model_frame.psf()
-        if len(psf_model.shape) == 2:
-            psf_model = psf_model[None, ...]
-
-        if len(psf_model.shape) == 2:  # only one image for all bands
-            psf_model = jnp.tile(psf_model, (obs_frame.bbox.shape[0], 1, 1))
-
-        psf_obs = obs_frame.psf()
-        if len(psf_obs.shape) == 2:
-            psf_obs = psf_obs[None, ...]
-
-        self.fft_shape_model_im = good_fft_size(padding * max(model_frame.bbox.shape))
-        fft_shape_model_psf = good_fft_size(padding * max(psf_model.shape))
-        fft_shape_obs_psf = good_fft_size(padding * max(psf_obs.shape))
-
-        # Fourier transform model and observation PSFs
-        model_kpsf = jnp.fft.fftshift(
-            transform(psf_model, (fft_shape_model_psf, fft_shape_model_psf), (-2, -1)), (-2)
-        )
-        obs_kpsf = jnp.fft.fftshift(
-            transform(psf_obs, (fft_shape_obs_psf, fft_shape_obs_psf), (-2, -1)), (-2)
-        )
+        # Extract rotation angle, flip, scale between WCSs
+        m_in = get_affine(model_frame.wcs)
+        m_out = get_affine(obs_frame.wcs)
[…]
