+++ src/scarlet2/measure.py
@@ -1,9 +1,11 @@
 """Measurement methods"""
 
+import copy
+
 import numpy as jnp
 import numpy.ma as ma
 
-from .frame import get_affine, get_scale_angle_flip
+from .frame import get_affine, get_scale_angle_flip_shift
 from .source import Component
 
 
@@ -194,14 +196,17 @@ class Moments(dict):
                 self[m, n - m] = (grid_y**m * grid_x ** (n - m) * model * weight).sum(axis=(-2, -1))
 
             if n == 1:
+                self._centroid = jnp.array((self[1, 0] / self[0, 0], self[0, 1] / self[0, 0]))
                 # shift grid to produce centered moments
                 if center is None:
-                    center = self[1, 0] / self[0, 0], self[0, 1] / self[0, 0]  # centroid
+                    center = self._centroid
                     self[1, 0] = jnp.zeros_like(self[1, 0])
                     self[0, 1] = jnp.zeros_like(self[0, 1])
                 else:
                     center = jnp.asarray(center)
                     # centroid wrt given center
+                    self._centroid[0] -= center[0] * self[0, 0]
+                    self._centroid[1] -= center[1] * self[0, 0]
                     self[1, 0] -= center[0] * self[0, 0]
                     self[0, 1] -= center[1] * self[0, 0]
                 if model.ndim == 3 and center[0].ndim == 1:
@@ -221,6 +226,16 @@ class Moments(dict):
         return max(key[0] for key in self.keys())
 
     @property
+    def flux(self):
+        """Determine flux from 0th moment
+
+        Returns
+        -------
+        float
+        """
+        return self[0, 0]
+
+    @property
     def centroid(self):
         """Determine centroid from moments
 
@@ -229,7 +244,7 @@ class Moments(dict):
         array
             Coordinates of the centroid in the pixel frame of the data that defines these moments
         """
-        return jnp.array((self[1, 0] / self[0, 0], self[0, 1] / self[0, 0]))
+        return self._centroid
 
     @property
     def size(self):
@@ -266,6 +281,36 @@ class Moments(dict):
             self[key] /= norm
         return self
 
+    def convolve(self, p):
+        """Convolve moments with moments `p`
+
+        The moments are changed in place.
+
+        See Melchior et al. (2010), "Weak gravitational lensing with Deimos", Equation 9
+
+        Parameters
+        ----------
+        p: Moments
+            Moments of the kernel to convolve with
+
+        Returns
+        -------
+        self
+        """
+
+        g_ = self
+        g = copy.deepcopy(g_)
+        n_min = min(p.order, g.order)
+
+        for n in range(n_min + 1):
+            for i in range(n + 1):
+                j = n - i
+                g_[i, j] = jnp.zeros_like(g_[i, j])
+                for k in range(i + 1):
+                    for l in range(j + 1):  # noqa: E741
+                        g_[i, j] += binomial(i, k) * binomial(j, l) * g[k, l] * p[i - k, j - l]
+        return self
+
     def deconvolve(self, p):
         """Deconvolve moments from moments `p`
 
@@ -292,23 +337,26 @@ class Moments(dict):
             g[1, 0] -= g[0, 0] * p[1, 0]
             g[0, 1] /= p[0, 0]
             g[1, 0] /= p[0, 0]
-            if n_min >= 2:
-                g[0, 2] -= g[0, 0] * p[0, 2] + 2 * g[0, 1] * p[0, 1]
-                g[1, 1] -= g[0, 0] * p[1, 1] + g[0, 1] * p[1, 0] + g[1, 0] * p[0, 1]
-                g[2, 0] -= g[0, 0] * p[2, 0] + 2 * g[1, 0] * p[1, 0]
-                if n_min >= 3:
[…]
