### 2025-11-17 578be14
added channel names to Observation label (#236)
