### 2025-11-25 e722d27
CorrelatedObservation with construction from correlation function
