### 2025-11-26 815f9b9
refactored chi^2 and likelihood
