### 2025-11-26 aec1737
classmethod for CorrelatedObservation; updated tutorial
