### 2025-11-26 714d83c
removed 1D power spectrum
