### 2025-11-26 bb3cb5d
Correlated noise (#195)
* updates

* fourier likelihood

* add demo notebook

* add minimal docstring and fix uppercase variable

* adding CorrelatedNoiseObservation class

* added correlation_function

* CorrelatedObservation with construction from correlation function

* refactored chi^2 and likelihood

* classmethod for CorrelatedObservation; updated tutorial

* removed 1D power spectrum

---------

Co-authored-by: Peter Melchior <peter.m.melchior@gmail.com>
