+++ src/scarlet2/observation.py
@@ -4,6 +4,7 @@ import jax.numpy as jnp
 
 from .bbox import Box, overlap_slices
 from .frame import Frame, get_affine
+from .measure import correlation_function
 from .module import Module
 from .renderer import (
     ChannelRenderer,
@@ -34,6 +35,8 @@ class Observation(Module):
     """Metadata to describe what view of the sky `data` amounts to"""
     renderer: (Renderer, eqx.nn.Sequential)
     """Renderer to translate from the model frame the observation frame"""
+    n: int
+    """Number of valid pixels in `data`"""
 
     def __init__(self, data, weights, psf=None, wcs=None, channels=None, renderer=None):
         # TODO: automatic conversion to jnp arrays
@@ -46,6 +49,9 @@ class Observation(Module):
         if self.weights is not None and self.weights.ndim == 2:
             # add a channel dimension if it is missing
             self.weights = self.weights[None, ...]
+
+        # number of unmasked pixels
+        self.n = jnp.prod(jnp.asarray(data.shape)) - jnp.sum(self.weights == 0)
 
         self.frame = Frame(Box(self.data.shape), psf, wcs, channels=channels)
         if renderer is None:
@@ -92,16 +98,12 @@ class Observation(Module):
         return self._log_likelihood(model, self.data)
 
     def _log_likelihood(self, model, data):
-        # rendered model
-        model_ = self.render(model)
-
         # normalization of the single-pixel likelihood:
         # 1 / [(2pi)^1/2 (sigma^2)^1/2]
         # with inverse variance weights: sigma^2 = 1/weight
         # full likelihood is sum over all (unmasked) pixels in data
-        n = jnp.prod(jnp.asarray(data.shape)) - jnp.sum(self.weights == 0)
-        log_norm = n / 2 * jnp.log(2 * jnp.pi)
-        log_like = -jnp.sum(self.weights * (model_ - data) ** 2) / 2
+        log_like = -self._chisquare(model) / 2
+        log_norm = self.n / 2 * jnp.log(2 * jnp.pi)
         return log_like - log_norm
 
     def goodness_of_fit(self, model):
@@ -123,8 +125,10 @@ class Observation(Module):
         float
         """
         # only use unmasked pixels in the data
-        n = jnp.prod(jnp.asarray(self.data.shape)) - jnp.sum(self.weights == 0)
-        return (self.weights * (self.render(model) - self.data) ** 2).sum() / n
+        return self._chisquare(model) / self.n
+
+    def _chisquare(self, model):
+        return jnp.sum(self.weights * (self.render(model) - self.data) ** 2)
 
     def match(self, frame, renderer=None):
         """Construct the mapping between `frame` (from the model) and this observation frame
@@ -209,6 +213,133 @@ class Observation(Module):
             chi_dict[i] = {"in": chi_in, "out": chi_out}
 
         return chi_dict
+
+
+class CorrelatedObservation(Observation):
+    """Content and definition of an observation with pixel correlations
+
+    The noise model is still assumed to be Gaussian, but with correlations between pixels.
+    The implementation computes the goodness of fit in Fourier space from the noise power spectrum to avoid
+    the expensive computation of/with an inverse banded matrix in configuration space.
+    """
+
+    power_spectrum: jnp.ndarray
+    """Noise power spectrum for :py:meth:`log_likelihood`"""
+
+    def __init__(
+        self,
+        data,
+        weights,
+        psf=None,
+        wcs=None,
+        channels=None,
+        renderer=None,
+        power_spectrum=None,
+        correlation_function=None,
+    ):
+        super().__init__(data, weights, psf=psf, wcs=wcs, channels=channels, renderer=renderer)
+
+        assert power_spectrum is not None or correlation_function is not None, (
+            "Provide either power_spectrum or correlation_function"
+        )
+
[…]
