+++ .github/workflows/asv-nightly.yml
@@ -26,7 +26,7 @@ jobs:
       with:
         python-version: ${{env.PYTHON_VERSION}}
     - name: Checkout main branch of the repository
-      uses: actions/checkout@v5
+      uses: actions/checkout@v6
       with:
         fetch-depth: 0
     - name: Install dependencies
