### 2025-12-02 eb75a36
decreased min_snr to 3; more stable testing for Gaussian shape (#238)
