+++ src/scarlet2/init.py
@@ -26,7 +26,7 @@ def _get_edge_pixels(img, box):
     return jnp.concatenate(edge, axis=1)
 
 
-def make_bbox(obs, center_pix, sizes=[11, 17, 25, 35, 47, 61, 77], min_snr=20, min_corr=0.99):  # noqa: B006
+def best_box_moments(obs, center_pix, sizes=[11, 17, 25, 35, 47, 61, 77], min_snr=3, min_corr=0.99):  # noqa: B006
     """Make a bounding box for source at center
 
     This method finds small box around the center so that the edge flux has a minimum SNR,
@@ -49,7 +49,7 @@ def make_bbox(obs, center_pix, sizes=[11, 17, 25, 35, 47, 61, 77], min_snr=20, m
 
     Returns
     -------
-    ~scarlet2.Box
+    :py:class:`~scarlet2.Box`, :py:class:`~scarlet2.Moments`
     """
     assert isinstance(obs, Observation)
     assert obs.weights is not None, "Observation weights are required"
@@ -61,6 +61,7 @@ def make_bbox(obs, center_pix, sizes=[11, 17, 25, 35, 47, 61, 77], min_snr=20, m
     # increase box size from list until SNR is below threshold or spectrum changes significantly
     peak_spectrum = pixel_spectrum(obs, center_pix, correct_psf=True)
     last_spectrum = jnp.empty(len(peak_spectrum))
+    moments = []
     for i in range(len(sizes)):
         box2d = Box((sizes[i], sizes[i]))
         box2d.set_center(center_pix.astype(int))
@@ -69,66 +70,41 @@ def make_bbox(obs, center_pix, sizes=[11, 17, 25, 35, 47, 61, 77], min_snr=20, m
         valid_edge_pixel = edge_pixels != 0
         edge_spectrum = jnp.sum(edge_pixels, axis=-1) / jnp.sum(valid_edge_pixel, axis=-1)
         weight_edge_pixels = _get_edge_pixels(obs.weights, box2d)
-        snr_edge_pixels = edge_pixels * jnp.sqrt(weight_edge_pixels)
+        snr_edge_pixels = jnp.abs(edge_pixels) * jnp.sqrt(weight_edge_pixels)
         mean_snr = jnp.sum(jnp.sum(snr_edge_pixels, axis=-1) / jnp.sum(valid_edge_pixel, axis=-1))
         spec_corr = (
             jnp.dot(edge_spectrum, peak_spectrum)
             / jnp.sqrt(jnp.dot(peak_spectrum, peak_spectrum))
             / jnp.sqrt(jnp.dot(edge_spectrum, edge_spectrum))
         )
+        m = boxed_moments(obs, center_pix, bbox=box2d)
 
         if mean_snr < min_snr or max(box2d.shape) > max(obs.frame.bbox.spatial.shape):
             break
 
-        if i > 0 and (spec_corr < min_corr or jnp.any(edge_spectrum > last_spectrum)):
+        if i > 0 and (
+            spec_corr < min_corr or jnp.any(edge_spectrum > last_spectrum) or jnp.any(jnp.isnan(m.size))
+        ):
             box2d = Box((sizes[i - 1], sizes[i - 1]))
             box2d.set_center(center_pix.astype(int))
+            m = moments[-1]
             break
 
         last_spectrum = edge_spectrum
+        moments.append(m)
 
     box = obs.frame.bbox[0] @ box2d
-    return box
+    m = standardize_moments(m, obs)
+    return box, m
 
 
-def compact_morphology(min_value=1e-6, max_value=1 - 1e-6):
-    """Create image of the point source morphology model, i.e. the most compact source possible
-
-    Parameters
-    ----------
-    min_value: float
[…]
