+++ .github/workflows/publish-to-pypi.yml
@@ -23,7 +23,7 @@ jobs:
     permissions:
       id-token: write
     steps:
-    - uses: actions/checkout@v5
+    - uses: actions/checkout@v6
     - name: Set up Python
       uses: actions/setup-python@v6
       with:
