### 2025-12-02 ff3ac6d
added mask to CorrelatedObservation; works now for multi-band data
