### 2025-12-02 3af4c7b
much faster correlation function
