### 2025-12-02 a77c11e
first stab at resampling observations
