+++ src/scarlet2/observation.py
@@ -2,13 +2,14 @@ import equinox as eqx
 import jax
 import jax.numpy as jnp
 
-from .bbox import Box, overlap_slices
+from .bbox import Box, insert_into, overlap_slices
 from .frame import Frame, get_affine
 from .measure import correlation_function
 from .module import Module
 from .renderer import (
     ChannelRenderer,
     ConvolutionRenderer,
+    LanczosResamplingRenderer,
     NoRenderer,
     Renderer,
     ResamplingRenderer,
@@ -293,7 +294,7 @@ class CorrelatedObservation(Observation):
         return jnp.sum((res_fft * jnp.conjugate(res_fft)).real / self.power_spectrum)
 
     @classmethod
-    def from_observation(cls, obs, length=50):
+    def from_observation(cls, obs, patch_size=50, maxlength=2, resample_to_frame=None):
         """Create a :py:class:`CorrelatedObservation` from :py:class:`Observation`
 
         The method will construct a new Observation instance with a modified likelihood that takes into
@@ -304,57 +305,105 @@ class CorrelatedObservation(Observation):
         ----------
         obs: :py:class:`Observation`
             Observation containing the data and original weight map
-        length: int
-            Linear size of the patch.
-            Computing the correlation function is expensive. as the correlations are usually short-range,
-            a small patch is usually sufficient for accurate results and easier to find from empty regions
-            in the image.
+        patch_size: int
+            Linear size of the patch for measuring the correlation function
+        maxlength: int
+            Maximum distance (in pixels) for the 2D correlation function
+        resample_to_frame: None, :py:class:`~scarlet2.Frame`
+            Frame describing the desired spatial sampling
 
         Returns
         -------
         :py:class:`CorrelatedObservation`
         """
-        # compute the pixel correlations in a noisy patch (without correlations from sources)
-        # 1) mask pixels with bright pixels or zero weights
-        img = obs.data
-        mask = obs.weights == 0
-        mask = mask.at[~mask].set(img[~mask] > 3 * jnp.sqrt(1 / obs.weights[~mask]))
-        # extend the mask to remove most of the outskirts of detected galaxies
-        kernel = jnp.ones((9, 9))
-        _correlate2d = lambda x, kernel: jax.scipy.signal.correlate2d(x, kernel, mode="same")
-        correlate3d = jax.vmap(_correlate2d, in_axes=(0, None), out_axes=0)
-        mask = correlate3d(mask, kernel) > 0
-        img_ = img.at[mask].set(0)
+        if resample_to_frame is not None:
+            # create a reverse renderer without PSF corrections or channel filtering
+            _obs_frame = Frame(obs.frame.bbox, psf=None, wcs=obs.frame.wcs, channels=obs.frame.channels)
+            _new_box = obs.frame.bbox[:-2] @ resample_to_frame.bbox.spatial
+            _model_frame = Frame(_new_box, psf=None, wcs=resample_to_frame.wcs, channels=obs.frame.channels)
+            _renderer = LanczosResamplingRenderer(_obs_frame, _model_frame, lanczos_order=9)
+            wcs = resample_to_frame.wcs
 
-        # 2) find patch of size length (at most image size) with the largest number of unmasked pixels
-        length = min(length, min(img.shape[-2:]))
-        shape = (length, length)
-        kernel = jnp.ones(shape)
-        # correlated with tophat = sliding sum
-        gaps = correlate3d(mask == 0, kernel)
+            # resample data
+            data = _renderer(obs.data)
 
-        # location of lower-left pixel of patch with fewest masked pixels
-        def best_patch(img, gaps):
-            # trim off L//2 to avoid the center of patch is to close to image border
[…]
