+++ src/scarlet2/renderer.py
@@ -352,11 +352,11 @@ class LanczosResamplingRenderer(Renderer):
     interpolant: Interpolant
     coords: jnp.ndarray = eqx.field(repr=False)
     warp: jnp.ndarray = eqx.field(repr=False)
-    _diff_kernel_fft: jnp.array = eqx.field(repr=False)
-    _fft_shape: int = eqx.field(repr=False)
+    _diff_kernel_fft: jnp.array = eqx.field(repr=False, default=None)
+    _fft_shape: int = eqx.field(repr=False, default=None)
 
-    def __init__(self, model_frame, obs_frame):
-        self.interpolant = Lanczos(7)
+    def __init__(self, model_frame, obs_frame, lanczos_order=5):
+        self.interpolant = Lanczos(lanczos_order)
         model_shape = model_frame.bbox.spatial.shape
         self.coords = jnp.stack(
             jnp.meshgrid(jnp.arange(model_shape[0]), jnp.arange(model_shape[1])), -1
@@ -364,46 +364,47 @@ class LanczosResamplingRenderer(Renderer):
         obs_shape = obs_frame.bbox.spatial.shape
         self.warp = obs_frame.convert_pixel_to(model_frame).reshape(obs_shape[0], obs_shape[1], 2)
 
-        # construct diff kernel in model_space
-        # create PSF model
-        psf = model_frame.psf()
-        psf_model = jnp.tile(psf, (obs_frame.bbox.shape[0], 1, 1)) if len(psf.shape) == 2 else psf
+        if model_frame.psf is not None and model_frame.wcs is not None:
+            # construct diff kernel in model_space
+            # create PSF model
+            psf = model_frame.psf()
+            psf_model = jnp.tile(psf, (obs_frame.bbox.shape[0], 1, 1)) if len(psf.shape) == 2 else psf
 
-        # resample obs psf in model pixel
-        psf_obs = obs_frame.psf()
-        # TODO: what is different between indices and meshgrid???
-        # coords_ = jnp.stack(jnp.indices(obs_psf.shape[-2:]), axis=-1).astype(jnp.float32)
-        coords_ = jnp.stack(
-            jnp.meshgrid(jnp.arange(psf_obs.shape[-2]), jnp.arange(psf_obs.shape[-1])), -1
-        ).astype(jnp.float32)
-        coords_in_model_space = obs_frame.convert_pixel_to(model_frame, pixel=coords_)
-        ylims = _minmax_int(coords_in_model_space[..., 0])
-        xlims = _minmax_int(coords_in_model_space[..., 1])
-        warp_ = jnp.stack(
-            jnp.meshgrid(
-                jnp.arange(ylims[0] - 1, ylims[1] - 1),
-                jnp.arange(xlims[0], xlims[1]),
-            ),
-            -1,
-        ).astype(jnp.float32)
-        warp__ = model_frame.convert_pixel_to(obs_frame, pixel=warp_).reshape(
-            warp_.shape[0], warp_.shape[1], 2
-        )
-        # interpolate observed to model pixels
-        psf_obs_interp = resample3d(psf_obs, coords=coords_, warp=warp__, interpolant=self.interpolant)
+            # resample obs psf in model pixel
+            psf_obs = obs_frame.psf()
+            # TODO: what is different between indices and meshgrid???
+            # coords_ = jnp.stack(jnp.indices(obs_psf.shape[-2:]), axis=-1).astype(jnp.float32)
+            coords_ = jnp.stack(
+                jnp.meshgrid(jnp.arange(psf_obs.shape[-2]), jnp.arange(psf_obs.shape[-1])), -1
+            ).astype(jnp.float32)
[…]
