### 2025-12-05 43f09ab
AsinhAutomaticNorm works with CorrelatedObservation
