### 2025-12-09 3541d4f
removed automatic obs.match(model_frame) in Frame.from_observations; refactored renderer
