### 2025-12-10 fa1d68e
set weight according to power spectrum; fix padding
