### 2025-12-10 1dc1f69
renderer set on demand
