### 2025-12-10 4dbfbbc
simplified test
