### 2025-12-10 1fa02a5
corrected model colors in plot.sources
