### 2025-12-13 c943efa
added Observation.name as consistent plot label
