### 2025-12-15 f4b070e
Correlated observation (#239)
* added mask to CorrelatedObservation; works now for multi-band data

* much faster correlation function

* first stab at resampling observations

* AsinhAutomaticNorm works with CorrelatedObservation

* removed automatic obs.match(model_frame) in Frame.from_observations; refactored renderer

* set weight according to power spectrum; fix padding

* renderer set on demand

* simplified test

* docs update: resampling for multi-res fits

* getting rid of ruff complaints (hopefully)

* corrected model colors in plot.sources

* added Observation.name as consistent plo
[…]
