+++ src/scarlet2/observation.py
@@ -2,14 +2,14 @@ import equinox as eqx
 import jax
 import jax.numpy as jnp
 
-from .bbox import Box, overlap_slices
+from .bbox import Box, insert_into, overlap_slices
 from .frame import Frame, get_affine
 from .measure import correlation_function
 from .module import Module
 from .renderer import (
     ChannelRenderer,
     ConvolutionRenderer,
-    NoRenderer,
+    LanczosResamplingRenderer,
     Renderer,
     ResamplingRenderer,
     TrimSpatialBox,
@@ -33,12 +33,14 @@ class Observation(Module):
     """Statistical weights (usually inverse variance) for :py:meth:`log_likelihood`"""
     frame: Frame
     """Metadata to describe what view of the sky `data` amounts to"""
-    renderer: (Renderer, eqx.nn.Sequential)
+    renderer: (None, Renderer, eqx.nn.Sequential)
     """Renderer to translate from the model frame the observation frame"""
     n: int
     """Number of valid pixels in `data`"""
+    name: str
+    """Name to describe the observation"""
 
-    def __init__(self, data, weights, psf=None, wcs=None, channels=None, renderer=None):
+    def __init__(self, data, weights, psf=None, wcs=None, channels=None, renderer=None, name=None):
         # TODO: automatic conversion to jnp arrays
         self.data = data
         if self.data.ndim == 2:
@@ -54,9 +56,8 @@ class Observation(Module):
         self.n = jnp.prod(jnp.asarray(data.shape)) - jnp.sum(self.weights == 0)
 
         self.frame = Frame(Box(self.data.shape), psf, wcs, channels=channels)
-        if renderer is None:
-            renderer = NoRenderer()
         self.renderer = renderer
+        self.name = name if name is not None else ""
 
         # (re)-import `VALIDATION_SWITCH` at runtime to avoid using a static/old value
         from .validation_utils import VALIDATION_SWITCH
@@ -80,6 +81,9 @@ class Observation(Module):
         array
             Prediction of the observation given the `model`. Has the same shape as :py:attr:`data`.
         """
+        assert self.renderer is not None, (
+            "Observation requires a renderer. Call Observation.match(model_frame) first"
+        )
         return self.renderer(model)
 
     def log_likelihood(self, model):
@@ -130,6 +134,21 @@ class Observation(Module):
     def _chisquare(self, model):
         return jnp.sum(self.weights * (self.render(model) - self.data) ** 2)
 
+    def check_set_renderer(self, frame):
+        """Check existence of :py:attr:`renderer`, or set it by calling :py:meth:`match`
+
+        Parameters
+        ----------
+        frame: Frame
+            The frame to match
+
+        Returns
+        -------
+        None
+        """
+        if self.renderer is None:
+            self.match(frame)
+
     def match(self, frame, renderer=None):
         """Construct the mapping between `frame` (from the model) and this observation frame
 
@@ -143,7 +162,7 @@ class Observation(Module):
 
         Returns
         -------
-        self
+        None
         """
         # choose the renderer
         if renderer is None:
@@ -173,7 +192,7 @@ class Observation(Module):
                 renderers.append(ResamplingRenderer(frame, self.frame))
 
             if len(renderers) == 0:
-                renderer = NoRenderer()
+                renderer = lambda x, key=None: x
             elif len(renderers) == 1:
                 renderer = renderers[0]
             else:
@@ -185,7 +204,6 @@ class Observation(Module):
                 "Renderer does not map model frame to observation frame"
             )
         object.__setattr__(self, "renderer", renderer)
-        return self
 
     def eval_chi_square_in_box_and_border(self, scene, border_width=3):
         """
[…]
