+++ .github/workflows/publish-benchmarks-pr.yml
@@ -27,7 +27,7 @@ jobs:
         echo "Conclusion: ${{ github.event.workflow_run.conclusion }}"
         echo "Event: ${{ github.event.workflow_run.event }}"
     - name: Download artifact
-      uses: dawidd6/action-download-artifact@v11
+      uses: dawidd6/action-download-artifact@v12
       with:
         name: benchmark-artifacts
         run_id: ${{ github.event.workflow_run.id }}
