### 2026-01-30 #9646 @Jammy2211
Scarlet2 is a powerful tool for galaxy morphology and scene modelling. I am familiar with older software packages (e.g. GALAPAGOS, based on GALFIT), and even a brief run-through of scarlet2 makes it clear that rapid progress has been made in this area. In particular, the software benefits from fast GPU-enabled runtime via JAX, automatic differentiation for sampling, and a concise interface for fitting complex galaxy scenes in large images.

The codebase appears to be written to a high standard, and the documentation (once fully functional) should allow new users to get started quickly. I therefore have no major concerns and am happy to recommend the software for publication. I do, however, leave a few comments below that the authors may wish to consider.

In addition, the paper is currently missing a “State of the Field” section, which should be added.

**Comments**

The quick start does not run, with the following error cropping up:

```
    sc2.plot.observation(obs, norm=norm, sky_coords=centers, show_psf=True, add_labels=True)
  File "/home/jammy/venv/Scarlet2/lib/python3.10/site-packages/scarlet2/plot.py", line 443, in observation
    psf_model = observation.frame.psf()
TypeError: 'jaxlib._jax.ArrayImpl' object is not callable
```

**Abstractions:** scarlet2 makes extensive use of object-oriented abstractions to perform essentially all tasks involved in galaxy morphology fitting. This enables a very concise and powerful interface for complex modelling in only a few lines of Python.

However, in some cases the abstraction felt too opaque, to the point where it was difficult to understand what model was actually being fit, or how to inspect or extend it.

For example, while running the quick start, I was unsure whether the model being fit involved Sérsic profiles or some other morphological representation. After reading through the full example (albeit relatively fast...), I still did not have a clear sense of the underlying parameterisation.

I then consulted the “How To…” section, which provides guidance on defining parameters and priors, but I still came away uncertain about what the default model in the quick start is. In the “Sample From Posterior” section, the corner plot shows parameters labelled “spectrum 0”, “spectrum 1”, etc., but I was unsure what physical components these correspond to.

After rereading the material in more detail, my impression is that the quick start is using point sources and the "spectrum" parameters are their brightness's as the primary free parameter.

I think this comment amounts to me saying that the quick start could be improved with a few refinements to make it more explicit what is being fit and what the free parameters represent.

I think it would also help if the quick start avoided try / except statements and if statements, as this also made it difficult for me to pair what the text was saying with the code, which meant I lost track of abstractions, for example

```
with sc2.Scene(model_frame) as scene:
    for i, center in enumerate(centers):
        if i == 0:  # we know source 0 is a star
            spectrum = sc2.init.pixel_spectrum(obs, center, correct_psf=True)
            sc2.PointSource(center, spectrum)
        else:
            try:
                spectrum, morph = sc2.init.from_gaussian_moments(obs, center)
            except ValueError:
                spectrum = sc2.init.pixel_spectrum(obs, center)
                morph = sc2.init.compact_morphology()

            sc2.Source(center, spectrum, morph)
```

This is API overload for someone new to the software!
