### 2026-02-01 e21e9cd
Bump dawidd6/action-download-artifact from 12 to 13
Bumps [dawidd6/action-download-artifact](https://github.com/dawidd6/action-download-artifact) from 12 to 13.
- [Release notes](https://github.com/dawidd6/action-download-artifact/releases)
- [Commits](https://github.com/dawidd6/action-download-artifact/compare/v12...v13)

---
updated-dependencies:
- dependency-name: dawidd6/action-download-artifact
  dependency-version: '13'
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
