### 2026-02-03 #9646 @pmelchior
@ivastar yes, that is doable.
