### 2026-02-06 6f6cc92
removed obsolete apply_constraint and OK parameter checks
