### 2026-02-09 02e99d0
monkey patching parameters onto Module; inference does not need parameter arg
