+++ src/scarlet2/module.py
@@ -43,29 +43,26 @@ class Module(eqx.Module):
         object.__setattr__(self, "parameters", parameters)
         return self
 
-    def get(self, parameters):
+    def get_parameters(self, ptree):
         """Get parameter arrays from this module
-
-        Parameters
-        ----------
-        parameters: :py:class:`Parameters`
-            List of parameters to search
 
         Returns
         -------
-        list
+        tuple
             requested data arrays for `parameters`
         """
-        assert isinstance(parameters, Parameters)
-        return parameters.extract_from(self)
+        if not hasattr(self, "parameters"):
+            return tuple()
 
-    def replace(self, parameters, values):
+        get_node = lambda node, p: node if isinstance(p, Parameter) else None
+        leaves = jtu.tree_leaves(jtu.tree_map(get_node, self, ptree))
+        return tuple(l for l in leaves if l is not None)
+
+    def replace_parameters(self, values):
         """Replace parameter arrays by another value
 
         Parameters
         ----------
-        parameters: :py:class:`Parameters`
-            List of parameters to search
         values: list
             List of values to replace the current parameter arrays with
             Needs to be in the same shape and order as the current parameter values
@@ -75,10 +72,10 @@ class Module(eqx.Module):
         :py:class:`Module`
             Modified module. All other module components are unchanged.
         """
-        where = lambda model: model.get(parameters)  # noqa: E731
+        where = lambda model: model.get_parameters()  # noqa: E731
         return eqx.tree_at(where, self, replace=values)
 
-    def get_filter_spec(self, parameters):
+    def get_filter_spec(self):
         """Get equinox filter_spec for all fields named in parameters
 
         Parameters
@@ -91,12 +88,13 @@ class Module(eqx.Module):
         list
             requested data arrays for `parameters`
         """
-        filtered = jax.tree_util.tree_map(lambda _: False, self)
-        where = lambda model: model.get(parameters)  # noqa: E731
-        values = (True,) * len(parameters)
-        filtered = eqx.tree_at(where, filtered, replace=values)
-        if all(jax.tree_util.tree_leaves(filtered)):
+        if not hasattr(self, "parameters"):
             return None
+
+        filtered = jax.tree_util.tree_map(lambda _: False, self)
+        where = lambda model: model.get_parameters()  # noqa: E731
+        values = (True,) * len(self.parameters)
+        filtered = eqx.tree_at(where, filtered, replace=values)
         return filtered
 
 
@@ -135,8 +133,6 @@ class Parameter:
             self.name = varname.argname("node", vars_only=False)
         else:
             self.name = name
-        self.node = node
-
         self.constraint = constraint
         if self.constraint is not None:
             try:
@@ -148,14 +144,68 @@ class Parameter:
 
         self.prior = prior
         self.stepsize = stepsize
+        self.node = node
 
         # add this source to the active scene
         try:
+            base = Parameterization.parameters.base
+            # go to pixel frame even if specified in sky coords
+            if hasattr(base, "frame"):
+                self.to_pixels(base.frame)
             Parameterization.parameters.__iadd__(self)
         except AttributeError as err:
             msg = "A Parameter instance should only be created within the context of Parameters\n"
             msg += "Use 'with Parameters(scene) as p: Parameter(...)'"
             raise RuntimeError(msg) from err
+
+    def to_pixels(self, frame):
+        """Convert parameter to pixel coordinates of the model frame
+
+        scarlet2 models are optimized in pixel coordinates (defined by the model
+        frame of :py:class:`~scarlet2.Scene`. Therefore parameters (or their priors,
+        stepsize, etc) that are defined in :py:mod:`astropy.units` or
[…]
