+++ src/scarlet2/scene.py
@@ -1,14 +1,14 @@
-import logging
 from collections import defaultdict
 
 import equinox as eqx
 import jax
 import jax.numpy as jnp
+import jax.tree_util as jtu
 
 from . import Scenery
 from .bbox import overlap_slices
 from .frame import Frame
-from .module import Module, Parameters
+from .module import Module, Parameter
 from .validation_utils import (
     ValidationError,
     ValidationInfo,
@@ -146,6 +146,11 @@ class Scene(Module):
         """
         # uses numpyro NUTS on all non-fixed parameters
         # requires that those have priors set
+
+        # no optimizable parameters: return self
+        if not hasattr(self, "parameters"):
+            return self
+
         try:
             import numpyro
             import numpyro.distributions as dist
@@ -183,8 +188,8 @@ class Scene(Module):
                 return self.obs._log_likelihood(self.model, value)
 
         # find all non-fixed parameters and their priors
-        parameters = self.parameters
-        priors = {p.name: p.prior for p in parameters}
+        ptree = self.parameters.as_tree()
+        priors = {p.name: p.prior for p in self.parameters}
         has_none = any(prior is None for prior in priors.values())
         if has_none:
             from pprint import pformat
@@ -194,9 +199,12 @@ class Scene(Module):
 
         # define the pyro model, where every parameter becomes a sample,
         # and the observations sample from their likelihood given the rendered model
+        sample_node = lambda node, p: numpyro.sample(p.name, p.prior) if isinstance(p, Parameter) else node
+
         def pyro_model(model):
-            samples = tuple(numpyro.sample(p.name, p.prior) for p in parameters)
-            model_ = model.replace(parameters, samples)
+            # samples = tuple(numpyro.sample(p.name, p.prior) for p in self.parameters)
+            # model_ = model.replace_parameters(samples)
+            model_ = jtu.tree_map(sample_node, model, ptree)
             pred = model_()  # create prediction once for all observations
             # dealing with multiple observations
             for i, obs_ in enumerate(observations):
@@ -212,7 +220,7 @@ class Scene(Module):
 
             from numpyro.infer.initialization import init_to_value
 
-            values = {p.name: p.node for p in parameters}
+            values = {p.name: p.node for p in self.parameters}
             init_strategy = partial(init_to_value, values=values)
 
         nuts_kernel = NUTS(pyro_model, init_strategy=init_strategy, **kwargs)
@@ -266,6 +274,10 @@ class Scene(Module):
         Scene
             The scene model with updated parameters
         """
+        # no optimizable parameters: return self
+        if not hasattr(self, "parameters"):
+            return self
+
         try:
             import optax
             import optax._src.base as base
@@ -273,16 +285,16 @@ class Scene(Module):
         except ImportError as err:
             raise ImportError("scarlet2.Scene.fit() requires optax and numpyro.") from err
 
-        parameters = self.parameters
         # making sure we can iterate
         if not isinstance(observations, (list, tuple)):
             observations = (observations,)
-        assert isinstance(parameters, Parameters)
 
         # make a stepsize tree
-        where = lambda model: model.get(parameters)
-        replace = tuple(p.stepsize for p in parameters)
-        steps = eqx.tree_at(where, self, replace=replace)
+        ptree = self.parameters.as_tree()
+        steps = jtu.tree_map(lambda node: node.stepsize if isinstance(node, Parameter) else None, ptree)
+        # where = lambda model: model.get_parameters()
+        # replace = tuple(p.stepsize for p in self.parameters)
+        # steps = eqx.tree_at(where, self, replace=replace)
 
         def scale_by_stepsize() -> base.GradientTransformation:
             # adapted from optax.scale_by_param_block_norm()
@@ -313,17 +325,19 @@ class Scene(Module):
         )
 
[…]
