+++ src/scarlet2/scene.py
@@ -1,3 +1,4 @@
+import copy
 from collections import defaultdict
 
 import equinox as eqx
@@ -188,7 +189,6 @@ class Scene(Module):
                 return self.obs._log_likelihood(self.model, value)
 
         # find all non-fixed parameters and their priors
-        ptree = self.parameters.as_tree()
         priors = {p.name: p.prior for p in self.parameters}
         has_none = any(prior is None for prior in priors.values())
         if has_none:
@@ -204,7 +204,7 @@ class Scene(Module):
         def pyro_model(model):
             # samples = tuple(numpyro.sample(p.name, p.prior) for p in self.parameters)
             # model_ = model.replace_parameters(samples)
-            model_ = jtu.tree_map(sample_node, model, ptree)
+            model_ = jtu.tree_map(sample_node, model, self.parameters.tree)
             pred = model_()  # create prediction once for all observations
             # dealing with multiple observations
             for i, obs_ in enumerate(observations):
@@ -289,12 +289,12 @@ class Scene(Module):
         if not isinstance(observations, (list, tuple)):
             observations = (observations,)
 
-        # make a stepsize tree
-        ptree = self.parameters.as_tree()
-        steps = jtu.tree_map(lambda node: node.stepsize if isinstance(node, Parameter) else None, ptree)
-        # where = lambda model: model.get_parameters()
-        # replace = tuple(p.stepsize for p in self.parameters)
-        # steps = eqx.tree_at(where, self, replace=replace)
+        # make schedule with the parameter stepsizes
+        steps = jtu.tree_map(
+            lambda node: node.stepsize if isinstance(node, Parameter) else None,
+            self.parameters.tree,
+            is_leaf=lambda x: x is None,
+        )
 
         def scale_by_stepsize() -> base.GradientTransformation:
             # adapted from optax.scale_by_param_block_norm()
@@ -325,11 +325,13 @@ class Scene(Module):
         )
 
         # transform to unconstrained parameters
-        scene = _constraint_replace(self, ptree, inv=True)
+        scene = _constraint_replace(self, self.parameters, inv=True)
 
         # get optimizer initialized with the optimization parameters
         filter_spec = jtu.tree_map(
-            lambda node: True if isinstance(node, Parameter) else False, ptree, is_leaf=lambda x: x is None
+            lambda node: isinstance(node, Parameter),
+            self.parameters.tree,
+            is_leaf=lambda x: x is None,
         )
         opt_state = optim.init(eqx.filter(scene, filter_spec))
 
@@ -337,7 +339,7 @@ class Scene(Module):
             for step in t:  # noqa: B007
                 # optimizer step
                 scene, loss, opt_state, convergence = _make_step(
-                    scene, observations, optim, opt_state, ptree, filter_spec=filter_spec
+                    scene, observations, self.parameters, filter_spec, optim, opt_state
                 )
 
                 # compute max change across all non-fixed parameters for convergence test
@@ -345,7 +347,7 @@ class Scene(Module):
 
                 # report current iteration results to callback
                 if callback is not None:
-                    scene_ = _constraint_replace(scene, ptree)
+                    scene_ = _constraint_replace(scene, self.parameters.tree)
                     callback(scene_, convergence, loss)
 
                 # Log the loss and max_change in the tqdm progress bar
@@ -355,7 +357,11 @@ class Scene(Module):
                 if max_change < e_rel:
                     break
 
-        returned_scene = _constraint_replace(scene, ptree)  # transform back to constrained variables
+        # transform back to constrained variables and attach parameters
+        returned_scene = _constraint_replace(scene, self.parameters)
+        new_params = copy.deepcopy(self.parameters)
[…]
