+++ src/scarlet2/module.py
@@ -1,6 +1,5 @@
 import astropy.units as u
 import equinox as eqx
-import jax
 import jax.numpy as jnp
 import jax.tree_util as jtu
 import varname
@@ -13,6 +12,11 @@ from .validation_utils import (
     ValidationResult,
     print_validation_results,
 )
+
+
+def _tree_select(func, pytree, *rest):
+    leaves = jtu.tree_leaves(jtu.tree_map(func, pytree, *rest, is_leaf=lambda x: x is None))
+    return tuple(_ for _ in leaves if _ is not None)
 
 
 class Module(eqx.Module):
@@ -43,59 +47,94 @@ class Module(eqx.Module):
         object.__setattr__(self, "parameters", parameters)
         return self
 
-    def get_parameters(self, ptree):
+    def get(self, p):
         """Get parameter arrays from this module
+
+        Parameters
+        ----------
+        p: (:py:class:`Parameters`, :py:class:`Parameter`, str)
+            optimization parameters to extract
 
         Returns
         -------
         tuple
             requested data arrays for `parameters`
         """
+        # this method is uesd during optimization where the module is a copy without parameter attribute
+        # hence is must be passed as argument
+        if isinstance(p, Parameters):
+            get_node = lambda node, pnode: node if isinstance(pnode, Parameter) else None
+            return _tree_select(get_node, self, p.tree)
+
+        # lookup by name or individual parameter is used for checking/validation
+        # as they don't know what module/parameters they belong to, it only works for modules with .parameters
         if not hasattr(self, "parameters"):
-            return tuple()
+            raise AttributeError("Module.get requires parameters attribute to be set for this module.")
+        parameters = self.parameters
 
-        get_node = lambda node, p: node if isinstance(p, Parameter) else None
-        leaves = jtu.tree_leaves(jtu.tree_map(get_node, self, ptree))
-        return tuple(l for l in leaves if l is not None)
+        if isinstance(p, Parameter):
+            get_node = lambda node, pnode: node if pnode is p else None
+        elif isinstance(p, str):
+            get_node = lambda node, pnode: node if (pnode is not None and pnode.name == p) else None
+        plist = _tree_select(get_node, self, parameters.tree)
+        if len(plist) == 0:
+            raise IndexError(f"Name {p} not found in parameters")
+        return plist[0]
 
-    def replace_parameters(self, values):
-        """Replace parameter arrays by another value
 
-        Parameters
-        ----------
-        values: list
-            List of values to replace the current parameter arrays with
-            Needs to be in the same shape and order as the current parameter values
+def _to_pixels(frame, field):
+    """Convert parameter to pixel coordinates of the model frame
 
-        Returns
-        -------
-        :py:class:`Module`
-            Modified module. All other module components are unchanged.
-        """
-        where = lambda model: model.get_parameters()  # noqa: E731
-        return eqx.tree_at(where, self, replace=values)
+    scarlet2 models are optimized in pixel coordinates (defined by the model
+    frame of :py:class:`~scarlet2.Scene`. Therefore parameters (or their priors,
+    stepsize, etc) that are defined in :py:mod:`astropy.units` or
+    :py:class:`astropy.SkyCoord` need to be transformed to pixel coordinates.
 
-    def get_filter_spec(self):
-        """Get equinox filter_spec for all fields named in parameters
+    See details in issue :issue:`51`.
 
-        Parameters
-        ----------
-        parameters: :py:class:`Parameters`
-            List of parameters to search
+    Parameters
+    ----------
+    frame: :py:class:`~scarlet2.Frame`
+        Frame to define sky coordinates
+    field: any
+        Attribute of parameter to be converted to pixel coordinates
 
-        Returns
-        -------
-        list
-            requested data arrays for `parameters`
-        """
-        if not hasattr(self, "parameters"):
[…]
