### 2026-02-11 14f3a73
moved sample & fit to infer.py
