+++ src/scarlet2/validation.py
@@ -1,6 +1,6 @@
+from .infer import FitValidator
 from .module import ParameterValidator
 from .observation import ObservationValidator
-from .scene import FitValidator
 from .source import SourceValidator
 from .validation_utils import ValidationResult
 
