+++ src/scarlet2/infer.py
@@ -1,0 +1,491 @@
+import copy
+from collections import defaultdict
+
+import equinox as eqx
+import jax
+import jax.numpy as jnp
+import jax.tree_util as jtu
+
+from .module import Parameter
+from .scene import Scene
+from .validation_utils import (
+    ValidationError,
+    ValidationInfo,
+    ValidationMethodCollector,
+    ValidationResult,
+    ValidationWarning,
+    print_validation_results,
+)
+
+
+def sample(scene, observations, seed=0, num_warmup=100, num_samples=200, progress_bar=True, **kwargs):
+    """Sample `parameters` of every source in `scene` to get posteriors given `observations`.
+
+    This method runs the HMC NUTS sampler from `numpyro` to get parameter
+    posteriors. It uses the likelihood of `observations` as well as the `prior`
+    attribute set for every :py:class:`~scarlet2.Parameter` in `parameters`.
+
+    Parameters
+    ----------
+    scene : :py:class:`~scarlet2.Scene`
+        The model of the scene.
+    observations: :py:class:`~scarlet2.Observation` or list
+        The observations to fit the models to.
+    seed: int, optional
+        RNG seed for the sampler
+    num_warmup: int, optional
+        Number of samples during HMC warm-up
+    num_samples: int, optional
+        Number of samples to create from tuned HMC
+    progress_bar: bool, optional
+        Whether to show a progress bar
+    **kwargs: dict, optional
+        Additional keyword arguments passed to the `numpyro.infer.NUTS` sampler.
+
+    Notes
+    -----
+    Requires `numpyro`
+
+    Returns
+    -------
+    numpyro.infer.mcmc.MCMC
+    """
+    # uses numpyro NUTS on all non-fixed parameters
+    # requires that those have priors set
+
+    # no optimizable parameters: return self
+    if not hasattr(scene, "parameters"):
+        return scene
+
+    try:
+        import numpyro
+        import numpyro.distributions as dist
+        import numpyro.distributions.constraints as constraints
+        from numpyro.infer import MCMC, NUTS
+    except ImportError as err:
+        raise ImportError("scarlet2.Scene.sample() requires numpyro.") from err
+
+    # making sure we can iterate
+    if not isinstance(observations, (list, tuple)):
+        observations = (observations,)
+
+    # helper class to turn observation likelihood(s) into numpyro distribution
+    class ObsDistribution(dist.Distribution):
+        support = constraints.real_vector
+
+        def __init__(self, obs, model, validate_args=None):
+            self.obs = obs
+            self.model = model
+            event_shape = jnp.shape(model)
+            super().__init__(
+                event_shape=event_shape,
+                validate_args=validate_args,
+            )
+
+        def sample(self, key, sample_shape=()):
+            raise NotImplementedError
+
+        def mean(self):
+            return self.obs.render(self.model)
+
+        @dist.util.validate_sample
+        def log_prob(self, value):
+            # numpyro needs sampling distribution of data (=value), not likelihood function of parameters
+            return self.obs._log_likelihood(self.model, value)
+
+    # find all non-fixed parameters and their priors
+    priors = {p.name: p.prior for p in scene.parameters}
+    has_none = any(prior is None for prior in priors.values())
+    if has_none:
+        from pprint import pformat
+
+        msg = f"All parameters need to have priors set. Got:\n{pformat(priors)}"
+        raise AttributeError(msg)
+
+    # define the pyro model, where every parameter becomes a sample,
+    # and the observations sample from their likelihood given the rendered model
+    sample_node = lambda node, p: numpyro.sample(p.name, p.prior) if isinstance(p, Parameter) else node
+
+    def pyro_model(model):
+        # samples = tuple(numpyro.sample(p.name, p.prior) for p in self.parameters)
+        # model_ = model.replace_parameters(samples)
[…]
