### 2026-02-11 64c636c
Renderer and Frame inherit from sc2.Module, not eqx
