### 2026-02-12 95da17f
ConvolutionRenderer has shift parameter
