+++ tests/scarlet2/test_scene_fit_checks.py
@@ -1,7 +1,8 @@
 from unittest.mock import patch
 
 import pytest
-from scarlet2.scene import FitValidator
+
+from scarlet2.infer import FitValidator, fit
 from scarlet2.validation_utils import (
     ValidationError,
     ValidationInfo,
@@ -25,10 +26,10 @@ def setup_validation():
         (10.0, ValidationError),
     ],
 )
-def test_check_goodness_of_fit(scene, parameters, good_obs, mocked_chi_value, expected):
+def test_check_goodness_of_fit(scene, good_obs, mocked_chi_value, expected):
     """Mocked goodness_of_fit return to produces the expected ValidationResult type"""
 
-    scene_ = scene.fit(good_obs, parameters, max_iter=1, e_rel=1e-4, progress_bar=True)
+    scene_ = fit(scene, good_obs, max_iter=1, e_rel=1e-4, progress_bar=True)
     checker = FitValidator(scene_, good_obs)
 
     with patch.object(type(good_obs), "goodness_of_fit", return_value=mocked_chi_value) as _:
@@ -45,10 +46,10 @@ def test_check_goodness_of_fit(scene, parameters, good_obs, mocked_chi_value, ex
         (10.0, ValidationError),
     ],
 )
-def test_check_chi_squared_in_box_and_border(scene, parameters, good_obs, mocked_chi_value, expected):
+def test_check_chi_squared_in_box_and_border(scene, good_obs, mocked_chi_value, expected):
     """Mocked chi-squared evaluation in box and border."""
 
-    scene_ = scene.fit(good_obs, parameters, max_iter=1, e_rel=1e-4, progress_bar=True)
+    scene_ = fit(scene, good_obs, max_iter=1, e_rel=1e-4, progress_bar=True)
     checker = FitValidator(scene_, good_obs)
     mock_return = {0: {"in": mocked_chi_value, "out": mocked_chi_value}}
 
