### 2026-02-13 093cfc4
updated notebooks to new API
