### 2026-02-13 b10332f
create backward compatible calls for Scene.fit/sample
