### 2026-02-14 a609e5d
expand parameters when source is added to scene
