### 2026-02-17 bbe1e18
fit optimizes scene and observations
