### 2026-02-21 d812f6b
parameters as dict; live lookup for "is node" not working for traced arrays
