+++ src/scarlet2/infer.py
@@ -1,12 +1,11 @@
-import copy
-from collections import defaultdict
+import operator
 
 import equinox as eqx
 import jax
 import jax.numpy as jnp
 import jax.tree_util as jtu
 
-from .module import Module, Parameter, Parameters
+from .module import Parameter
 from .scene import Scene
 from .validation_utils import (
     ValidationError,
@@ -16,25 +15,6 @@ from .validation_utils import (
     ValidationWarning,
     print_validation_results,
 )
-
-
-# helper class to optimize scene and observation parameters
-class _Model(Module):
-    scene: Scene
-    observations: list
-
-    def __init__(self, scene, observations):
-        self.scene = scene
-        self.observations = observations
-
-        # define empty parameters
-        Parameters(self)
-        # add scene and observation parameters if present
-        if hasattr(scene, "parameters"):
-            object.__setattr__(self.parameters.tree, "scene", scene.parameters.tree)
-        for i, obs in enumerate(observations):
-            if hasattr(obs, "parameters"):
-                self.parameters.tree.observations[i] = obs.parameters.tree
 
 
 def sample(scene, observations, seed=0, num_warmup=100, num_samples=200, progress_bar=True, **kwargs):
@@ -156,9 +136,21 @@ def sample(scene, observations, seed=0, num_warmup=100, num_samples=200, progres
     return mcmc
 
 
+def _dict_to_eqx_module(name: str, data: dict):
+    # Build field annotations dynamically
+    annotations = {k: type(v) for k, v in data.items()}
+
+    # Create a class that inherits from eqx.Module
+    cls = type(name, (eqx.Module,), {"__annotations__": annotations})
+
+    # Instantiate with the dict values
+    return cls(**data)
+
+
 def fit(
     scene,
     observations,
+    *args,
     schedule=None,
     max_iter=100,
     e_rel=1e-4,
@@ -177,6 +169,8 @@ def fit(
         The model of the scene.
     observations: :py:class:`~scarlet2.Observation` or list
         The observations to fit the model to.
+    *args: list, optional
+        Additional arguments passed. Only used for backwards (v0.3) compatibility.
     schedule: callable, optional
         A function that maps optimizer step count to value. See :py:class:`optax.Schedule` for details.
     max_iter: int, optional
@@ -213,21 +207,24 @@ def fit(
     # making sure we can iterate
     if not isinstance(observations, (list, tuple)):
         observations = (observations,)
+    obs_params = {}
     for obs in observations:
         obs.check_set_renderer(scene.frame)
+        obs_params.update(obs.parameters)
 
     # scene and observations can have parameters: combine them into one model
-    model = _Model(scene, observations)
-    parameters = model.parameters
+    parameters = scene.parameters | obs_params
     if len(parameters) == 0:
-        return model
+        return scene
 
-    # make schedule with the parameter stepsizes
-    steps = jtu.tree_map(
-        lambda node: node.stepsize if isinstance(node, Parameter) else None,
-        parameters.tree,
-        is_leaf=lambda x: x is None,
-    )
+    # construct eqx.Module containing all parameter arrays
+    values = {name: node for name, (node, params) in parameters.items()}
+    values = _dict_to_eqx_module("ParamModel", values)
+    # same tree structure but for param specs so that we can use them below
+    treedef = jtu.tree_structure(values)
+    params = tuple(param for name, (node, param) in parameters.items())
+    params = jtu.build_tree(treedef, params)
+    steps = jtu.tree_map(lambda param: param.stepsize, params)
 
     def scale_by_stepsize() -> base.GradientTransformation:
         # adapted from optax.scale_by_param_block_norm()
@@ -257,29 +254,25 @@ def fit(
         scale_by_stepsize(),
     )
 
-    # transform to unconstrained parameters (undo monkey patch: no .parameters)
[…]
