+++ src/scarlet2/module.py
@@ -1,3 +1,5 @@
+import re
+
 import astropy.units as u
 import equinox as eqx
 import jax.numpy as jnp
@@ -5,18 +7,12 @@ import jax.tree_util as jtu
 import varname
 from astropy.coordinates import SkyCoord
 
-from . import Parameterization
+from . import Parameterization, parameter_registry
 from .validation_utils import (
     ValidationError,
     ValidationMethodCollector,
     ValidationResult,
-    print_validation_results,
 )
-
-
-def _tree_select(func, pytree, *rest):
-    leaves = jtu.tree_leaves(jtu.tree_map(func, pytree, *rest, is_leaf=lambda x: x is None))
-    return tuple(_ for _ in leaves if _ is not None)
 
 
 class Module(eqx.Module):
@@ -26,92 +22,68 @@ class Module(eqx.Module):
     and adds extra functionality to deal with optimizable parameters.
     """
 
+    registry_key: str = eqx.field(init=False, default="", repr=False)
+
     def __call__(self):
         """Evaluate the model"""
         raise NotImplementedError
 
-    def set_parameters(self, parameters):
-        """Define parameters for this module
-
-        Parameters
-        ----------
-        parameters: :py:class:`Parameters`
-            Optimization parameters
+    @property
+    def parameters(self):
+        """Parameters defined for this module
 
         Returns
         -------
-        self
+        dict
+            name: (node, param) mapping for all parameters
         """
-        # Monkey patching parameters onto base model:
-        # allows to move the parameters with the model without eqx replicating them during fit/sample
-        # TODO
-        # parameters_ = copy.deepcopy(parameters)
-        # parameters_.base = self
-        object.__setattr__(self, "parameters", parameters)
-        return self
+        return parameter_registry.get(self.registry_key, dict())
 
-    def get(self, p):
-        """Get parameter(s) `p` from this module
-
-        Parameters
-        ----------
-        p: (:py:class:`Parameters`, :py:class:`Parameter`, str)
-            optimization parameters to extract
+    def get_parameters(self):
+        """Get parameter(s) from this module
 
         Returns
         -------
-        tuple
-            requested data arrays for `parameters`
+        dict
+            requested data arrays for `parameters`, ordered by their `name`
         """
-        # this method is used during optimization/sampling where the module is a copy without .parameters
-        # hence is must be passed as argument
-        if isinstance(p, Parameters):
-            get_node = lambda node, pnode: node if isinstance(pnode, Parameter) else None
-            return _tree_select(get_node, self, p.tree)
+        return {name: node for name, (node, param) in self.parameters.items()}
 
-        # lookup by name or individual parameter is used for checking/validation
-        # as they don't know what module/parameters they belong to, it only works for modules with .parameters
-        if not hasattr(self, "parameters"):
-            raise AttributeError("Module.get requires parameters attribute to be set for this module.")
-        parameters = self.parameters
-
-        if isinstance(p, Parameter):
-            get_node = lambda node, pnode: node if pnode is p else None
-        elif isinstance(p, str):
-            get_node = lambda node, pnode: node if (pnode is not None and pnode.name == p) else None
-        plist = _tree_select(get_node, self, parameters.tree)
-        if len(plist) == 0:
-            raise IndexError(f"Name {p} not found in parameters")
-        return plist[0]
-
-    def replace(self, p, v):
-        """Replace parameter(s) `p` from this module with `v`
+    def replace(self, values):
+        """Replace parameter(s) from this module with `values`
 
         Parameters
         ----------
-        p: (:py:class:`Parameters`, :py:class:`Parameter`, str)
-            optimization parameters to replace
-        v: (jnp.array, list[jnp.array])
[…]

[…]
