### 2026-02-21 716f1c1
fixed sample method and fit validation test
