+++ src/scarlet2/infer.py
@@ -1,11 +1,15 @@
 import operator
+from pprint import pformat
 
 import equinox as eqx
 import jax
 import jax.numpy as jnp
 import jax.tree_util as jtu
+import numpyro
+import numpyro.distributions as dist
+import numpyro.distributions.constraints as constraints
+from numpyro.infer import MCMC, NUTS
 
-from .module import Parameter
 from .scene import Scene
 from .validation_utils import (
     ValidationError,
@@ -17,7 +21,43 @@ from .validation_utils import (
 )
 
 
-def sample(scene, observations, seed=0, num_warmup=100, num_samples=200, progress_bar=True, **kwargs):
+# helper class to turn observation likelihood(s) into numpyro distribution
+class _ObsDistribution(dist.Distribution):
+    support = constraints.real_vector
+
+    def __init__(self, obs, model, validate_args=None):
+        self.obs = obs
+        self.model = model
+        event_shape = jnp.shape(model)
+        super().__init__(
+            event_shape=event_shape,
+            validate_args=validate_args,
+        )
+
+    def sample(self, key, sample_shape=()):
+        raise NotImplementedError
+
+    def mean(self):
+        return self.obs.render(self.model)
+
+    @dist.util.validate_sample
+    def log_prob(self, value):
+        # numpyro needs sampling distribution of data (=value), not likelihood function of parameters
+        return self.obs._log_likelihood(self.model, value)
+
+
+def _dict_to_eqx_module(name: str, data: dict):
+    # Build field annotations dynamically
+    annotations = {k: type(v) for k, v in data.items()}
+
+    # Create a class that inherits from eqx.Module
+    cls = type(name, (eqx.Module,), {"__annotations__": annotations})
+
+    # Instantiate with the dict values
+    return cls(**data)
+
+
+def sample(scene, observations, *args, seed=0, num_warmup=100, num_samples=200, progress_bar=True, **kwargs):
     """Sample `parameters` of every source in `scene` to get posteriors given `observations`.
 
     This method runs the HMC NUTS sampler from `numpyro` to get parameter
@@ -30,6 +70,8 @@ def sample(scene, observations, seed=0, num_warmup=100, num_samples=200, progres
         The model of the scene.
     observations: :py:class:`~scarlet2.Observation` or list
         The observations to fit the models to.
+    *args: list, optional
+        Additional arguments passed. Only used for backwards (v0.3) compatibility.
     seed: int, optional
         RNG seed for the sampler
     num_warmup: int, optional
@@ -49,102 +91,59 @@ def sample(scene, observations, seed=0, num_warmup=100, num_samples=200, progres
     -------
     numpyro.infer.mcmc.MCMC
     """
-    # uses numpyro NUTS on all non-fixed parameters
-    # requires that those have priors set
-
-    # no optimizable parameters: return self
-    if not hasattr(scene, "parameters"):
-        return scene
-
-    try:
-        import numpyro
-        import numpyro.distributions as dist
-        import numpyro.distributions.constraints as constraints
-        from numpyro.infer import MCMC, NUTS
-    except ImportError as err:
-        raise ImportError("scarlet2.Scene.sample() requires numpyro.") from err
-
     # making sure we can iterate
     if not isinstance(observations, (list, tuple)):
         observations = (observations,)
+    obs_params = {}
     for obs in observations:
         obs.check_set_renderer(scene.frame)
+        obs_params.update(obs.parameters)
 
-    # helper class to turn observation likelihood(s) into numpyro distribution
-    class ObsDistribution(dist.Distribution):
-        support = constraints.real_vector
-
-        def __init__(self, obs, model, validate_args=None):
-            self.obs = obs
-            self.model = model
-            event_shape = jnp.shape(model)
-            super().__init__(
[…]
