### 2026-02-21 bc98b49
allowing Parameters to be reused and reset
