+++ tests/scarlet2/test_quickstart.py
@@ -3,84 +3,41 @@
 # ruff: noqa: D103
 # ruff: noqa: D106
 
-from functools import partial
-
-import jax.numpy as jnp
 import numpyro.distributions as dist
-from huggingface_hub import hf_hub_download
-from numpyro.distributions import constraints
+import pytest
 from numpyro.infer.initialization import init_to_sample
 
-from scarlet2 import fit, init
-from scarlet2.frame import Frame
-from scarlet2.module import Parameter, Parameters, relative_step
-from scarlet2.observation import Observation
-from scarlet2.psf import ArrayPSF, GaussianPSF
-from scarlet2.scene import Scene
-from scarlet2.source import Source
-from scarlet2.validation_utils import set_validation
-
-# turn off automatic validation checks
-set_validation(False)
-
-filename = hf_hub_download(
-    repo_id="astro-data-lab/scarlet-test-data", filename="hsc_cosmos_35.npz", repo_type="dataset"
+from scarlet2 import Parameter, Parameters
+from scarlet2.infer import fit
+from scarlet2.validation_utils import (
+    set_validation,
 )
-file = jnp.load(filename)
-data = jnp.asarray(file["images"])
-channels = [str(f) for f in file["filters"]]
-centers = [(src["y"], src["x"]) for src in file["catalog"]]  # Note: y/x convention!
-weights = jnp.asarray(1 / file["variance"])
-psf = jnp.asarray(file["psfs"])
-
-_ = GaussianPSF(0.7)
-obs = Observation(data, weights, psf=ArrayPSF(jnp.asarray(psf)), channels=channels)
-model_frame = Frame.from_observations(obs)
-
-with Scene(model_frame) as scene:
-    for center in centers:
-        center = jnp.array(center)
-        try:
-            spectrum, morph = init.from_gaussian_moments(obs, center, min_corr=0.99)
-        except ValueError:
-            spectrum = init.pixel_spectrum(obs, center)
-            morph = init.compact_morphology()
-
-        Source(center, spectrum, morph)
 
 
-def test_fit():
-    spec_step = partial(relative_step, factor=0.05)
+@pytest.fixture(autouse=True)
+def setup_validation():
+    """Automatically disable validation for all tests. This permits the creation
+    of intentionally invalid Observation objects."""
+    set_validation(False)
 
-    # fitting
-    with Parameters(scene):
-        for i in range(len(scene.sources)):
-            Parameter(
-                scene.sources[i].spectrum,
-                name=f"spectrum_{i}",
-                constraint=constraints.positive,
-                stepsize=spec_step,
-            )
-            Parameter(
-                scene.sources[i].morphology, name=f"morph_{i}", constraint=constraints.positive, stepsize=0.1
-            )
+
+def test_fit(scene, good_obs):
     max_iter = 10
-    scene_ = fit(scene, obs, max_iter=max_iter, progress_bar=False)
+    scene_ = fit(scene, good_obs, max_iter=max_iter, progress_bar=False)
     return scene_
 
 
-def test_sample(scene):
-    # use old API version to call sample()
-    with Parameters(scene):
+def test_sample(scene, good_obs):
+    with Parameters(scene):  # new parameters: only the one spectrum gets sampled
         p = scene.sources[0].spectrum
         prior = dist.Normal(p, scale=1).to_event(1)
         Parameter(p, name="spectrum:0", prior=prior)
     init_strategy = init_to_sample
     scene.sample(
-        obs, num_samples=10, num_warmup=10, dense_mass=True, init_strategy=init_strategy, progress_bar=False
+        good_obs,
+        num_samples=10,
+        num_warmup=10,
+        dense_mass=True,
[…]
