### 2026-02-21 8ca846a
reset Parameters if Scene gets new sources
