### 2026-02-22 c76f3a9
enable the use of observation parameters in fit
