### 2026-02-22 ab759af
enable the use of observation parameters in sample
