### 2026-02-24 8770dee
remove restriction of jax==0.6.2; confirmed working on python=3.12 and jax=0.7.2
