+++ pyproject.toml
@@ -18,7 +18,7 @@ dynamic = ["version"]
 requires-python = ">=3.10"
 dependencies = [
     "equinox",
-    "jax==0.6.2",
+    "jax",
     "astropy",
     "numpy",
     "matplotlib",
