### 2026-02-25 f2b671e
remove proximal thresholds and add proper edge treatment in bspline_convolve
