### 2026-02-26 7178946
added scales to starelet_reconstruction; docs  call for constraint/prior
