### 2026-02-26 b2c091d
Stable starlets (#249)
* remove proximal thresholds and add proper edge treatment in bspline_convolve

* added scales to starelet_reconstruction; docs  call for constraint/prior
