### 2026-02-26 07bbfa2
allow constraint and prior on the same parameter (closes #247)
