### 2026-03-02 #9646 @pmelchior
Thanks to @AlexandreAdam and @Jammy2211 for your constructive reviews. I'll respond to the issues in turn:

### Review 1

* I added CONTRIBUTING.md and a link to the docs in README.md in the latest commit to main.
* The inconsistency between release version and documentation is fixed. This was a misconfigured RTD, which defaulted to "latest" rather than "stable" branch. This should also address the `AttributeError` in the Quick Start Guide.
* I added a comparison to other similar modeling codes in astronomy to the latest version of the paper.
* I lifted the temporary restrictions to jax=0.6.2 (see https://github.com/pmelchior/scarlet2/pull/246).
* The HDF5 file with the model fit is in the correct place. The notebook on posterior sampling is in the directory "howto" and therefore must search in its parent directory.

### Review 2

* The "state of the field" section has been added to the new paper draft.
* I can't reproduce the error in "plot.observation". Was this in a development version?
* It's very valuable to hear of the difficulty in understanding what type of model is being fit. I am preparing a new quick start guide, which only deals with the essentials, while the current, more involved example will be moved to "A real-world example" notebook.
