+++ README.md
@@ -1,5 +1,8 @@
 # _scarlet2_
 
+[![PyPI](https://img.shields.io/pypi/v/scarlet2)](https://pypi.python.org/pypi/scarlet2/)
+[![Documentation Status](https://img.shields.io/badge/docs-latest-brightgreen.svg?style=flat&logo=readthedocs)](https://scarlet2.readthedocs.io/)
+
 _scarlet2_ is an open-source python library for modeling astronomical sources from multi-band, multi-epoch, and
 multi-instrument data. It provides non-parametric and parametric models, can handle source overlap (aka blending), and
 can integrate neural network priors. It's designed to be modular, flexible, and powerful.
