### 2026-03-02 a6a6de5
respond to JOSS review (#253)
