+++ CONTRIBUTING.md
@@ -1,0 +1,217 @@
+# Contributing to scarlet2
+
+Thank you for your interest in contributing to **scarlet2**. Whether you're fixing a bug, proposing a new
+feature, improving documentation, or sharing a use case, your contributions are welcome.
+
+---
+
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [Ways to Contribute](#ways-to-contribute)
+- [Getting Started](#getting-started)
+- [Development Workflow](#development-workflow)
+- [Coding Standards](#coding-standards)
+- [Tests](#tests)
+- [Documentation](#documentation)
+- [Submitting a Pull Request](#submitting-a-pull-request)
+- [Reporting Issues](#reporting-issues)
+- [Questions and Discussions](#questions-and-discussions)
+
+---
+
+## Code of Conduct
+
+This project follows a standard open-source code of conduct. Please be respectful, constructive, and inclusive
+in all interactions. Harassment or discriminatory behavior of any kind will not be tolerated.
+
+---
+
+## Ways to Contribute
+
+- **Bug reports**: Found something broken? Open an [issue](https://github.com/pmelchior/scarlet2/issues).
+- **Feature requests**: Have an idea for a new model, prior, or utility? Start
+  a [Discussion](https://github.com/pmelchior/scarlet2/discussions) first to gauge interest.
+- **Code contributions**: Fix a bug, implement a feature, or improve performance.
+- **Documentation**: Improve docstrings, tutorials, or the ReadTheDocs pages.
+- **Benchmarks**: Add or improve benchmarks for model performance.
+- **Examples and notebooks**: Share use cases (strong lensing, transients, deblending, etc.) as notebooks in
+  `docs/`.
+
+---
+
+## Getting Started
+
+### Prerequisites
+
+scarlet2 is built on [JAX](https://github.com/google/jax)
+and [equinox](https://github.com/patrick-kidger/equinox). Before setting up the dev environment, install JAX
+with the appropriate `jaxlib` for your platform (CPU, GPU, or TPU) following
+the [official JAX installation instructions](https://github.com/google/jax#installation).
+
+### Fork and Clone
+
+1. Fork the repository on GitHub.
+2. Clone your fork locally:
+   ```bash
+   git clone https://github.com/<your-username>/scarlet2.git
+   cd scarlet2
+   ```
+3. Add the upstream remote:
+   ```bash
+   git remote add upstream https://github.com/pmelchior/scarlet2.git
+   ```
+
+### Install in Development Mode
+
+Install scarlet2 along with its optional dependencies:
+
+```bash
+pip install -e ".[dev]"
+```
+
+For the full feature set (optimization, sampling, and I/O), also install:
+
+```bash
+pip install optax numpyro h5py
+```
+
+---
+
+## Development Workflow
+
+1. **Sync with upstream** before starting work:
+   ```bash
+   git fetch upstream
+   git checkout main
+   git merge upstream/main
+   ```
+
+2. **Create a feature branch**:
+   ```bash
+   git checkout -b my-feature-or-fix
+   ```
+
+3. Make your changes, write tests, and update documentation as needed.
+
+4. **Run tests** to make sure nothing is broken (see [Tests](#tests)).
+
+5. **Commit** with a clear, descriptive message:
+   ```bash
+   git commit -m "Fix: correct morphology normalization in Frame.render"
+   ```
+
+6. **Push** your branch and open a pull request against `main`.
+
+---
+
+## Coding Standards
+
+- **Style**: Follow [PEP 8](https://peps.python.org/pep-0008/). You can use `ruff` or `flake8` for linting.
+- **Type hints**: Add type annotations where they improve clarity.
+- **Docstrings**: Use NumPy-style docstrings for all public functions and classes.
[…]
