+++ paper/paper.bib
@@ -169,3 +169,99 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{galfit,
+       author = {{Peng}, Chien Y. and {Ho}, Luis C. and {Impey}, Chris D. and {Rix}, Hans-Walter},
+        title = "{Detailed Structural Decomposition of Galaxy Images}",
+      journal = {\aj},
+     keywords = {Galaxies: Bulges, Galaxies: Fundamental Parameters, Galaxies: Nuclei, Galaxies: Structure, Techniques: Image Processing, Techniques: Photometric, Astrophysics},
+         year = 2002,
+        month = jul,
+       volume = {124},
+       number = {1},
+        pages = {266-293},
+          doi = {10.1086/340952},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0204182},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2002AJ....124..266P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{galapagos,
+       author = {{Barden}, Marco and {H{\"a}u{\ss}ler}, Boris and {Peng}, Chien Y. and {McIntosh}, Daniel H. and {Guo}, Yicheng},
+        title = "{GALAPAGOS: from pixels to parameters}",
+      journal = {\mnras},
+     keywords = {methods: data analysis, surveys, galaxies: statistics, galaxies: structure, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Extragalactic Astrophysics},
+         year = 2012,
+        month = may,
+       volume = {422},
+       number = {1},
+        pages = {449-468},
+          doi = {10.1111/j.1365-2966.2012.20619.x},
+archivePrefix = {arXiv},
+       eprint = {1203.1831},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2012MNRAS.422..449B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{galapagos2,
+       author = {{H{\"a}u{\ss}ler}, Boris and {Bamford}, Steven P. and {Vika}, Marina and {Rojas}, Alex L. and {Barden}, Marco and {Kelvin}, Lee S. and {Alpaslan}, Mehmet and {Robotham}, Aaron S.~G. and {Driver}, Simon P. and {Baldry}, I.~K. and {Brough}, Sarah and {Hopkins}, Andrew M. and {Liske}, Jochen and {Nichol}, Robert C. and {Popescu}, Cristina C. and {Tuffs}, Richard J.},
+        title = "{MegaMorph - multiwavelength measurement of galaxy structure: complete S{\'e}rsic profile information from modern surveys}",
+      journal = {\mnras},
+     keywords = {methods: data analysis, techniques: image processing, galaxies: fundamental parameters, galaxies: structure, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2013,
+        month = mar,
+       volume = {430},
+       number = {1},
+        pages = {330-369},
+          doi = {10.1093/mnras/sts633},
+archivePrefix = {arXiv},
+       eprint = {1212.3332},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2013MNRAS.430..330H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@software{tractor,
[…]
