### 2026-03-02 a0f8eb1
Added JOSS badge
