### 2026-03-02 #9646 @pmelchior
@ivastar The new version of the paper is online in the `joss` branch of scarlet2.
