### 2026-03-03 4c9b7fe
better treatment and validation of invalid pixels in observation (redo of #89)
