### 2026-03-03 bc401a1
better treatment and validation of invalid pixels in observation (redo of #89) (#255)
