### 2026-03-03 54f18af
Parameters as dict (#245)
* removed obsolete apply_constraint and OK parameter checks

* monkey patching parameters onto Module; inference does not need parameter arg

* added tree form of parameters

* removed node from Parameter: lookup by tree structure

* moved sample & fit to infer.py

* Renderer and Frame inherit from sc2.Module, not eqx

* ConvolutionRenderer has shift parameter

* updated notebooks to new API

* create backward compatible calls for
Scene.fit/sample

* expand parameters when source is added to scene

* fit optimizes scene and observations

* parameters as dict; live loo
[…]
