+++ src/scarlet2/module.py
@@ -1,20 +1,18 @@
-import logging
+import re
 
 import astropy.units as u
 import equinox as eqx
-import jax
 import jax.numpy as jnp
 import jax.tree_util as jtu
 import varname
 from astropy.coordinates import SkyCoord
+from numpyro.distributions.transforms import biject_to
 
-from . import Parameterization
+from . import Parameterization, parameter_registry
 from .validation_utils import (
     ValidationError,
-    ValidationInfo,
     ValidationMethodCollector,
     ValidationResult,
-    print_validation_results,
 )
 
 
@@ -25,75 +23,149 @@ class Module(eqx.Module):
     and adds extra functionality to deal with optimizable parameters.
     """
 
+    registry_key: str = eqx.field(init=False, default="", repr=False)
+
     def __call__(self):
         """Evaluate the model"""
         raise NotImplementedError
 
-    def make_parameters(self):
-        """Construct :py:class:`Parameters` for this module"""
-        return Parameters(self)
-
-    def get(self, parameters):
-        """Get parameter arrays from this module
-
-        Parameters
-        ----------
-        parameters: :py:class:`Parameters`
-            List of parameters to search
+    @property
+    def parameters(self):
+        """Parameters defined for this module
 
         Returns
         -------
-        list
+        dict
+            name: (node, param) mapping for all parameters
+        """
+        return parameter_registry.get(self.registry_key, dict())
+
+    def get(self, name=None):
+        """Get parameter(s) from this module
+
+        Parameters
+        ----------
+        name: str, optional
+            Name of parameter. If not set, returns all parameters.
+
+        Returns
+        -------
+        dict
             requested data arrays for `parameters`
         """
-        assert isinstance(parameters, Parameters)
-        return parameters.extract_from(self)
+        leaves = jtu.tree_leaves(self)
+        if name is None:
+            return {name: leaves[idx] for name, (idx, param) in self.parameters.items()}
+        else:
+            if name in self.parameters:
+                idx, param = self.parameters[name]
+                return leaves[idx]
+            else:
+                return None
 
-    def replace(self, parameters, values):
-        """Replace parameter arrays by another value
-
-        Parameters
-        ----------
-        parameters: :py:class:`Parameters`
-            List of parameters to search
-        values: list
-            List of values to replace the current parameter arrays with
-            Needs to be in the same shape and order as the current parameter values
-
-        Returns
-        -------
-        :py:class:`Module`
-            Modified module. All other module components are unchanged.
-        """
-        where = lambda model: model.get(parameters)  # noqa: E731
-        return eqx.tree_at(where, self, replace=values)
-
-    def get_filter_spec(self, parameters):
-        """Get equinox filter_spec for all fields named in parameters
+    def set(self, values):
+        """Set parameter(s) from this module with `values`
 
         Parameters
         ----------
-        parameters: :py:class:`Parameters`
-            List of parameters to search
+        values: dict[str,jnp.array]
+            values to replace parameters with, identified by their `name`
 
         Returns
         -------
-        list
-            requested data arrays for `parameters`
+        Module:
+            new module with parameter(s) replaced by `values`
         """
-        filtered = jax.tree_util.tree_map(lambda _: False, self)
-        where = lambda model: model.get(parameters)  # noqa: E731
-        values = (True,) * len(parameters)
-        filtered = eqx.tree_at(where, filtered, replace=values)
-        if all(jax.tree_util.tree_leaves(filtered)):
-            return None
-        return filtered
+        # .get_parameters produces values as dict, but infer.fit requires a values dataclass
+        values_ = values if isinstance(values, dict) else values.__dict__
+
[…]
