### 2026-03-04 ad3002a
added simplified quickstart; improved validation messages
