+++ src/scarlet2/observation.py
@@ -467,72 +467,6 @@ class ObservationValidator(metaclass=ValidationMethodCollector):
     def __init__(self, observation: Observation):
         self.observation = observation
 
-    def check_weights_non_negative(self) -> ValidationResult:
-        """Check that the weights in the observation are non-negative.
-
-        Returns
-        -------
-        ValidationResult
-            A subclass of ValidationResult indicating the result of the check.
-        """
-        if self.observation.weights is not None and (self.observation.weights < 0).any():
-            return ValidationError(
-                message="Weights in the observation must be non-negative.",
-                check=self.__class__.__name__,
-                context={"observation.weights": self.observation.weights},
-            )
-        else:
-            return ValidationInfo(
-                message="Weights in the observation are non-negative.",
-                check=self.__class__.__name__,
-            )
-
-    def check_weights_finite(self) -> ValidationResult:
-        """Check that the weights in the observation are finite.
-
-        Returns
-        -------
-        ValidationResult
-            A subclass of ValidationResult indicating the result of the check.
-        """
-        if self.observation.weights is not None and jnp.isinf(self.observation.weights).any():
-            return ValidationError(
-                message="Weights in the observation must be finite.",
-                check=self.__class__.__name__,
-                context={"observation.weights": self.observation.weights},
-            )
-        else:
-            return ValidationInfo(
-                message="Weights in the observation are finite.",
-                check=self.__class__.__name__,
-            )
-
-    def check_data_and_weights_shape(self) -> ValidationResult:
-        """Check that the data and weights have the same shape.
-
-        Returns
-        -------
-        ValidationResult
-            A subclass of ValidationResult indicating the result of the check.
-        """
-        if (
-            self.observation.weights is not None
-            and self.observation.data.shape != self.observation.weights.shape
-        ):
-            return ValidationError(
-                message="Data and weights must have the same shape.",
-                check=self.__class__.__name__,
-                context={
-                    "observation.data.shape": self.observation.data.shape,
-                    "observation.weights.shape": self.observation.weights.shape,
-                },
-            )
-        else:
-            return ValidationInfo(
-                message="Data and weights have the same shape.",
-                check=self.__class__.__name__,
-            )
-
     def check_num_channels_matches_data(self) -> ValidationResult:
         """Check that the number of channels in the observation matches the data.
 
@@ -561,6 +495,58 @@ class ObservationValidator(metaclass=ValidationMethodCollector):
                 check=self.__class__.__name__,
             )
 
+    def check_data_and_weights_shape(self) -> ValidationResult:
+        """Check that the data and weights exist and have the same shape.
+
+        Returns
+        -------
+        ValidationResult
+            A subclass of ValidationResult indicating the result of the check.
+        """
+        if self.observation.data is None:
+            return ValidationError(message="Observation.data are not set.", check=self.__class__.__name__)
+        if self.observation.weights is None:
+            return ValidationError(message="Observation.weights are not set.", check=self.__class__.__name__)
+        if self.observation.data.shape != self.observation.weights.shape:
+            return ValidationError(
+                message="Observation data and weights must have the same shape.",
+                check=self.__class__.__name__,
+                context={
+                    "observation.data.shape": self.observation.data.shape,
+                    "observation.weights.shape": self.observation.weights.shape,
+                },
+            )
+        else:
+            return ValidationInfo(
+                message="Data and weights have the same shape.",
+                check=self.__class__.__name__,
+            )
+
+    def check_weights_finite_and_non_negative(self) -> ValidationResult:
+        """Check that the weights in the observation are finite and non-negative.
+
+        Returns
+        -------
+        ValidationResult
[…]
