### 2026-03-05 0fb850a
added simplified quickstart; improved validation messages (#256)
* added simplified quickstart; improved validation messages

* switched to Sersic model; fixed error at the end
