### 2026-03-05 3ea3ae5
Better docs (#257)
* reactivated Parameter validation; copy editing in quickstart

* clean up notebooks
