### 2026-03-05 66d51cb
automatic conversion to jnp.array of type float
