+++ src/scarlet2/observation.py
@@ -39,13 +39,12 @@ class Observation(Module):
     """Name to describe the observation"""
 
     def __init__(self, data, weights, psf=None, wcs=None, channels=None, renderer=None, name=None):
-        # TODO: automatic conversion to jnp arrays
-        self.data = data
+        self.data = jnp.asarray(data).astype(float)
         if self.data.ndim == 2:
             # add a channel dimension if it is missing
             self.data = self.data[None, ...]
 
-        self.weights = weights
+        self.weights = jnp.asarray(weights).astype(float)
         if self.weights is not None and self.weights.ndim == 2:
             # add a channel dimension if it is missing
             self.weights = self.weights[None, ...]
