### 2026-03-05 b28184b
Add option to skip resampling the PSF in order to use oversampled PSF models with CorrelatedObservation
