### 2026-03-05 5c9677c
Add option to skip resampling the PSF with CorrelatedObservations (#258)
* Add option to skip resampling the PSF in order to use oversampled PSF models with CorrelatedObservation

* updated docstring

---------

Co-authored-by: Charlotte Ward <cvw5890@E1-055867.science.psu.edu>
Co-authored-by: Peter Melchior <peter.m.melchior@gmail.com>
