### 2026-03-05 1ccd413
replaced jax.tree_util by jax.tree (closes 259)
