### 2026-03-06 622f2dc
added UnitDisk constraint for ellipticity
