### 2026-03-06 54dd88b
Finishing touches for v0.5 (#260)
* automatic conversion to jnp.array of type float

* exposed kernel_fft in renderers (potentially optimizable)

* replaced jax.tree_util by jax.tree (closes 259)

* removed obsolete dependence on scipy.fftpack

* added UnitDisk constraint for ellipticity
