### 2026-03-06 a732515
Added constraint to API
