### 2026-03-06 8407d8d
silence ruff and align use recommended ruff lint rules with ruff format
