### 2026-03-06 98a9bc8
more ruff nonsense
