### 2026-03-12 #9646 @Jammy2211
Also happy to approve!
