### 2026-03-13 3f4c2e8
changes from copier update; import _version.py
