### 2026-03-13 64d4c31
silence ruff and align use recommended ruff lint rules with ruff format (#261)
* silence ruff and align use recommended ruff lint rules with ruff format

* more ruff nonsense

* changes from copier update; import _version.py
