### 2026-03-13 b9e15b2
added missing files from copier update (#262)
* added missing files from copier update

* added emptied notebook
