+++ tests/scarlet2/test_packaging.py
@@ -1,0 +1,6 @@
+import scarlet2
+
+
+def test_version():
+    """Check to see that we can get the package version"""
+    assert scarlet2.__version__ is not None
