### 2026-03-16 3d83355
added get_multiresolution_support
