### 2026-03-16 2b4767c
added detect.py
