+++ src/scarlet2/detect.py
@@ -1,0 +1,215 @@
+"""Detection utilities: connected-pixel footprint extraction and peak finding.
+
+Translated from scarlet v1's ``detect_pybind11.cc``. Uses NumPy rather than
+JAX because detection involves dynamic data structures (variable-length peak
+lists, irregularly shaped footprints) that are not compatible with JAX's JIT.
+Both NumPy and JAX arrays are accepted as inputs.
+"""
+
+from dataclasses import dataclass
+from typing import List
+
+import numpy as np
+
+
+@dataclass
+class Peak:
+    """A peak (local maximum) in a :class:`Footprint`.
+
+    Attributes
+    ----------
+    y : int
+        Row index in the full image.
+    x : int
+        Column index in the full image.
+    flux : float
+        Pixel value at the peak location.
+    """
+
+    y: int
+    x: int
+    flux: float
+
+
+@dataclass
+class Footprint:
+    """A detected footprint (connected region above threshold) in an image.
+
+    Attributes
+    ----------
+    footprint : ndarray of bool, shape (height, width)
+        Boolean mask of the footprint pixels, sized to the bounding box.
+    peaks : list of Peak
+        Peaks found within this footprint, sorted brightest-first.
+    bounds : ndarray of int, shape (4,)
+        Bounding box ``[y_min, y_max, x_min, x_max]`` in the full image.
+    """
+
+    footprint: np.ndarray
+    peaks: List[Peak]
+    bounds: np.ndarray
+
+
+def get_connected_pixels(i, j, image, unchecked, footprint, bounds, thresh=0):
+    """Find all pixels 4-connected to ``(i, j)`` that exceed ``thresh``.
+
+    Uses an iterative flood-fill to avoid Python's recursion limit on large
+    images. Modifies ``unchecked``, ``footprint``, and ``bounds`` in-place.
+
+    Parameters
+    ----------
+    i, j : int
+        Seed pixel coordinates (row, column).
+    image : 2D ndarray
+        The image to search.
+    unchecked : 2D boolean ndarray
+        Tracks unvisited pixels; updated in-place.
+    footprint : 2D boolean ndarray
+        Accumulates the footprint mask; updated in-place.
+    bounds : ndarray of int, shape (4,)
+        Bounding box ``[y_min, y_max, x_min, x_max]``; updated in-place.
+    thresh : float, optional
+        Pixels must strictly exceed this value to join the footprint.
+    """
+    height, width = image.shape
+    stack = [(i, j)]
+    while stack:
+        ci, cj = stack.pop()
+        if not unchecked[ci, cj]:
+            continue
+        unchecked[ci, cj] = False
+        if image[ci, cj] > thresh:
+            footprint[ci, cj] = True
+            if ci < bounds[0]:
+                bounds[0] = ci
+            elif ci > bounds[1]:
+                bounds[1] = ci
+            if cj < bounds[2]:
+                bounds[2] = cj
+            elif cj > bounds[3]:
+                bounds[3] = cj
+            if ci > 0:
+                stack.append((ci - 1, cj))
+            if ci < height - 1:
+                stack.append((ci + 1, cj))
+            if cj > 0:
+                stack.append((ci, cj - 1))
+            if cj < width - 1:
+                stack.append((ci, cj + 1))
+
+
+def get_peaks(image, min_separation, y0=0, x0=0):
+    """Find local maxima in an image patch.
+
+    A pixel is a local maximum if its value strictly exceeds all 8-connected
+    neighbours that lie within the image boundary.
+
+    Parameters
+    ----------
+    image : 2D ndarray
+        The (possibly masked) image patch to search for peaks.
+    min_separation : float
+        Minimum pixel distance between peaks. When two peaks are closer than
+        this, the dimmer one is removed.
[…]
