### 2026-03-18 6a3e4bb
added hierachical source detection; Box now standard python class
