+++ src/scarlet2/detect.py
@@ -1,15 +1,26 @@
 """Detection utilities: connected-pixel footprint extraction and peak finding.
 
-Translated from scarlet v1's ``detect_pybind11.cc``. Uses NumPy rather than
-JAX because detection involves dynamic data structures (variable-length peak
-lists, irregularly shaped footprints) that are not compatible with JAX's JIT.
-Both NumPy and JAX arrays are accepted as inputs.
+Low-level routines (``get_connected_pixels``, ``get_footprints``) are
+translated from scarlet v1's ``detect_pybind11.cc``.  Higher-level helpers
+(``get_wavelets``, ``QuadTreeRegion``, ``get_peaks``, …) are adapted from
+scarlet v1's ``detect.py``.
+
+Uses NumPy rather than JAX because detection involves dynamic data structures
+(variable-length peak lists, irregularly shaped footprints) that are not
+compatible with JAX's JIT.  Both NumPy and JAX arrays are accepted as inputs.
 """
 
-from dataclasses import dataclass
-from typing import List
+from dataclasses import dataclass, field
+from typing import List, Tuple
 
 import numpy as np
+
+from .bbox import Box, overlap_slices
+from .wavelets import get_multiresolution_support, starlet_transform
+
+# ---------------------------------------------------------------------------
+# Data classes
+# ---------------------------------------------------------------------------
 
 
 @dataclass
@@ -50,11 +61,16 @@ class Footprint:
     bounds: np.ndarray
 
 
+# ---------------------------------------------------------------------------
+# Low-level detection (translated from detect_pybind11.cc)
+# ---------------------------------------------------------------------------
+
+
 def get_connected_pixels(i, j, image, unchecked, footprint, bounds, thresh=0):
     """Find all pixels 4-connected to ``(i, j)`` that exceed ``thresh``.
 
     Uses an iterative flood-fill to avoid Python's recursion limit on large
-    images. Modifies ``unchecked``, ``footprint``, and ``bounds`` in-place.
+    images.  Modifies ``unchecked``, ``footprint``, and ``bounds`` in-place.
 
     Parameters
     ----------
@@ -98,7 +114,7 @@ def get_connected_pixels(i, j, image, unchecked, footprint, bounds, thresh=0):
                 stack.append((ci, cj + 1))
 
 
-def get_peaks(image, min_separation, y0=0, x0=0):
+def _get_patch_peaks(image, min_separation, y0=0, x0=0):
     """Find local maxima in an image patch.
 
     A pixel is a local maximum if its value strictly exceeds all 8-connected
@@ -109,7 +125,7 @@ def get_peaks(image, min_separation, y0=0, x0=0):
     image : 2D ndarray
         The (possibly masked) image patch to search for peaks.
     min_separation : float
-        Minimum pixel distance between peaks. When two peaks are closer than
+        Minimum pixel distance between peaks.  When two peaks are closer than
         this, the dimmer one is removed.
     y0, x0 : int, optional
         Row and column offsets added to peak coordinates so they refer to the
@@ -176,7 +192,7 @@ def get_footprints(image, min_separation, min_area, thresh=0):
     Parameters
     ----------
     image : 2D array-like
-        The image to detect sources in. Accepts NumPy or JAX arrays.
+        The image to detect sources in.  Accepts NumPy or JAX arrays.
     min_separation : float
         Minimum pixel separation between peaks within a footprint.
     min_area : int
@@ -207,9 +223,681 @@ def get_footprints(image, min_separation, min_area, thresh=0):
                 if sub_fp.sum() >= min_area:
                     patch = image[bounds[0] : bounds[1] + 1, bounds[2] : bounds[3] + 1].copy()
[…]
