+++ tests/scarlet2/test_detect.py
@@ -1,0 +1,222 @@
+# ruff: noqa: D101
+# ruff: noqa: D102
+# ruff: noqa: D103
+
+import numpy as np
+import pytest
+
+from scarlet2.detect import (
+    Footprint,
+    Peak,
+    SceneSource,
+    bounds_to_bbox,
+    box_intersect,
+    build_source_list,
+    flatten_source_list,
+    footprint_intersect,
+    get_connected_pixels,
+    get_footprints,
+)
+from scarlet2.wavelets import get_multiresolution_support, starlet_transform
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+
+def _make_blob(shape, center, radius, peak_value, background=0.0):
+    """Return an image with a filled square blob."""
+    img = np.full(shape, background, dtype=np.float32)
+    y0, y1 = max(0, center[0] - radius), min(shape[0], center[0] + radius + 1)
+    x0, x1 = max(0, center[1] - radius), min(shape[1], center[1] + radius + 1)
+    img[y0:y1, x0:x1] = peak_value / 2
+    img[center] = peak_value
+    return img
+
+
+def _detect_coeffs(img, scales=3):
+    """Return masked starlet coefficients for a 2-D image."""
+    coeffs = np.asarray(starlet_transform(img, scales=scales))
+    sigma = float(np.std(img[img > 0])) * 0.1
+    M = get_multiresolution_support(img, coeffs, sigma)
+    return M * coeffs
+
+
+# ---------------------------------------------------------------------------
+# get_connected_pixels
+# ---------------------------------------------------------------------------
+
+
+def test_get_connected_pixels_basic():
+    img = np.zeros((10, 10))
+    img[3:6, 3:6] = 1.0
+    unchecked = np.ones((10, 10), dtype=bool)
+    footprint = np.zeros((10, 10), dtype=bool)
+    bounds = np.array([4, 4, 4, 4], dtype=int)
+
+    get_connected_pixels(4, 4, img, unchecked, footprint, bounds, thresh=0)
+
+    assert footprint[3:6, 3:6].all()
+    assert not footprint[0, 0]
+    assert bounds[0] == 3 and bounds[1] == 5
+    assert bounds[2] == 3 and bounds[3] == 5
+
+
+def test_get_connected_pixels_thresh():
+    img = np.zeros((5, 5))
+    img[2, 2] = 0.5  # below thresh=1
+    unchecked = np.ones((5, 5), dtype=bool)
+    footprint = np.zeros((5, 5), dtype=bool)
+    bounds = np.array([2, 2, 2, 2], dtype=int)
+
+    get_connected_pixels(2, 2, img, unchecked, footprint, bounds, thresh=1.0)
+
+    assert not footprint.any()
+
+
+# ---------------------------------------------------------------------------
+# get_footprints
+# ---------------------------------------------------------------------------
+
+
+def test_get_footprints_two_blobs():
+    img = np.zeros((30, 30), dtype=np.float32)
+    img[4:7, 4:7] = 5.0
+    img[5, 5] = 10.0
+    img[20:23, 20:23] = 3.0
+    img[21, 21] = 6.0
+
+    fps = get_footprints(img, min_separation=2, min_area=4, thresh=0)
+
+    assert len(fps) == 2
+    centers = {(fp.peaks[0].y, fp.peaks[0].x) for fp in fps}
+    assert (5, 5) in centers
+    assert (21, 21) in centers
+
+
+def test_get_footprints_returns_footprint_type():
+    img = np.zeros((10, 10), dtype=np.float32)
[…]
