### 2026-03-18 bbb1fce
Footprint.bounds consitent with Box.bounds; scales option during detection
