+++ src/scarlet2/detect.py
@@ -52,13 +52,14 @@ class Footprint:
         Boolean mask of the footprint pixels, sized to the bounding box.
     peaks : list of Peak
         Peaks found within this footprint, sorted brightest-first.
-    bounds : ndarray of int, shape (4,)
-        Bounding box ``[y_min, y_max, x_min, x_max]`` in the full image.
+    bounds : tuple of two (min, max) pairs
+        Bounding box ``((y_min, y_max), (x_min, x_max))`` in the full image,
+        with exclusive end coordinates (consistent with :class:`~scarlet2.bbox.Box`).
     """
 
     footprint: np.ndarray
     peaks: List[Peak]
-    bounds: np.ndarray
+    bounds: Tuple[Tuple[int, int], Tuple[int, int]]
 
 
 # ---------------------------------------------------------------------------
@@ -82,8 +83,9 @@ def get_connected_pixels(i, j, image, unchecked, footprint, bounds, thresh=0):
         Tracks unvisited pixels; updated in-place.
     footprint : 2D boolean ndarray
         Accumulates the footprint mask; updated in-place.
-    bounds : ndarray of int, shape (4,)
-        Bounding box ``[y_min, y_max, x_min, x_max]``; updated in-place.
+    bounds : list of two [min, max] pairs
+        Bounding box ``[[y_min, y_max], [x_min, x_max]]`` with exclusive end
+        coordinates; updated in-place.
     thresh : float, optional
         Pixels must strictly exceed this value to join the footprint.
     """
@@ -96,14 +98,14 @@ def get_connected_pixels(i, j, image, unchecked, footprint, bounds, thresh=0):
         unchecked[ci, cj] = False
         if image[ci, cj] > thresh:
             footprint[ci, cj] = True
-            if ci < bounds[0]:
-                bounds[0] = ci
-            elif ci > bounds[1]:
-                bounds[1] = ci
-            if cj < bounds[2]:
-                bounds[2] = cj
-            elif cj > bounds[3]:
-                bounds[3] = cj
+            if ci < bounds[0][0]:
+                bounds[0][0] = ci
+            if ci + 1 > bounds[0][1]:
+                bounds[0][1] = ci + 1
+            if cj < bounds[1][0]:
+                bounds[1][0] = cj
+            if cj + 1 > bounds[1][1]:
+                bounds[1][1] = cj + 1
             if ci > 0:
                 stack.append((ci - 1, cj))
             if ci < height - 1:
@@ -214,21 +216,22 @@ def get_footprints(image, min_separation, min_area, thresh=0):
 
     for i in range(height):
         for j in range(width):
-            bounds = np.array([i, i, j, j], dtype=int)
+            bounds = [[i, i + 1], [j, j + 1]]
             get_connected_pixels(i, j, image, unchecked, footprint, bounds, thresh)
-            sub_h = bounds[1] - bounds[0] + 1
-            sub_w = bounds[3] - bounds[2] + 1
+            (y0, y1), (x0, x1) = bounds
+            sub_h = y1 - y0
+            sub_w = x1 - x0
             if sub_h * sub_w > min_area:
-                sub_fp = footprint[bounds[0] : bounds[1] + 1, bounds[2] : bounds[3] + 1]
+                sub_fp = footprint[y0:y1, x0:x1]
                 if sub_fp.sum() >= min_area:
-                    patch = image[bounds[0] : bounds[1] + 1, bounds[2] : bounds[3] + 1].copy()
+                    patch = image[y0:y1, x0:x1].copy()
                     patch[~sub_fp] = 0
-                    peaks = _get_patch_peaks(
-                        patch, min_separation, y0=int(bounds[0]), x0=int(bounds[2])
-                    )
[…]
