### 2026-03-19 d60bb46
inverse variance weighting; nearest peak association across scales
