### 2026-03-20 b6abeae
detection footprint can be limited to existing catalog
