### 2026-03-20 cc5d581
added plotting of detection footprints; removed `sky_coords` in favor of `add_peaks`
