### 2026-03-23 398b58d
added `init.hierarchical_sources`; exposed kwargs and added docstrongs in detect.py
