### 2026-03-26 dd3092b
wavelet sigma calculation avoids image boundary; minimum box size scale dependent
