### 2026-03-26 bac00cf
SNR-based box size increase
