### 2026-03-27 7e3faab
slight refactor of refine
