### 2026-03-30 #9646 @ivastar
@pmelchior, a couple of very minor comments:

Line 52: sounds like a statement that GALFIT is in IDL. GALFIT is actually in C, GALAPAGOS-2 is in IDL.

Line 59: A statement about the lack of non-parametric implementations. How about `imcascade` https://github.com/tbmiller-astro/imcascade

In the acknowledgements, please include citation for the major dependencies of scarlet2 that are not already mentioned in the text.

Finally, for `equinox` I recommend including the arXiv paper:
@article{kidger2021equinox,
  author  = {Kidger, Patrick and Garcia, Cristian},
  title   = {Equinox: neural networks in JAX via callable PyTrees and filtered transformations},
  journal = {arXiv preprint arXiv:2111.00254},
  year    = {2021}
}
