### 2026-03-31 d5dcca0
added citations for code dependencies
