+++ paper/paper.bib
@@ -137,10 +137,10 @@ archivePrefix = {arXiv},
 }
 
 @article{kidger2021equinox,
-    author={Patrick Kidger and Cristian Garcia},
-    title={{E}quinox: neural networks in {JAX} via callable {P}y{T}rees and filtered transformations},
-    year={2021},
-    journal={Differentiable Programming workshop at Neural Information Processing Systems 2021}
+  author = {Kidger, Patrick and Garcia, Cristian},
+  title = {Equinox: neural networks in JAX via callable PyTrees and filtered transformations},
+  journal = {arXiv preprint arXiv:2111.00254},
+  year = {2021}
 }
 
 @software{deepmind2020jax,
@@ -261,6 +261,51 @@ archivePrefix = {arXiv},
   url       = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
 }
 
+@Article{numpy,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ month         = sep,
+ journal       = {Nature},
+ volume        = {585},
+ number        = {7825},
+ pages         = {357--362},
+ doi           = {10.1038/s41586-020-2649-2},
+ publisher     = {Springer Science and Business Media {LLC}},
+ url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@Article{matplotlib,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
+@Book{h5py,
+        keywords = {python, hdf5},
+        year = {2013},
+        publisher = {O'Reilly},
+        title = {Python and HDF5},
+        author = {Andrew Collette}
+}
 
 
 
