### 2026-03-31 #9646 @pmelchior
@ivastar made all the changes you suggested, except for `imcascade`. It's not non-parametric, it replaces expensive Sersic functions by mixtures of Gaussians, which makes convolutions analytic (`ngmix` does that as well).
