### 2026-03-31 #9646 @ivastar
@pmelchior please go ahead with the remaining tasks on the list above. Please make sure that the authors, title and the license on Zenodo match those of the JOSS paper. Then I'll do the rest.
