+++ .github/workflows/testing-and-coverage.yml
@@ -34,6 +34,6 @@ jobs:
       run: |
         python -m pytest --cov=scarlet2 --cov-report=xml
     - name: Upload coverage report to codecov
-      uses: codecov/codecov-action@v5
+      uses: codecov/codecov-action@v6
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
