### 2026-04-03 #9646 @pmelchior
Double-checked ORCIDs; included new affiliations for co-authors, so please make sure you use the latest commit of the paper.

The version number for the release is v0.5.1.

The version is archived on Zenodo, DOI: [10.5281/zenodo.19410794](https://doi.org/10.5281/zenodo.19410794). Because the archive was created through Zenodo's GitHub integration from a new release, there was no option to define authors or license. In terms of citations, I will add a CITATION.cff file to scarlet2, which points to this JOSS paper (with the authors as listed), not the Zenodo archive.
