### 2026-04-08 #9646 @ivastar
Hi @pmelchior! I believe you can edit the Zenodo archive once it has been uploaded. Specifically can you please change the title do it matches the paper?
