### 2026-04-10 402d262
use union of boxes/footprints across scales
