### 2026-04-10 00b77d9
more defensive plotting
