### 2026-04-10 ca7c512
corrected matching with limit_to and SNR-based box size increase
