+++ src/scarlet2/detect.py
@@ -184,7 +184,7 @@ def _get_patch_peaks(image, min_separation, y0=0, x0=0):
     return peaks
 
 
-def footprints(image, min_separation, min_area, thresh=0):
+def footprints(image, min_separation=0, min_area=9, thresh=0):
     """Detect footprints and their peaks in an image.
 
     Iterates over every pixel, flood-fills each connected region of pixels
@@ -195,9 +195,9 @@ def footprints(image, min_separation, min_area, thresh=0):
     ----------
     image : 2D array-like
         The image to detect sources in.  Accepts NumPy or JAX arrays.
-    min_separation : float
+    min_separation : float, optional
         Minimum pixel separation between peaks within a footprint.
-    min_area : int
+    min_area : int, optional
         Minimum number of pixels a footprint must contain to be kept.
     thresh : float, optional
         Detection threshold; pixels must strictly exceed this value.
@@ -607,7 +607,7 @@ def get_detect_wavelets(images, variance, max_scale=3, K=3):
     return M * _coeffs, np.asarray(sigma_j)
 
 
-def get_blend_trees(detect, scales=None, min_separation=0, min_area=4, thresh=0):
+def get_blend_trees(detect, scales=None, min_separation=0, min_area=9, thresh=0):
     """Build a :class:`QuadTreeRegion` for each wavelet scale in ``detect``.
 
     Parameters
@@ -653,7 +653,7 @@ def get_blend_trees(detect, scales=None, min_separation=0, min_area=4, thresh=0)
     return trees, all_footprints
 
 
-def get_blend_structures(detect, scales=None, min_separation=0, min_area=4, thresh=0):
+def get_blend_structures(detect, scales=None, min_separation=0, min_area=9, thresh=0):
     """Build :class:`SingleScaleStructure` objects for the third wavelet scale.
 
     Each structure at the largest scale is linked to all overlapping footprints at
@@ -737,7 +737,7 @@ class HierarchicalFootprint:
     children: List["HierarchicalFootprint"] = field(default_factory=list)
 
 
-def hierarchical_footprints(detect, flatten=False, scales=None, limit_to=None, sigma_scales=None, K=3, min_separation=0, min_area=4, thresh=0):
+def hierarchical_footprints(detect, flatten=True, scales=None, limit_to=None, sigma_scales=None, K=3, min_separation=0, min_area=9, thresh=0):
     """Decompose a detection image into a hierarchy of :class:`HierarchicalFootprint` objects.
 
     Starting from the largest starlet scale and working down to the smallest scale,
@@ -804,14 +804,16 @@ def hierarchical_footprints(detect, flatten=False, scales=None, limit_to=None, s
     def _snr_bbox(fp, scale):
         """Bounding box grown to the SNR-predicted extent of an exponential profile.
 
-        At scale ``j`` the B-spline kernel half-support is ``2*2^j`` pixels.
-        For a profile I(r) = I0*exp(-r/h) the radius at which the profile
-        drops to the K*sigma threshold is h*ln(S), where S = I0/(K*sigma_j).
-        Normalising so that S=K gives the minimum half-size 2*2^j, the full
-        formula is ``half_size = 2*2^j * ln(S) / ln(K)``.  The result is
-        unioned with the tight footprint bbox so the box never shrinks.
+        For a profile I(r) = I0*exp(-r/h), the footprint boundary lies at the
+        detection threshold K*sigma_j, so h = r_foot / ln(S) where
+        S = I0/(K*sigma_j) and r_foot is the mean distance from the peak to the
+        edge of the footprint bbox.  The box is grown to the radius
+        where the profile drops to the noise level (1*sigma_j):
[…]
