### 2026-04-13 9442938
external catalog match in catalog order; returns None for non-detections
