### 2026-04-14 #9646 @ivastar
@pmelchior everything looks good! The Editor in Chief will now review the submission and do the final publication magic.

@AlexandreAdam, @Jammy2211 thank you for you reviewing this submission! These efforts make JOSS possible.
