### 2026-04-21 eda4028
docstrings for catalog matching
