+++ src/scarlet2/detect.py
@@ -14,6 +14,7 @@ from dataclasses import dataclass, field
 from typing import List, Tuple
 
 import numpy as np
+from scipy.optimize import linear_sum_assignment
 
 from .bbox import Box, overlap_slices
 from .wavelets import get_multiresolution_support, starlet_transform
@@ -752,6 +753,21 @@ def hierarchical_footprints(detect, flatten=True, scales=None, catalog=None, sig
     top-level source.  This catches compact sources that are below the noise at
     the largest scale but significant at finer scales.
 
+    If ``catalog`` is provided, the detected sources are matched to the catalog
+    positions via a global bipartite assignment (Hungarian algorithm).  The cost
+    of assigning catalog entry ``i`` to source ``j`` is the squared distance from
+    the catalog position to the nearest peak of source ``j``, restricted to cases
+    where the catalog position falls inside source ``j``'s footprint mask.
+    Catalog entries with no containing footprint are returned as ``None``.
+
+    If ``sigma_scales`` is provided, each footprint's bounding box is grown beyond
+    the detection threshold to the noise level.  Assuming an exponential profile
+    I(r) = I0*exp(-r/h), the scale length h is estimated from the footprint size
+    as h = r_foot / ln(S), where r_foot is the distance from the peak to the
+    farthest edge of the footprint bounding box and S = I0 / (K*sigma_j).  The
+    box is grown to the radius where the profile reaches the noise level (1*sigma_j):
+    half_size = r_foot * ln(S*K) / ln(S).
+
     Parameters
     ----------
     detect : ndarray, shape (scales+1, H, W)
@@ -762,13 +778,14 @@ def hierarchical_footprints(detect, flatten=True, scales=None, catalog=None, sig
         Indices into ``detect`` specifying which planes to use.  If ``None``
         (default) all planes are used.
     catalog : list of (y, x) tuples, optional
-        If given, only footprints that contain at least one of these pixel
-        positions are returned.
+        If given, the output is catalog-indexed: one entry per catalog position,
+        matched to the best overlapping detected source, or ``None`` if undetected.
+        Matching is a global optimal assignment — each source is assigned to at
+        most one catalog entry.
     sigma_scales : array-like, shape (max_scale+1,), optional
         Per-scale noise estimate from :func:`~scarlet2.detect.get_detect_wavelets`.
-        When provided, each footprint's bounding box is grown using an
-        SNR-based exponential-profile formula so that the box captures the
-        full extent of the source down to the detection threshold.
+        When provided, each footprint's bounding box is grown to the noise level
+        using the SNR-based exponential-profile formula described above.
     K : float, optional
         Detection threshold multiplier used when ``sigma_scales`` is given.
         Must match the value passed to :func:`~scarlet2.detect.get_detect_wavelets`.
@@ -782,10 +799,13 @@ def hierarchical_footprints(detect, flatten=True, scales=None, catalog=None, sig
 
     Returns
     -------
-    sources : list of HierarchicalFootprint
-        Top-level sources, each potentially carrying a tree of children.
-        Sources detected only at finer scales appear as additional top-level
-        entries with their ``scale`` set accordingly.
+    sources : list of HierarchicalFootprint or None
+        When ``catalog`` is ``None``: top-level sources, each potentially carrying
+        a tree of children.  Sources detected only at finer scales appear as
+        additional top-level entries with their ``scale`` set accordingly.
+        When ``catalog`` is given: catalog-length list where each entry is the
+        matched :class:`HierarchicalFootprint`, or ``None`` if no source was
+        detected at that catalog position.
     """
     if scales is None:
         scales = scale_indices = list(range(len(detect)))
@@ -955,34 +975,28 @@ def hierarchical_footprints(detect, flatten=True, scales=None, catalog=None, sig
[…]
