### 2026-04-21 aec17c0
allow overwriting of sources by direct assignment
