### 2026-04-22 cb454ed
added footprint splitting with watershed
