### 2026-04-22 0e4cfaf
peak and scale hierarchies combined
