+++ src/scarlet2/detect.py
@@ -283,6 +283,33 @@ def footprint_intersect(footprint1, box1, footprint2, box2):
     return np.sum(footprint1[slices1] * footprint2[slices2]) > 0
 
 
+def footprint_iou(source1, source2):
+    """Compute intersection over union (IoU) between two source footprints.
+
+    Parameters
+    ----------
+    source1, source2 : :class:`HierarchicalFootprint`
+        Sources with a `.bbox` (:class:`~scarlet2.bbox.Box`) and `.footprint`
+        (boolean ndarray sized to that bbox).
+
+    Returns
+    -------
+    iou : float
+        IoU in ``[0, 1]``.  Returns ``0`` if the bounding boxes do not overlap.
+    """
+    def _mask(source):
+        fp = source.footprint
+        return fp.footprint if isinstance(fp, Footprint) else fp
+
+    if not box_intersect(source1.bbox, source2.bbox):
+        return 0.0
+    slices1, slices2 = overlap_slices(source1.bbox, source2.bbox)
+    mask1, mask2 = _mask(source1), _mask(source2)
+    intersection = int(np.sum(mask1[slices1] & mask2[slices2]))
+    union = int(np.sum(mask1)) + int(np.sum(mask2)) - intersection
+    return float(intersection / union) if union > 0 else 0.0
+
+
 # ---------------------------------------------------------------------------
 # QuadTree
 # ---------------------------------------------------------------------------
@@ -835,20 +862,20 @@ class HierarchicalFootprint:
     children: List["HierarchicalFootprint"] = field(default_factory=list)
 
 
-def hierarchical_footprints(detect, flatten=True, scales=None, catalog=None, sigma_scales=None, K=3, min_separation=0, min_area=9, thresh=0):
+def hierarchical_footprints(detect, flatten=True, scales=None, catalog=None, sigma_scales=None, K=3, split_peaks=True, min_separation=0, min_area=9, thresh=0):
     """Decompose a detection image into a hierarchy of :class:`HierarchicalFootprint` objects.
 
-    Starting from the largest starlet scale and working down to the smallest scale,
-    each source is refined by finding the fine-scale footprint that contains its
-    current center — that footprint's peak becomes the new center.  Any other
-    overlapping fine-scale footprints whose peak lies inside the parent's footprint
-    mask become child footprints.  Children are excluded from the parent's subsequent
-    refinement steps to prevent double-counting.
+    Iterates from the largest starlet scale to the smallest.  At each scale,
+    every detected footprint is matched to the best-overlapping source already
+    registered from larger scales (measured by IoU).  If the registered source's
+    center lies inside the new footprint, it is a *primary* match: the source
+    center is refined and its footprint is grown to the union.  Otherwise the new
+    footprint becomes a *child* of the best-matching source.  Footprints with no
+    overlap with any registered source are promoted to new top-level sources.
 
-    After the top-down pass, a sweep over every finer scale promotes any footprint
-    whose primary peak is not already covered by an existing source to a new
-    top-level source.  This catches compact sources that are below the noise at
-    the largest scale but significant at finer scales.
+    When ``split_peaks`` is ``True`` (default), footprints that contain more than
+    one peak are split into separate sources by using a watershed algorithm.
+    Otherwise, additional peaks become children of the originating footprint.
 
     If ``catalog`` is provided, the detected sources are matched to the catalog
     positions via a global bipartite assignment (Hungarian algorithm).  The cost
@@ -860,9 +887,9 @@ def hierarchical_footprints(detect, flatten=True, scales=None, catalog=None, sig
     If ``sigma_scales`` is provided, each footprint's bounding box is grown beyond
[…]
