### 2026-04-22 2315ad5
performance optimization; added scipy dependency
