+++ src/scarlet2/detect.py
@@ -1,9 +1,7 @@
 """Detection utilities: connected-pixel footprint extraction and peak finding.
 
-Low-level routines (``get_connected_pixels``, ``get_footprints``) are
-translated from scarlet v1's ``detect_pybind11.cc``.  Higher-level helpers
-(``get_wavelets``, ``QuadTreeRegion``, ``get_peaks``, …) are adapted from
-scarlet v1's ``detect.py``.
+Higher-level helpers (``get_wavelets``, ``QuadTreeRegion``, ``get_peaks``, …)
+are adapted from scarlet v1's ``detect.py``.
 
 Uses NumPy rather than JAX because detection involves dynamic data structures
 (variable-length peak lists, irregularly shaped footprints) that are not
@@ -16,7 +14,7 @@ from typing import List, Tuple
 import heapq
 
 import numpy as np
-from scipy.ndimage import binary_fill_holes
+from scipy.ndimage import binary_fill_holes, find_objects, label as ndimage_label
 
 
 from scipy.optimize import linear_sum_assignment
@@ -69,57 +67,8 @@ class Footprint:
 
 
 # ---------------------------------------------------------------------------
-# Low-level detection (translated from detect_pybind11.cc)
+# Low-level detection
 # ---------------------------------------------------------------------------
-
-
-def get_connected_pixels(i, j, image, unchecked, footprint, bounds, thresh=0):
-    """Find all pixels 4-connected to ``(i, j)`` that exceed ``thresh``.
-
-    Uses an iterative flood-fill to avoid Python's recursion limit on large
-    images.  Modifies ``unchecked``, ``footprint``, and ``bounds`` in-place.
-
-    Parameters
-    ----------
-    i, j : int
-        Seed pixel coordinates (row, column).
-    image : 2D ndarray
-        The image to search.
-    unchecked : 2D boolean ndarray
-        Tracks unvisited pixels; updated in-place.
-    footprint : 2D boolean ndarray
-        Accumulates the footprint mask; updated in-place.
-    bounds : list of two [min, max] pairs
-        Bounding box ``[[y_min, y_max], [x_min, x_max]]`` with exclusive end
-        coordinates; updated in-place.
-    thresh : float, optional
-        Pixels must strictly exceed this value to join the footprint.
-    """
-    height, width = image.shape
-    stack = [(i, j)]
-    while stack:
-        ci, cj = stack.pop()
-        if not unchecked[ci, cj]:
-            continue
-        unchecked[ci, cj] = False
-        if image[ci, cj] > thresh:
-            footprint[ci, cj] = True
-            if ci < bounds[0][0]:
-                bounds[0][0] = ci
-            if ci + 1 > bounds[0][1]:
-                bounds[0][1] = ci + 1
-            if cj < bounds[1][0]:
-                bounds[1][0] = cj
-            if cj + 1 > bounds[1][1]:
-                bounds[1][1] = cj + 1
-            if ci > 0:
-                stack.append((ci - 1, cj))
-            if ci < height - 1:
-                stack.append((ci + 1, cj))
-            if cj > 0:
-                stack.append((ci, cj - 1))
-            if cj < width - 1:
-                stack.append((ci, cj + 1))
 
 
 def _get_patch_peaks(image, min_separation, y0=0, x0=0):
@@ -193,9 +142,9 @@ def _get_patch_peaks(image, min_separation, y0=0, x0=0):
 def footprints(image, min_separation=0, min_area=9, thresh=0):
     """Detect footprints and their peaks in an image.
 
-    Iterates over every pixel, flood-fills each connected region of pixels
[…]
