### 2026-04-22 #9646 @warrickball
Indeed! I've just returned from a long Easter break (made three days longer by travel disruption last week) and will start the final editorial checks soon. There are multiple papers in my publication queue, so it might take me a few days to get started, but I'll try my best to propose any fixes by the end of this weekend.
