### 2026-04-22 53dfb1b
lots of small fixes for new starlet init
