### 2026-04-23 731f066
more robust data loading
