### 2026-04-23 4d7f706
added Module.replace to help with constraint violations
