### 2026-04-23 5cef8ba
added detection notebook; updated docs to use new detection (multi-res init is still missing)
