+++ docs/1-howto.md
@@ -3,6 +3,7 @@
 ```{toctree}
 :maxdepth: 2
 
+howto/detect
 howto/sampling
 howto/priors
 howto/correlated
