+++ src/scarlet2/detect.py
@@ -10,7 +10,6 @@ compatible with JAX's JIT.  Both NumPy and JAX arrays are accepted as inputs.
 
 import heapq
 from dataclasses import dataclass, field
-from typing import List, Tuple
 
 import numpy as np
 from scipy.ndimage import binary_fill_holes, find_objects
@@ -60,8 +59,8 @@ class Footprint:
     """
 
     footprint: np.ndarray
-    peaks: List[Peak]
-    bounds: Tuple[Tuple[int, int], Tuple[int, int]]
+    peaks: list[Peak]
+    bounds: tuple[tuple[int, int], tuple[int, int]]
 
 
 # ---------------------------------------------------------------------------
@@ -192,11 +191,14 @@ def box_intersect(box1, box2):
 
     Parameters
     ----------
-    box1, box2 : Box
+    box1: Box
+        First box
+    box2: Box
+        Second box
 
     Returns
     -------
-    overlap : bool
+    bool
     """
     overlap = box1 & box2
     return overlap.shape[0] != 0 and overlap.shape[1] != 0
@@ -287,7 +289,6 @@ class QuadTreeRegion:
         self.capacity = capacity
         self.depth = depth
         self.detect = detect
-        self.debug = detect is not None
 
     def footprint_image(self, bbox=None):
         """Return a 2-D image of all footprint masks in the tree.
@@ -350,10 +351,11 @@ class QuadTreeRegion:
         Parameters
         ----------
         footprints : list of Footprint
+            Footprints to add bounding boxes for.
 
         Returns
         -------
-        self : QuadTreeRegion
+        self
         """
         for fp in footprints:
             box = Box.from_bounds(*fp.bounds)
@@ -363,21 +365,12 @@ class QuadTreeRegion:
 
     def split(self):
         """Sub-divide this region into four quadrants."""
-        import matplotlib.pyplot as plt
 
         height, width = self.bbox.shape
         h2 = height // 2
         w2 = width // 2
         h3 = height - h2
         w3 = width - w2
-
-        if self.debug:
-            fig, ax = plt.subplots()
-            ax.imshow(self.detect[2], cmap="Greys")
-            ax.set_title(self.depth)
-            draw_box(self.bbox, ax, "r")
-            for box in self.boxes:
-                draw_box(box, ax, "b")
 
         origin = self.bbox.origin
         self.sub_regions = [
@@ -463,6 +456,7 @@ class SingleScaleStructure:
         scale : int
             Wavelet scale of the primary footprint.
         footprint : Footprint
+            The footprint at the primary scale.
         """
         self.scale = scale
         self.footprint = footprint
@@ -476,7 +470,9 @@ class SingleScaleStructure:
         Parameters
         ----------
         scale : int
+            The wavelet scale of this structure.
         footprint : Footprint
+            The footprint at the primary scale.
         """
         if scale not in self.peaks:
             self.peaks[scale] = []
@@ -490,7 +486,9 @@ class SingleScaleStructure:
         Parameters
         ----------
         scale : int
+            The wavelet scale of this structure.
         tree : QuadTreeRegion
+            The tree at this scale.
 
         Returns
         -------
@@ -545,15 +543,17 @@ def get_wavelets(images, variance, max_scale=3):
     return np.array(coeffs)
 
 
-def get_detect_wavelets(images, variance, max_scale=3, K=3, image_type="ground"):
+def get_detect_wavelets(image, variance, max_scale=3, K=3, image_type="ground"):
     """Get starlet coefficients of a detection image for source finding.
 
     The detection image is inverse varianced weighted sum of `images` across all bands.
 
     Parameters
     ----------
[…]
