+++ pyproject.toml
@@ -113,6 +113,8 @@ ignore = [
     "UP015", # Allow redundant open parameters
     "UP028", # Allow yield in for loop
     "E731", # Allow assignment of lambda expressions
+    "N803", # Allow uppercase arguments 
+    "N806", # Allow uppercase variables
 ]
 
 
