### 2026-04-23 ae41b94
improved Footprint repr; more docs updates
