+++ .pre-commit-config.yaml
@@ -11,6 +11,7 @@ repos:
         description: Compare current template version against latest
         verbose: true
     # Clear output from jupyter notebooks so that only the input cells are committed.
+    # Avoid reporting error when this hook is triggered
   - repo: local
     hooks:
       - id: jupyter-nb-clear-output
@@ -20,7 +21,7 @@ repos:
         exclude: ^docs/pre_executed
         stages: [pre-commit]
         language: system
-        entry: jupyter nbconvert --clear-output
+        entry: jupyter nbconvert --clear-output || exit 0
     # Prevents committing directly branches named 'main' and 'master'.
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.4.0
@@ -34,7 +35,7 @@ repos:
         args: ['--maxkb=500']
     # Verify that pyproject.toml is well formed
   - repo: https://github.com/abravalheri/validate-pyproject
-    rev: v0.12.1
+    rev: v0.25
     hooks:
       - id: validate-pyproject
         name: Validate pyproject.toml
