### 2026-04-24 ade2964
added deprecation warning of add_labels; updated correlated noise howto
