### 2026-04-24 20bab49
replaced `jupyter nbconvert` with nbstripout to silence the hook failure
