### 2026-04-24 af9077a
generalized plot.channel_map for arbitrary channel numbers
