### 2026-04-25 #9646 @warrickball
I've started doing some initial checks and flagged a few minor things that you should fix before I make a detailed pass at the text of the paper (though at a glance it looks like it'll need very few changes).

* The library JAX is usually styled that way ("JAX", e.g. in [their docs](https://docs.jax.dev/)). Please use this consistently throughout the paper, including the title (which will mean updating the Zenodo archive title, too). I'll change the title of this issue.
* There are a few references that use NASA ADS macros for the journal titles. Please replace these with the full journal names (e.g. *The Astronomical Journal* instead of `\aj`), as you already have in many of the references.
* Though this isn't a requirement, I'm used to seeing a bit more detail in the affiliations, particularly which academic department an author is associated with. If you do update the affiliations, please include the country, too, though I'm also happy to add this alongside any other editorial tweaks I propose via PR.
* This also isn't a requirement, but I noticed that you haven't included the authors ORCIDs in the Zenodo archive, though they're present in the paper. We encourage authors to include ORCIDs in both places.

Thanks!
