### 2026-04-26 dbd70bc
added HierarchicalSourceInfo as return type; updated multi-observation howto
