+++ src/scarlet2/init.py
@@ -1,14 +1,19 @@
 """Helper methods to initialize sources"""
 
 import operator
+import warnings
+from dataclasses import dataclass
 from functools import reduce
 
+import astropy
 import astropy.units as u
 import jax.numpy as jnp
+from equinox import tree_pformat
 
 from . import Scenery, measure
 from .bbox import Box, insert_into
-from .detect import get_detect_wavelets, hierarchical_footprints
+from .detect import HierarchicalFootprint, get_detect_wavelets, hierarchical_footprints
+from .frame import get_relative_jacobian_shift
 from .morphology import GaussianMorphology
 from .observation import Observation
 
@@ -455,7 +460,60 @@ def _sort_spectra(spectra, channels):
     return spectrum
 
 
+@dataclass
+class HierarchicalSourceInfo:
+    """Information for initializing a hierarchical source.
+
+    Attributes
+    ----------
+    peak: astropy.coordinates.SkyCoord
+        Sky position of detected peak
+    center: astropy.coordinates.SkyCoord
+        Sky position of the center of the bounding box
+    spectrum: jnp.array
+        1-D array of per-channel flux, shape ``(model.frame.C,)``
+    morphology: jnp.array
+        2-D intensity map, clipped to the footprint bounding box, normalized to maximum of 1
+    footprint: HierarchicalFootprint
+        The footprint from :func:`~scarlet2.detect.hierarchical_footprints`
+    """
+
+    peak: astropy.coordinates.SkyCoord
+    center: astropy.coordinates.SkyCoord
+    spectrum: jnp.array
+    morphology: jnp.array
+    footprint: HierarchicalFootprint
+
+    def __repr__(self):
+        return tree_pformat(self)
+
+
 def _footprints_to_sources(obs, detect, footprints, scales, catalog):
+    try:
+        frame = Scenery.scene.frame
+    except AttributeError:
+        print("Attributes defined in sky coordinates can only be created within the context of a Scene")
+        print("Use 'with Scene(frame) as scene: (...)'")
+        raise
+
+    # we need a proper renderer to determine placement in the model frame
+    obs.check_set_renderer(frame)
+    # spatial differences
+    jac, shift = get_relative_jacobian_shift(frame, obs.frame)
+    if not jnp.allclose(jnp.linalg.det(jac), 1) or not jnp.allclose(shift, shift.astype(int)):
+        warnings.warn(
+            "hierarchical_sources does not support resampling. Align `obs` first!",
+            stacklevel=2,
+        )
+    # map between frame and obs channels: assumes clean 1-to-1 mapping
+    if frame.channels != obs.frame.channels:
+        frame_channels = list(frame.channels)
+        obs_channels = list(obs.frame.channels)
+        obs_idx = {c: i for i, c in enumerate(obs_channels)}
+        channel_map = {i: obs_idx[c] for i, c in enumerate(frame_channels) if c in obs_idx}
+        frame_channels = jnp.asarray(list(channel_map.keys()))
+        obs_channels = jnp.asarray(list(channel_map.values()))
+
     _images = jnp.copy(obs.data)
     shape = obs.frame.bbox.spatial.shape
     res = [
@@ -466,13 +524,15 @@ def _footprints_to_sources(obs, detect, footprints, scales, catalog):
             if fp is not None and fp.scale == scale:
                 footprint_map = insert_into(jnp.zeros(shape), jnp.asarray(fp.footprint), fp.bbox)
                 source_detect = detect[scale] * footprint_map
-                # MLE of flux per band given the morphology from source_detect
+                # spectrum: MLE of flux per band given the morphology from source_detect
                 spectrum = (_images * source_detect[None, :, :]).sum(axis=(-2, -1)) / jnp.sum(
                     source_detect**2
                 )
[…]
