### 2026-04-26 4487452
switch to jax.numpy; use flat spectrum and NNLS for forced photometry
