### 2026-04-28 154b63f
minor docstring fix
