+++ paper/paper.bib
@@ -172,7 +172,7 @@ archivePrefix = {arXiv},
 @ARTICLE{galfit,
        author = {{Peng}, Chien Y. and {Ho}, Luis C. and {Impey}, Chris D. and {Rix}, Hans-Walter},
         title = "{Detailed Structural Decomposition of Galaxy Images}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {Galaxies: Bulges, Galaxies: Fundamental Parameters, Galaxies: Nuclei, Galaxies: Structure, Techniques: Image Processing, Techniques: Photometric, Astrophysics},
          year = 2002,
         month = jul,
@@ -190,7 +190,7 @@ archivePrefix = {arXiv},
 @ARTICLE{galapagos,
        author = {{Barden}, Marco and {H{\"a}u{\ss}ler}, Boris and {Peng}, Chien Y. and {McIntosh}, Daniel H. and {Guo}, Yicheng},
         title = "{GALAPAGOS: from pixels to parameters}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: data analysis, surveys, galaxies: statistics, galaxies: structure, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Extragalactic Astrophysics},
          year = 2012,
         month = may,
@@ -208,7 +208,7 @@ archivePrefix = {arXiv},
 @ARTICLE{galapagos2,
        author = {{H{\"a}u{\ss}ler}, Boris and {Bamford}, Steven P. and {Vika}, Marina and {Rojas}, Alex L. and {Barden}, Marco and {Kelvin}, Lee S. and {Alpaslan}, Mehmet and {Robotham}, Aaron S.~G. and {Driver}, Simon P. and {Baldry}, I.~K. and {Brough}, Sarah and {Hopkins}, Andrew M. and {Liske}, Jochen and {Nichol}, Robert C. and {Popescu}, Cristina C. and {Tuffs}, Richard J.},
         title = "{MegaMorph - multiwavelength measurement of galaxy structure: complete S{\'e}rsic profile information from modern surveys}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: data analysis, techniques: image processing, galaxies: fundamental parameters, galaxies: structure, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2013,
         month = mar,
@@ -239,7 +239,7 @@ archivePrefix = {ascl},
 @ARTICLE{astrophot,
        author = {{Stone}, Connor J. and {Courteau}, St{\'e}phane and {Cuillandre}, Jean-Charles and {Hezaveh}, Yashar and {Perreault-Levasseur}, Laurence and {Arora}, Nikhil},
         title = "{AstroPhot: Fitting Everything Everywhere All at Once in Astronomical Images}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {galaxies: general, galaxies: photometry, stars: imaging, software: data analysis, techniques: image processing, techniques: photometric, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
          year = 2023,
         month = aug,
