### 2026-04-29 a7e4a01
Initial editorial tweaks for publication in JOSS (#266)
