+++ paper/paper.bib
@@ -9,13 +9,15 @@
     year = 2013,
     month = oct,
     volume = 558,
+    eid = {A33},
+    pages = {A33},
     doi = {10.1051/0004-6361/201322068},
     url = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
 }
 
 @ARTICLE{oldag-2024,
     author = {{Oldag}, Drew and {DeLucchi}, Melissa and {Beebe}, Wilson and {Branton}, Doug and {Campos}, Sandro and {Chandler}, Colin Orion and {Christofferson}, Carl and {Connolly}, Andrew and {Kubica}, Jeremy and {Lynn}, Olivia and {Malanchev}, Konstantin and {Malz}, Alex I. and {Mandelbaum}, Rachel and {McGuire}, Sean and {Wenneman}, Chris},
-    title = "{A Python Project Template for Healthy Scientific Software}",
+    title = {A {Python} Project Template for Healthy Scientific Software},
     journal = {Research Notes of the American Astronomical Society},
     keywords = {Open source software, 1866},
     year = 2024,
@@ -31,7 +33,7 @@
 
 @ARTICLE{siegel-2025,
        author = {{Siegel}, Jared C. and {Melchior}, Peter},
-        title = "{Spatially Resolved Galaxy{\textendash}Dust Modeling with Coupled Data-driven Priors}",
+        title = {Spatially Resolved Galaxy{\textendash}Dust Modeling with Coupled Data-driven Priors},
       journal = {The Astrophysical Journal},
      keywords = {Interstellar dust extinction, Neural networks, Extragalactic astronomy, 837, 1933, 506, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2025,
@@ -106,6 +108,7 @@ archivePrefix = {arXiv},
   title={Composable Effects for Flexible and Accelerated Probabilistic Programming in NumPyro},
   author={Phan, Du and Pradhan, Neeraj and Jankowiak, Martin},
   journal={arXiv preprint arXiv:1912.11554},
+  doi={10.48550/arXiv.1912.11554},
   year={2019}
 }
 
@@ -139,7 +142,10 @@ archivePrefix = {arXiv},
 @article{kidger2021equinox,
   author = {Kidger, Patrick and Garcia, Cristian},
   title = {Equinox: neural networks in JAX via callable PyTrees and filtered transformations},
-  journal = {arXiv preprint arXiv:2111.00254},
+  journal = {Differentiable Programming workshop at Neural Information Processing Systems 2021},
+  doi = {10.48550/arXiv.2111.00254},
+  archivePrefix = {arXiv},
+  eprint = {2111.00254},
   year = {2021}
 }
 
@@ -152,7 +158,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{yao-2025,
[…]
